% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subProcessors.R
\name{accessField}
\alias{accessField}
\title{Access the field in a list}
\usage{
accessField(d, field, natype = NA)
}
\arguments{
\item{d}{Input data list}

\item{field}{Field name to access in each element}

\item{natype}{What to fill in when field is unavailable}
}
\value{
A vector of elements
}
\description{
This function accesses named field within the elements of a list. If an element
lacks the field, it's filled in by natype.
}
\keyword{internal}
