% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_value_object.R
\docType{class}
\name{AnnotationValueObject}
\alias{AnnotationValueObject}
\title{AnnotationValueObject}
\format{
An \code{R6Class} generator object
}
\description{
AnnotationValueObject Class

AnnotationValueObject Class
}
\details{
Create a new AnnotationValueObject
}
\keyword{(JSON}
\keyword{(if}
\keyword{AnnotationValueObject}
\keyword{Deserialize}
\keyword{Initialize}
\keyword{JSON}
\keyword{Return}
\keyword{To}
\keyword{Validate}
\keyword{a}
\keyword{all}
\keyword{an}
\keyword{any).}
\keyword{are}
\keyword{class.}
\keyword{fields}
\keyword{format)}
\keyword{if}
\keyword{in}
\keyword{input}
\keyword{instance}
\keyword{internal}
\keyword{into}
\keyword{invalid}
\keyword{list}
\keyword{new}
\keyword{of}
\keyword{respect}
\keyword{string}
\keyword{the}
\keyword{to}
\keyword{true}
\keyword{valid.}
\keyword{values}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer optional}

\item{\code{classUri}}{character optional}

\item{\code{className}}{character optional}

\item{\code{termUri}}{character optional}

\item{\code{termName}}{character optional}

\item{\code{parentName}}{character optional}

\item{\code{parentDescription}}{character optional}

\item{\code{parentLink}}{character optional}

\item{\code{parentOfParentName}}{character optional}

\item{\code{parentOfParentDescription}}{character optional}

\item{\code{parentOfParentLink}}{character optional}

\item{\code{description}}{character optional}

\item{\code{evidenceCode}}{character optional}

\item{\code{objectClass}}{character optional}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnnotationValueObject-new}{\code{AnnotationValueObject$new()}}
\item \href{#method-AnnotationValueObject-toJSON}{\code{AnnotationValueObject$toJSON()}}
\item \href{#method-AnnotationValueObject-fromJSON}{\code{AnnotationValueObject$fromJSON()}}
\item \href{#method-AnnotationValueObject-toJSONString}{\code{AnnotationValueObject$toJSONString()}}
\item \href{#method-AnnotationValueObject-fromJSONString}{\code{AnnotationValueObject$fromJSONString()}}
\item \href{#method-AnnotationValueObject-validateJSON}{\code{AnnotationValueObject$validateJSON()}}
\item \href{#method-AnnotationValueObject-toString}{\code{AnnotationValueObject$toString()}}
\item \href{#method-AnnotationValueObject-isValid}{\code{AnnotationValueObject$isValid()}}
\item \href{#method-AnnotationValueObject-getInvalidFields}{\code{AnnotationValueObject$getInvalidFields()}}
\item \href{#method-AnnotationValueObject-print}{\code{AnnotationValueObject$print()}}
\item \href{#method-AnnotationValueObject-clone}{\code{AnnotationValueObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationValueObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationValueObject-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new AnnotationValueObject class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationValueObject$new(
  id = NULL,
  classUri = NULL,
  className = NULL,
  termUri = NULL,
  termName = NULL,
  parentName = NULL,
  parentDescription = NULL,
  parentLink = NULL,
  parentOfParentName = NULL,
  parentOfParentDescription = NULL,
  parentOfParentLink = NULL,
  description = NULL,
  evidenceCode = NULL,
  objectClass = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{classUri}}{classUri}

\item{\code{className}}{className}

\item{\code{termUri}}{termUri}

\item{\code{termName}}{termName}

\item{\code{parentName}}{parentName}

\item{\code{parentDescription}}{parentDescription}

\item{\code{parentLink}}{parentLink}

\item{\code{parentOfParentName}}{parentOfParentName}

\item{\code{parentOfParentDescription}}{parentOfParentDescription}

\item{\code{parentOfParentLink}}{parentOfParentLink}

\item{\code{description}}{description}

\item{\code{evidenceCode}}{evidenceCode}

\item{\code{objectClass}}{objectClass}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationValueObject-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationValueObject-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationValueObject$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
AnnotationValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationValueObject-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationValueObject-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of AnnotationValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationValueObject$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of AnnotationValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationValueObject-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationValueObject-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationValueObject$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
AnnotationValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationValueObject-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationValueObject-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of AnnotationValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationValueObject$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of AnnotationValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationValueObject-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationValueObject-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to AnnotationValueObject and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationValueObject$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationValueObject-toString"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationValueObject-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationValueObject$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of AnnotationValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationValueObject-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationValueObject-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationValueObject$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationValueObject-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationValueObject-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationValueObject$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationValueObject-print"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationValueObject-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationValueObject$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationValueObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationValueObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationValueObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
