% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openapi_generated.R
\docType{class}
\name{ApiResponse}
\alias{ApiResponse}
\title{ApiResponse}
\format{
An \code{R6Class} generator object
}
\description{
ApiResponse Class
}
\details{
Gemma RESTful API

This website documents the usage of the Gemma REST API(https://gemma.msl.ubc.ca/rest/v2/). Here you can find example script usage of the API, as well as graphical interface for each endpoint, with description of its parameters and the endpoint URL.  The documentation of the underlying java code can be found here(https://gemma.msl.ubc.ca/resources/apidocs/ubic/gemma/web/services/rest/package-summary.html). See the links section(https://gemma.msl.ubc.ca/resources/restapidocs/#footer) in the footer of this page for other relevant links.  Use of this webpage and Gemma web services, including the REST API, is subject to these terms and conditions(https://pavlidislab.github.io/Gemma/terms.html). Please read these in full before continuing to use this webpage or any other part of the Gemma system.

The version of the OpenAPI document: 2.4.1
Contact: pavlab-support@msl.ubc.ca
Generated by: https://openapi-generator.tech
}
\keyword{ApiResponse}
\keyword{Initialize}
\keyword{a}
\keyword{class.}
\keyword{internal}
\keyword{new}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{content}}{The deserialized response body.}

\item{\code{response}}{The raw response from the endpoint.}

\item{\code{status_code}}{The HTTP response status code.}

\item{\code{status_code_desc}}{The brief descriptoin of the HTTP response status code.}

\item{\code{headers}}{The HTTP response headers.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ApiResponse-new}{\code{ApiResponse$new()}}
\item \href{#method-ApiResponse-clone}{\code{ApiResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ApiResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ApiResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiResponse$new(
  content = NULL,
  response = NULL,
  status_code = NULL,
  status_code_desc = NULL,
  headers = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{content}}{The deserialized response body.}

\item{\code{response}}{The raw response from the endpoint.}

\item{\code{status_code}}{The HTTP response status code.}

\item{\code{status_code_desc}}{The brief description of the HTTP response status code.}

\item{\code{headers}}{The HTTP response headers.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApiResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ApiResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
