% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_design_value_object.R
\docType{class}
\name{ArrayDesignValueObject}
\alias{ArrayDesignValueObject}
\title{ArrayDesignValueObject}
\format{
An \code{R6Class} generator object
}
\description{
ArrayDesignValueObject Class

ArrayDesignValueObject Class
}
\details{
Create a new ArrayDesignValueObject
}
\keyword{(JSON}
\keyword{(if}
\keyword{ArrayDesignValueObject}
\keyword{Deserialize}
\keyword{Initialize}
\keyword{JSON}
\keyword{Return}
\keyword{To}
\keyword{Validate}
\keyword{a}
\keyword{all}
\keyword{an}
\keyword{any).}
\keyword{are}
\keyword{class.}
\keyword{fields}
\keyword{format)}
\keyword{if}
\keyword{in}
\keyword{input}
\keyword{instance}
\keyword{internal}
\keyword{into}
\keyword{invalid}
\keyword{list}
\keyword{new}
\keyword{of}
\keyword{respect}
\keyword{string}
\keyword{the}
\keyword{to}
\keyword{true}
\keyword{valid.}
\keyword{values}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer optional}

\item{\code{lastUpdated}}{character optional}

\item{\code{troubled}}{character optional}

\item{\code{lastTroubledEvent}}{\link{AuditEventValueObject} optional}

\item{\code{needsAttention}}{character optional}

\item{\code{lastNeedsAttentionEvent}}{\link{AuditEventValueObject} optional}

\item{\code{curationNote}}{character optional}

\item{\code{lastNoteUpdateEvent}}{\link{AuditEventValueObject} optional}

\item{\code{blackListed}}{character optional}

\item{\code{color}}{character optional}

\item{\code{dateCached}}{character optional}

\item{\code{description}}{character optional}

\item{\code{designElementCount}}{integer optional}

\item{\code{expressionExperimentCount}}{integer optional}

\item{\code{hasBlatAssociations}}{character optional}

\item{\code{hasGeneAssociations}}{character optional}

\item{\code{hasSequenceAssociations}}{character optional}

\item{\code{isAffymetrixAltCdf}}{character optional}

\item{\code{isMerged}}{character optional}

\item{\code{isMergee}}{character optional}

\item{\code{isSubsumed}}{character optional}

\item{\code{isSubsumer}}{character optional}

\item{\code{lastGeneMapping}}{character optional}

\item{\code{lastRepeatMask}}{character optional}

\item{\code{lastSequenceAnalysis}}{character optional}

\item{\code{lastSequenceUpdate}}{character optional}

\item{\code{name}}{character optional}

\item{\code{numGenes}}{character optional}

\item{\code{numProbeAlignments}}{character optional}

\item{\code{numProbeSequences}}{character optional}

\item{\code{numProbesToGenes}}{character optional}

\item{\code{shortName}}{character optional}

\item{\code{switchedExpressionExperimentCount}}{integer optional}

\item{\code{taxon}}{character optional}

\item{\code{taxonID}}{integer optional}

\item{\code{technologyType}}{character optional}

\item{\code{troubleDetails}}{character optional}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ArrayDesignValueObject-new}{\code{ArrayDesignValueObject$new()}}
\item \href{#method-ArrayDesignValueObject-toJSON}{\code{ArrayDesignValueObject$toJSON()}}
\item \href{#method-ArrayDesignValueObject-fromJSON}{\code{ArrayDesignValueObject$fromJSON()}}
\item \href{#method-ArrayDesignValueObject-toJSONString}{\code{ArrayDesignValueObject$toJSONString()}}
\item \href{#method-ArrayDesignValueObject-fromJSONString}{\code{ArrayDesignValueObject$fromJSONString()}}
\item \href{#method-ArrayDesignValueObject-validateJSON}{\code{ArrayDesignValueObject$validateJSON()}}
\item \href{#method-ArrayDesignValueObject-toString}{\code{ArrayDesignValueObject$toString()}}
\item \href{#method-ArrayDesignValueObject-isValid}{\code{ArrayDesignValueObject$isValid()}}
\item \href{#method-ArrayDesignValueObject-getInvalidFields}{\code{ArrayDesignValueObject$getInvalidFields()}}
\item \href{#method-ArrayDesignValueObject-print}{\code{ArrayDesignValueObject$print()}}
\item \href{#method-ArrayDesignValueObject-clone}{\code{ArrayDesignValueObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayDesignValueObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayDesignValueObject-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ArrayDesignValueObject class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayDesignValueObject$new(
  id = NULL,
  lastUpdated = NULL,
  troubled = NULL,
  lastTroubledEvent = NULL,
  needsAttention = NULL,
  lastNeedsAttentionEvent = NULL,
  curationNote = NULL,
  lastNoteUpdateEvent = NULL,
  blackListed = NULL,
  color = NULL,
  dateCached = NULL,
  description = NULL,
  designElementCount = NULL,
  expressionExperimentCount = NULL,
  hasBlatAssociations = NULL,
  hasGeneAssociations = NULL,
  hasSequenceAssociations = NULL,
  isAffymetrixAltCdf = NULL,
  isMerged = NULL,
  isMergee = NULL,
  isSubsumed = NULL,
  isSubsumer = NULL,
  lastGeneMapping = NULL,
  lastRepeatMask = NULL,
  lastSequenceAnalysis = NULL,
  lastSequenceUpdate = NULL,
  name = NULL,
  numGenes = NULL,
  numProbeAlignments = NULL,
  numProbeSequences = NULL,
  numProbesToGenes = NULL,
  shortName = NULL,
  switchedExpressionExperimentCount = NULL,
  taxon = NULL,
  taxonID = NULL,
  technologyType = NULL,
  troubleDetails = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{lastUpdated}}{lastUpdated}

\item{\code{troubled}}{troubled}

\item{\code{lastTroubledEvent}}{lastTroubledEvent}

\item{\code{needsAttention}}{needsAttention}

\item{\code{lastNeedsAttentionEvent}}{lastNeedsAttentionEvent}

\item{\code{curationNote}}{curationNote}

\item{\code{lastNoteUpdateEvent}}{lastNoteUpdateEvent}

\item{\code{blackListed}}{blackListed}

\item{\code{color}}{color}

\item{\code{dateCached}}{dateCached}

\item{\code{description}}{description}

\item{\code{designElementCount}}{designElementCount}

\item{\code{expressionExperimentCount}}{expressionExperimentCount}

\item{\code{hasBlatAssociations}}{hasBlatAssociations}

\item{\code{hasGeneAssociations}}{hasGeneAssociations}

\item{\code{hasSequenceAssociations}}{hasSequenceAssociations}

\item{\code{isAffymetrixAltCdf}}{isAffymetrixAltCdf}

\item{\code{isMerged}}{isMerged}

\item{\code{isMergee}}{isMergee}

\item{\code{isSubsumed}}{isSubsumed}

\item{\code{isSubsumer}}{isSubsumer}

\item{\code{lastGeneMapping}}{lastGeneMapping}

\item{\code{lastRepeatMask}}{lastRepeatMask}

\item{\code{lastSequenceAnalysis}}{lastSequenceAnalysis}

\item{\code{lastSequenceUpdate}}{lastSequenceUpdate}

\item{\code{name}}{name}

\item{\code{numGenes}}{numGenes}

\item{\code{numProbeAlignments}}{numProbeAlignments}

\item{\code{numProbeSequences}}{numProbeSequences}

\item{\code{numProbesToGenes}}{numProbesToGenes}

\item{\code{shortName}}{shortName}

\item{\code{switchedExpressionExperimentCount}}{switchedExpressionExperimentCount}

\item{\code{taxon}}{taxon}

\item{\code{taxonID}}{taxonID}

\item{\code{technologyType}}{technologyType}

\item{\code{troubleDetails}}{troubleDetails}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayDesignValueObject-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayDesignValueObject-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayDesignValueObject$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ArrayDesignValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayDesignValueObject-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayDesignValueObject-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ArrayDesignValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayDesignValueObject$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ArrayDesignValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayDesignValueObject-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayDesignValueObject-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayDesignValueObject$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ArrayDesignValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayDesignValueObject-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayDesignValueObject-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ArrayDesignValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayDesignValueObject$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ArrayDesignValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayDesignValueObject-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayDesignValueObject-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ArrayDesignValueObject and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayDesignValueObject$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayDesignValueObject-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayDesignValueObject-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayDesignValueObject$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ArrayDesignValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayDesignValueObject-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayDesignValueObject-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayDesignValueObject$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayDesignValueObject-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayDesignValueObject-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayDesignValueObject$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayDesignValueObject-print"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayDesignValueObject-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayDesignValueObject$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayDesignValueObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayDesignValueObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayDesignValueObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
