% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_sequence.R
\docType{class}
\name{BioSequence}
\alias{BioSequence}
\title{BioSequence}
\format{
An \code{R6Class} generator object
}
\description{
BioSequence Class

BioSequence Class
}
\details{
Create a new BioSequence
}
\keyword{(JSON}
\keyword{(if}
\keyword{BioSequence}
\keyword{Deserialize}
\keyword{Initialize}
\keyword{JSON}
\keyword{Return}
\keyword{To}
\keyword{Validate}
\keyword{a}
\keyword{all}
\keyword{an}
\keyword{any).}
\keyword{are}
\keyword{class.}
\keyword{fields}
\keyword{format)}
\keyword{if}
\keyword{in}
\keyword{input}
\keyword{instance}
\keyword{internal}
\keyword{into}
\keyword{invalid}
\keyword{list}
\keyword{new}
\keyword{of}
\keyword{respect}
\keyword{string}
\keyword{the}
\keyword{to}
\keyword{true}
\keyword{valid.}
\keyword{values}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character \link{optional}}

\item{\code{description}}{character \link{optional}}

\item{\code{id}}{integer \link{optional}}

\item{\code{length}}{integer \link{optional}}

\item{\code{sequence}}{character \link{optional}}

\item{\code{isApproximateLength}}{character \link{optional}}

\item{\code{isCircular}}{character \link{optional}}

\item{\code{polymerType}}{\link{PolymerType} \link{optional}}

\item{\code{type}}{\link{SequenceType} \link{optional}}

\item{\code{fractionRepeats}}{numeric \link{optional}}

\item{\code{sequenceDatabaseEntry}}{\link{DatabaseEntry} \link{optional}}

\item{\code{taxon}}{\link{Taxon} \link{optional}}

\item{\code{bioSequence2GeneProduct}}{list(\link{BioSequence2GeneProduct}) \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BioSequence-new}{\code{BioSequence$new()}}
\item \href{#method-BioSequence-toJSON}{\code{BioSequence$toJSON()}}
\item \href{#method-BioSequence-fromJSON}{\code{BioSequence$fromJSON()}}
\item \href{#method-BioSequence-toJSONString}{\code{BioSequence$toJSONString()}}
\item \href{#method-BioSequence-fromJSONString}{\code{BioSequence$fromJSONString()}}
\item \href{#method-BioSequence-validateJSON}{\code{BioSequence$validateJSON()}}
\item \href{#method-BioSequence-toString}{\code{BioSequence$toString()}}
\item \href{#method-BioSequence-isValid}{\code{BioSequence$isValid()}}
\item \href{#method-BioSequence-getInvalidFields}{\code{BioSequence$getInvalidFields()}}
\item \href{#method-BioSequence-print}{\code{BioSequence$print()}}
\item \href{#method-BioSequence-clone}{\code{BioSequence$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioSequence-new"></a>}}
\if{latex}{\out{\hypertarget{method-BioSequence-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new BioSequence class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioSequence$new(
  name = NULL,
  description = NULL,
  id = NULL,
  length = NULL,
  sequence = NULL,
  isApproximateLength = NULL,
  isCircular = NULL,
  polymerType = NULL,
  type = NULL,
  fractionRepeats = NULL,
  sequenceDatabaseEntry = NULL,
  taxon = NULL,
  bioSequence2GeneProduct = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}

\item{\code{description}}{description}

\item{\code{id}}{id}

\item{\code{length}}{length}

\item{\code{sequence}}{sequence}

\item{\code{isApproximateLength}}{isApproximateLength}

\item{\code{isCircular}}{isCircular}

\item{\code{polymerType}}{polymerType}

\item{\code{type}}{type}

\item{\code{fractionRepeats}}{fractionRepeats}

\item{\code{sequenceDatabaseEntry}}{sequenceDatabaseEntry}

\item{\code{taxon}}{taxon}

\item{\code{bioSequence2GeneProduct}}{bioSequence2GeneProduct}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioSequence-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BioSequence-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioSequence$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
BioSequence in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioSequence-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BioSequence-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of BioSequence
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioSequence$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of BioSequence
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioSequence-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-BioSequence-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioSequence$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
BioSequence in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioSequence-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-BioSequence-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of BioSequence
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioSequence$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of BioSequence
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioSequence-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BioSequence-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to BioSequence and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioSequence$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioSequence-toString"></a>}}
\if{latex}{\out{\hypertarget{method-BioSequence-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioSequence$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of BioSequence
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioSequence-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-BioSequence-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioSequence$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioSequence-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-BioSequence-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioSequence$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioSequence-print"></a>}}
\if{latex}{\out{\hypertarget{method-BioSequence-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioSequence$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioSequence-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BioSequence-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BioSequence$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
