% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subProcessors.R
\name{blank_processor}
\alias{blank_processor}
\title{A blank processor that returns data as is}
\usage{
blank_processor(data)
}
\arguments{
\item{data}{any data}
}
\value{
Data as is
}
\description{
A blank processor that returns data as is
}
\keyword{internal}
