% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic.R
\docType{class}
\name{Characteristic}
\alias{Characteristic}
\title{Characteristic}
\format{
An \code{R6Class} generator object
}
\description{
Characteristic Class

Characteristic Class
}
\details{
Create a new Characteristic
}
\keyword{(JSON}
\keyword{(if}
\keyword{Characteristic}
\keyword{Deserialize}
\keyword{Initialize}
\keyword{JSON}
\keyword{Return}
\keyword{To}
\keyword{Validate}
\keyword{a}
\keyword{all}
\keyword{an}
\keyword{any).}
\keyword{are}
\keyword{class.}
\keyword{fields}
\keyword{format)}
\keyword{if}
\keyword{in}
\keyword{input}
\keyword{instance}
\keyword{internal}
\keyword{into}
\keyword{invalid}
\keyword{list}
\keyword{new}
\keyword{of}
\keyword{respect}
\keyword{string}
\keyword{the}
\keyword{to}
\keyword{true}
\keyword{valid.}
\keyword{values}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character \link{optional}}

\item{\code{description}}{character \link{optional}}

\item{\code{id}}{integer \link{optional}}

\item{\code{category}}{character \link{optional}}

\item{\code{categoryUri}}{character \link{optional}}

\item{\code{evidenceCode}}{\link{GOEvidenceCode} \link{optional}}

\item{\code{originalValue}}{character \link{optional}}

\item{\code{value}}{character \link{optional}}

\item{\code{valueUri}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Characteristic-new}{\code{Characteristic$new()}}
\item \href{#method-Characteristic-toJSON}{\code{Characteristic$toJSON()}}
\item \href{#method-Characteristic-fromJSON}{\code{Characteristic$fromJSON()}}
\item \href{#method-Characteristic-toJSONString}{\code{Characteristic$toJSONString()}}
\item \href{#method-Characteristic-fromJSONString}{\code{Characteristic$fromJSONString()}}
\item \href{#method-Characteristic-validateJSON}{\code{Characteristic$validateJSON()}}
\item \href{#method-Characteristic-toString}{\code{Characteristic$toString()}}
\item \href{#method-Characteristic-isValid}{\code{Characteristic$isValid()}}
\item \href{#method-Characteristic-getInvalidFields}{\code{Characteristic$getInvalidFields()}}
\item \href{#method-Characteristic-print}{\code{Characteristic$print()}}
\item \href{#method-Characteristic-clone}{\code{Characteristic$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-new"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Characteristic class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$new(
  name = NULL,
  description = NULL,
  id = NULL,
  category = NULL,
  categoryUri = NULL,
  evidenceCode = NULL,
  originalValue = NULL,
  value = NULL,
  valueUri = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}

\item{\code{description}}{description}

\item{\code{id}}{id}

\item{\code{category}}{category}

\item{\code{categoryUri}}{categoryUri}

\item{\code{evidenceCode}}{evidenceCode}

\item{\code{originalValue}}{originalValue}

\item{\code{value}}{value}

\item{\code{valueUri}}{valueUri}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Characteristic in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Characteristic
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Characteristic
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Characteristic in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Characteristic
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Characteristic
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Characteristic and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Characteristic
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-print"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
