% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_value_object.R
\docType{class}
\name{CharacteristicValueObject}
\alias{CharacteristicValueObject}
\title{CharacteristicValueObject}
\format{
An \code{R6Class} generator object
}
\description{
CharacteristicValueObject Class

CharacteristicValueObject Class
}
\details{
Create a new CharacteristicValueObject
}
\keyword{(JSON}
\keyword{(if}
\keyword{CharacteristicValueObject}
\keyword{Deserialize}
\keyword{Initialize}
\keyword{JSON}
\keyword{Return}
\keyword{To}
\keyword{Validate}
\keyword{a}
\keyword{all}
\keyword{an}
\keyword{any).}
\keyword{are}
\keyword{class.}
\keyword{fields}
\keyword{format)}
\keyword{if}
\keyword{in}
\keyword{input}
\keyword{instance}
\keyword{internal}
\keyword{into}
\keyword{invalid}
\keyword{list}
\keyword{new}
\keyword{of}
\keyword{respect}
\keyword{string}
\keyword{the}
\keyword{to}
\keyword{true}
\keyword{valid.}
\keyword{values}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer \link{optional}}

\item{\code{urlId}}{character \link{optional}}

\item{\code{alreadyPresentInDatabase}}{character \link{optional}}

\item{\code{alreadyPresentOnGene}}{character \link{optional}}

\item{\code{category}}{character \link{optional}}

\item{\code{categoryUri}}{character \link{optional}}

\item{\code{child}}{character \link{optional}}

\item{\code{numTimesUsed}}{integer \link{optional}}

\item{\code{ontologyUsed}}{character \link{optional}}

\item{\code{privateGeneCount}}{integer \link{optional}}

\item{\code{publicGeneCount}}{integer \link{optional}}

\item{\code{root}}{character \link{optional}}

\item{\code{taxon}}{character \link{optional}}

\item{\code{value}}{character \link{optional}}

\item{\code{valueUri}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CharacteristicValueObject-new}{\code{CharacteristicValueObject$new()}}
\item \href{#method-CharacteristicValueObject-toJSON}{\code{CharacteristicValueObject$toJSON()}}
\item \href{#method-CharacteristicValueObject-fromJSON}{\code{CharacteristicValueObject$fromJSON()}}
\item \href{#method-CharacteristicValueObject-toJSONString}{\code{CharacteristicValueObject$toJSONString()}}
\item \href{#method-CharacteristicValueObject-fromJSONString}{\code{CharacteristicValueObject$fromJSONString()}}
\item \href{#method-CharacteristicValueObject-validateJSON}{\code{CharacteristicValueObject$validateJSON()}}
\item \href{#method-CharacteristicValueObject-toString}{\code{CharacteristicValueObject$toString()}}
\item \href{#method-CharacteristicValueObject-isValid}{\code{CharacteristicValueObject$isValid()}}
\item \href{#method-CharacteristicValueObject-getInvalidFields}{\code{CharacteristicValueObject$getInvalidFields()}}
\item \href{#method-CharacteristicValueObject-print}{\code{CharacteristicValueObject$print()}}
\item \href{#method-CharacteristicValueObject-clone}{\code{CharacteristicValueObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicValueObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicValueObject-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new CharacteristicValueObject class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicValueObject$new(
  id = NULL,
  urlId = NULL,
  alreadyPresentInDatabase = NULL,
  alreadyPresentOnGene = NULL,
  category = NULL,
  categoryUri = NULL,
  child = NULL,
  numTimesUsed = NULL,
  ontologyUsed = NULL,
  privateGeneCount = NULL,
  publicGeneCount = NULL,
  root = NULL,
  taxon = NULL,
  value = NULL,
  valueUri = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{urlId}}{urlId}

\item{\code{alreadyPresentInDatabase}}{alreadyPresentInDatabase}

\item{\code{alreadyPresentOnGene}}{alreadyPresentOnGene}

\item{\code{category}}{category}

\item{\code{categoryUri}}{categoryUri}

\item{\code{child}}{child}

\item{\code{numTimesUsed}}{numTimesUsed}

\item{\code{ontologyUsed}}{ontologyUsed}

\item{\code{privateGeneCount}}{privateGeneCount}

\item{\code{publicGeneCount}}{publicGeneCount}

\item{\code{root}}{root}

\item{\code{taxon}}{taxon}

\item{\code{value}}{value}

\item{\code{valueUri}}{valueUri}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicValueObject-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicValueObject-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicValueObject$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CharacteristicValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicValueObject-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicValueObject-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of CharacteristicValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicValueObject$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of CharacteristicValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicValueObject-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicValueObject-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicValueObject$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CharacteristicValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicValueObject-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicValueObject-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of CharacteristicValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicValueObject$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of CharacteristicValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicValueObject-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicValueObject-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to CharacteristicValueObject and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicValueObject$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicValueObject-toString"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicValueObject-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicValueObject$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of CharacteristicValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicValueObject-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicValueObject-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicValueObject$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicValueObject-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicValueObject-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicValueObject$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicValueObject-print"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicValueObject-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicValueObject$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicValueObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicValueObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicValueObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
