% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{checkBounds}
\alias{checkBounds}
\title{Replace missing data with NAs}
\usage{
checkBounds(x)
}
\arguments{
\item{x}{Data}
}
\value{
Data or NA in case of an out of bounds error
}
\description{
Replace missing data with NAs
}
\keyword{internal}
