% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openapi_generated.R
\docType{class}
\name{ContrastResultValueObject}
\alias{ContrastResultValueObject}
\title{ContrastResultValueObject}
\format{
An \code{R6Class} generator object
}
\description{
ContrastResultValueObject Class

ContrastResultValueObject Class
}
\details{
Create a new ContrastResultValueObject
}
\keyword{(JSON}
\keyword{(if}
\keyword{ContrastResultValueObject}
\keyword{Deserialize}
\keyword{Initialize}
\keyword{JSON}
\keyword{Return}
\keyword{To}
\keyword{Validate}
\keyword{a}
\keyword{all}
\keyword{an}
\keyword{any).}
\keyword{are}
\keyword{class.}
\keyword{fields}
\keyword{format)}
\keyword{if}
\keyword{in}
\keyword{input}
\keyword{instance}
\keyword{internal}
\keyword{into}
\keyword{invalid}
\keyword{list}
\keyword{new}
\keyword{of}
\keyword{respect}
\keyword{string}
\keyword{the}
\keyword{to}
\keyword{true}
\keyword{valid.}
\keyword{values}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pvalue}}{numeric optional}

\item{\code{coefficient}}{numeric optional}

\item{\code{logFoldChange}}{numeric optional}

\item{\code{factorValue}}{\link{FactorValueBasicValueObject} optional}

\item{\code{secondFactorValue}}{\link{FactorValueBasicValueObject} optional}

\item{\code{tstat}}{numeric optional}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ContrastResultValueObject-new}{\code{ContrastResultValueObject$new()}}
\item \href{#method-ContrastResultValueObject-toJSON}{\code{ContrastResultValueObject$toJSON()}}
\item \href{#method-ContrastResultValueObject-fromJSON}{\code{ContrastResultValueObject$fromJSON()}}
\item \href{#method-ContrastResultValueObject-toJSONString}{\code{ContrastResultValueObject$toJSONString()}}
\item \href{#method-ContrastResultValueObject-fromJSONString}{\code{ContrastResultValueObject$fromJSONString()}}
\item \href{#method-ContrastResultValueObject-validateJSON}{\code{ContrastResultValueObject$validateJSON()}}
\item \href{#method-ContrastResultValueObject-toString}{\code{ContrastResultValueObject$toString()}}
\item \href{#method-ContrastResultValueObject-isValid}{\code{ContrastResultValueObject$isValid()}}
\item \href{#method-ContrastResultValueObject-getInvalidFields}{\code{ContrastResultValueObject$getInvalidFields()}}
\item \href{#method-ContrastResultValueObject-print}{\code{ContrastResultValueObject$print()}}
\item \href{#method-ContrastResultValueObject-clone}{\code{ContrastResultValueObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContrastResultValueObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-ContrastResultValueObject-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ContrastResultValueObject class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContrastResultValueObject$new(
  pvalue = NULL,
  coefficient = NULL,
  logFoldChange = NULL,
  factorValue = NULL,
  secondFactorValue = NULL,
  tstat = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pvalue}}{pvalue}

\item{\code{coefficient}}{coefficient}

\item{\code{logFoldChange}}{logFoldChange}

\item{\code{factorValue}}{factorValue}

\item{\code{secondFactorValue}}{secondFactorValue}

\item{\code{tstat}}{tstat}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContrastResultValueObject-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ContrastResultValueObject-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContrastResultValueObject$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ContrastResultValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContrastResultValueObject-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ContrastResultValueObject-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ContrastResultValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContrastResultValueObject$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ContrastResultValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContrastResultValueObject-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ContrastResultValueObject-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContrastResultValueObject$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ContrastResultValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContrastResultValueObject-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ContrastResultValueObject-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ContrastResultValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContrastResultValueObject$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ContrastResultValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContrastResultValueObject-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ContrastResultValueObject-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ContrastResultValueObject and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContrastResultValueObject$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContrastResultValueObject-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ContrastResultValueObject-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContrastResultValueObject$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ContrastResultValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContrastResultValueObject-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ContrastResultValueObject-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContrastResultValueObject$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContrastResultValueObject-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ContrastResultValueObject-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContrastResultValueObject$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContrastResultValueObject-print"></a>}}
\if{latex}{\out{\hypertarget{method-ContrastResultValueObject-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContrastResultValueObject$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContrastResultValueObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ContrastResultValueObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContrastResultValueObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
