% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curation_details.R
\docType{class}
\name{CurationDetails}
\alias{CurationDetails}
\title{CurationDetails}
\format{
An \code{R6Class} generator object
}
\description{
CurationDetails Class

CurationDetails Class
}
\details{
Create a new CurationDetails
}
\keyword{(JSON}
\keyword{(if}
\keyword{CurationDetails}
\keyword{Deserialize}
\keyword{Initialize}
\keyword{JSON}
\keyword{Return}
\keyword{To}
\keyword{Validate}
\keyword{a}
\keyword{all}
\keyword{an}
\keyword{any).}
\keyword{are}
\keyword{class.}
\keyword{fields}
\keyword{format)}
\keyword{if}
\keyword{in}
\keyword{input}
\keyword{instance}
\keyword{internal}
\keyword{into}
\keyword{invalid}
\keyword{list}
\keyword{new}
\keyword{of}
\keyword{respect}
\keyword{string}
\keyword{the}
\keyword{to}
\keyword{true}
\keyword{valid.}
\keyword{values}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer \link{optional}}

\item{\code{lastUpdated}}{character \link{optional}}

\item{\code{lastNeedsAttentionEvent}}{\link{AuditEvent} \link{optional}}

\item{\code{needsAttention}}{character \link{optional}}

\item{\code{lastTroubledEvent}}{\link{AuditEvent} \link{optional}}

\item{\code{troubled}}{character \link{optional}}

\item{\code{lastNoteUpdateEvent}}{\link{AuditEvent} \link{optional}}

\item{\code{curationNote}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CurationDetails-new}{\code{CurationDetails$new()}}
\item \href{#method-CurationDetails-toJSON}{\code{CurationDetails$toJSON()}}
\item \href{#method-CurationDetails-fromJSON}{\code{CurationDetails$fromJSON()}}
\item \href{#method-CurationDetails-toJSONString}{\code{CurationDetails$toJSONString()}}
\item \href{#method-CurationDetails-fromJSONString}{\code{CurationDetails$fromJSONString()}}
\item \href{#method-CurationDetails-validateJSON}{\code{CurationDetails$validateJSON()}}
\item \href{#method-CurationDetails-toString}{\code{CurationDetails$toString()}}
\item \href{#method-CurationDetails-isValid}{\code{CurationDetails$isValid()}}
\item \href{#method-CurationDetails-getInvalidFields}{\code{CurationDetails$getInvalidFields()}}
\item \href{#method-CurationDetails-print}{\code{CurationDetails$print()}}
\item \href{#method-CurationDetails-clone}{\code{CurationDetails$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CurationDetails-new"></a>}}
\if{latex}{\out{\hypertarget{method-CurationDetails-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new CurationDetails class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CurationDetails$new(
  id = NULL,
  lastUpdated = NULL,
  lastNeedsAttentionEvent = NULL,
  needsAttention = NULL,
  lastTroubledEvent = NULL,
  troubled = NULL,
  lastNoteUpdateEvent = NULL,
  curationNote = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{lastUpdated}}{lastUpdated}

\item{\code{lastNeedsAttentionEvent}}{lastNeedsAttentionEvent}

\item{\code{needsAttention}}{needsAttention}

\item{\code{lastTroubledEvent}}{lastTroubledEvent}

\item{\code{troubled}}{troubled}

\item{\code{lastNoteUpdateEvent}}{lastNoteUpdateEvent}

\item{\code{curationNote}}{curationNote}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CurationDetails-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CurationDetails-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CurationDetails$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CurationDetails in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CurationDetails-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CurationDetails-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of CurationDetails
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CurationDetails$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of CurationDetails
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CurationDetails-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CurationDetails-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CurationDetails$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CurationDetails in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CurationDetails-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-CurationDetails-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of CurationDetails
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CurationDetails$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of CurationDetails
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CurationDetails-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-CurationDetails-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to CurationDetails and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CurationDetails$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CurationDetails-toString"></a>}}
\if{latex}{\out{\hypertarget{method-CurationDetails-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CurationDetails$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of CurationDetails
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CurationDetails-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-CurationDetails-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CurationDetails$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CurationDetails-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-CurationDetails-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CurationDetails$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CurationDetails-print"></a>}}
\if{latex}{\out{\hypertarget{method-CurationDetails-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CurationDetails$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CurationDetails-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CurationDetails-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CurationDetails$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
