% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_entry.R
\docType{class}
\name{DatabaseEntry}
\alias{DatabaseEntry}
\title{DatabaseEntry}
\format{
An \code{R6Class} generator object
}
\description{
DatabaseEntry Class

DatabaseEntry Class
}
\details{
Create a new DatabaseEntry
}
\keyword{(JSON}
\keyword{(if}
\keyword{DatabaseEntry}
\keyword{Deserialize}
\keyword{Initialize}
\keyword{JSON}
\keyword{Return}
\keyword{To}
\keyword{Validate}
\keyword{a}
\keyword{all}
\keyword{an}
\keyword{any).}
\keyword{are}
\keyword{class.}
\keyword{fields}
\keyword{format)}
\keyword{if}
\keyword{in}
\keyword{input}
\keyword{instance}
\keyword{internal}
\keyword{into}
\keyword{invalid}
\keyword{list}
\keyword{new}
\keyword{of}
\keyword{respect}
\keyword{string}
\keyword{the}
\keyword{to}
\keyword{true}
\keyword{valid.}
\keyword{values}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{accession}}{character \link{optional}}

\item{\code{accessionVersion}}{character \link{optional}}

\item{\code{id}}{integer \link{optional}}

\item{\code{externalDatabase}}{\link{ExternalDatabase} \link{optional}}

\item{\code{uri}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DatabaseEntry-new}{\code{DatabaseEntry$new()}}
\item \href{#method-DatabaseEntry-toJSON}{\code{DatabaseEntry$toJSON()}}
\item \href{#method-DatabaseEntry-fromJSON}{\code{DatabaseEntry$fromJSON()}}
\item \href{#method-DatabaseEntry-toJSONString}{\code{DatabaseEntry$toJSONString()}}
\item \href{#method-DatabaseEntry-fromJSONString}{\code{DatabaseEntry$fromJSONString()}}
\item \href{#method-DatabaseEntry-validateJSON}{\code{DatabaseEntry$validateJSON()}}
\item \href{#method-DatabaseEntry-toString}{\code{DatabaseEntry$toString()}}
\item \href{#method-DatabaseEntry-isValid}{\code{DatabaseEntry$isValid()}}
\item \href{#method-DatabaseEntry-getInvalidFields}{\code{DatabaseEntry$getInvalidFields()}}
\item \href{#method-DatabaseEntry-print}{\code{DatabaseEntry$print()}}
\item \href{#method-DatabaseEntry-clone}{\code{DatabaseEntry$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseEntry-new"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseEntry-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DatabaseEntry class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseEntry$new(
  accession = NULL,
  accessionVersion = NULL,
  id = NULL,
  externalDatabase = NULL,
  uri = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{accession}}{accession}

\item{\code{accessionVersion}}{accessionVersion}

\item{\code{id}}{id}

\item{\code{externalDatabase}}{externalDatabase}

\item{\code{uri}}{uri}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseEntry-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseEntry-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseEntry$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
DatabaseEntry in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseEntry-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseEntry-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of DatabaseEntry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseEntry$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of DatabaseEntry
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseEntry-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseEntry-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseEntry$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
DatabaseEntry in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseEntry-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseEntry-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of DatabaseEntry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseEntry$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of DatabaseEntry
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseEntry-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseEntry-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to DatabaseEntry and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseEntry$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseEntry-toString"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseEntry-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseEntry$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of DatabaseEntry
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseEntry-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseEntry-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseEntry$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseEntry-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseEntry-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseEntry$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseEntry-print"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseEntry-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseEntry$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseEntry-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseEntry-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseEntry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
