% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openapi_generated.R
\docType{class}
\name{DatabaseEntryArg}
\alias{DatabaseEntryArg}
\title{DatabaseEntryArg}
\format{
An \code{R6Class} generator object
}
\description{
DatabaseEntryArg Class
}
\keyword{DatabaseEntryArg}
\keyword{DatabaseEntryArg.}
\keyword{Deserialize}
\keyword{JSON}
\keyword{JSON.}
\keyword{Returns}
\keyword{Serialize}
\keyword{Validate}
\keyword{an}
\keyword{input}
\keyword{instance}
\keyword{instance.}
\keyword{internal}
\keyword{into}
\keyword{of}
\keyword{representation}
\keyword{respect}
\keyword{string}
\keyword{string.}
\keyword{the}
\keyword{to}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new DatabaseEntryArg.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DatabaseEntryArg-new}{\code{DatabaseEntryArg$new()}}
\item \href{#method-DatabaseEntryArg-fromJSONString}{\code{DatabaseEntryArg$fromJSONString()}}
\item \href{#method-DatabaseEntryArg-fromJSON}{\code{DatabaseEntryArg$fromJSON()}}
\item \href{#method-DatabaseEntryArg-toJSONString}{\code{DatabaseEntryArg$toJSONString()}}
\item \href{#method-DatabaseEntryArg-toJSON}{\code{DatabaseEntryArg$toJSON()}}
\item \href{#method-DatabaseEntryArg-validateJSON}{\code{DatabaseEntryArg$validateJSON()}}
\item \href{#method-DatabaseEntryArg-toString}{\code{DatabaseEntryArg$toString()}}
\item \href{#method-DatabaseEntryArg-print}{\code{DatabaseEntryArg$print()}}
\item \href{#method-DatabaseEntryArg-clone}{\code{DatabaseEntryArg$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseEntryArg-new"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseEntryArg-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DatabaseEntryArg.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseEntryArg$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "character", "integer"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseEntryArg-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseEntryArg-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of DatabaseEntryArg.
An alias to the method \code{fromJSON} .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseEntryArg$fromJSONString(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of DatabaseEntryArg.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseEntryArg-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseEntryArg-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of DatabaseEntryArg.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseEntryArg$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of DatabaseEntryArg.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseEntryArg-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseEntryArg-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize DatabaseEntryArg to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseEntryArg$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the DatabaseEntryArg.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseEntryArg-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseEntryArg-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize DatabaseEntryArg to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseEntryArg$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the DatabaseEntryArg.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseEntryArg-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseEntryArg-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to DatabaseEntryArg and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseEntryArg$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseEntryArg-toString"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseEntryArg-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseEntryArg$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseEntryArg-print"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseEntryArg-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseEntryArg$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseEntryArg-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseEntryArg-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseEntryArg$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
