% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{encode}
\alias{encode}
\title{URL encode a string safely}
\usage{
encode(url)
}
\arguments{
\item{url}{The string to URL encode. Vectors are delimited by a comma.}
}
\value{
A URL encoding of url
}
\description{
URL encode a string safely
}
\keyword{internal}
