% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental_factor_value_object.R
\docType{class}
\name{ExperimentalFactorValueObject}
\alias{ExperimentalFactorValueObject}
\title{ExperimentalFactorValueObject}
\format{
An \code{R6Class} generator object
}
\description{
ExperimentalFactorValueObject Class

ExperimentalFactorValueObject Class
}
\details{
Create a new ExperimentalFactorValueObject
}
\keyword{(JSON}
\keyword{(if}
\keyword{Deserialize}
\keyword{ExperimentalFactorValueObject}
\keyword{Initialize}
\keyword{JSON}
\keyword{Return}
\keyword{To}
\keyword{Validate}
\keyword{a}
\keyword{all}
\keyword{an}
\keyword{any).}
\keyword{are}
\keyword{class.}
\keyword{fields}
\keyword{format)}
\keyword{if}
\keyword{in}
\keyword{input}
\keyword{instance}
\keyword{internal}
\keyword{into}
\keyword{invalid}
\keyword{list}
\keyword{new}
\keyword{of}
\keyword{respect}
\keyword{string}
\keyword{the}
\keyword{to}
\keyword{true}
\keyword{valid.}
\keyword{values}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer optional}

\item{\code{category}}{character optional}

\item{\code{categoryUri}}{character optional}

\item{\code{description}}{character optional}

\item{\code{factorValues}}{character optional}

\item{\code{name}}{character optional}

\item{\code{numValues}}{integer optional}

\item{\code{type}}{character optional}

\item{\code{values}}{list(\link{FactorValueValueObject}) optional}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExperimentalFactorValueObject-new}{\code{ExperimentalFactorValueObject$new()}}
\item \href{#method-ExperimentalFactorValueObject-toJSON}{\code{ExperimentalFactorValueObject$toJSON()}}
\item \href{#method-ExperimentalFactorValueObject-fromJSON}{\code{ExperimentalFactorValueObject$fromJSON()}}
\item \href{#method-ExperimentalFactorValueObject-toJSONString}{\code{ExperimentalFactorValueObject$toJSONString()}}
\item \href{#method-ExperimentalFactorValueObject-fromJSONString}{\code{ExperimentalFactorValueObject$fromJSONString()}}
\item \href{#method-ExperimentalFactorValueObject-validateJSON}{\code{ExperimentalFactorValueObject$validateJSON()}}
\item \href{#method-ExperimentalFactorValueObject-toString}{\code{ExperimentalFactorValueObject$toString()}}
\item \href{#method-ExperimentalFactorValueObject-isValid}{\code{ExperimentalFactorValueObject$isValid()}}
\item \href{#method-ExperimentalFactorValueObject-getInvalidFields}{\code{ExperimentalFactorValueObject$getInvalidFields()}}
\item \href{#method-ExperimentalFactorValueObject-print}{\code{ExperimentalFactorValueObject$print()}}
\item \href{#method-ExperimentalFactorValueObject-clone}{\code{ExperimentalFactorValueObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExperimentalFactorValueObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExperimentalFactorValueObject-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ExperimentalFactorValueObject class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentalFactorValueObject$new(
  id = NULL,
  category = NULL,
  categoryUri = NULL,
  description = NULL,
  factorValues = NULL,
  name = NULL,
  numValues = NULL,
  type = NULL,
  values = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{category}}{category}

\item{\code{categoryUri}}{categoryUri}

\item{\code{description}}{description}

\item{\code{factorValues}}{factorValues}

\item{\code{name}}{name}

\item{\code{numValues}}{numValues}

\item{\code{type}}{type}

\item{\code{values}}{values}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExperimentalFactorValueObject-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ExperimentalFactorValueObject-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentalFactorValueObject$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ExperimentalFactorValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExperimentalFactorValueObject-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ExperimentalFactorValueObject-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ExperimentalFactorValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentalFactorValueObject$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ExperimentalFactorValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExperimentalFactorValueObject-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ExperimentalFactorValueObject-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentalFactorValueObject$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ExperimentalFactorValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExperimentalFactorValueObject-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ExperimentalFactorValueObject-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ExperimentalFactorValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentalFactorValueObject$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ExperimentalFactorValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExperimentalFactorValueObject-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ExperimentalFactorValueObject-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ExperimentalFactorValueObject and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentalFactorValueObject$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExperimentalFactorValueObject-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ExperimentalFactorValueObject-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentalFactorValueObject$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ExperimentalFactorValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExperimentalFactorValueObject-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ExperimentalFactorValueObject-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentalFactorValueObject$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExperimentalFactorValueObject-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ExperimentalFactorValueObject-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentalFactorValueObject$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExperimentalFactorValueObject-print"></a>}}
\if{latex}{\out{\hypertarget{method-ExperimentalFactorValueObject-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentalFactorValueObject$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExperimentalFactorValueObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExperimentalFactorValueObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentalFactorValueObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
