% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_database.R
\docType{class}
\name{ExternalDatabase}
\alias{ExternalDatabase}
\title{ExternalDatabase}
\format{
An \code{R6Class} generator object
}
\description{
ExternalDatabase Class

ExternalDatabase Class
}
\details{
Create a new ExternalDatabase
}
\keyword{(JSON}
\keyword{(if}
\keyword{Deserialize}
\keyword{ExternalDatabase}
\keyword{Initialize}
\keyword{JSON}
\keyword{Return}
\keyword{To}
\keyword{Validate}
\keyword{a}
\keyword{all}
\keyword{an}
\keyword{any).}
\keyword{are}
\keyword{class.}
\keyword{fields}
\keyword{format)}
\keyword{if}
\keyword{in}
\keyword{input}
\keyword{instance}
\keyword{internal}
\keyword{into}
\keyword{invalid}
\keyword{list}
\keyword{new}
\keyword{of}
\keyword{respect}
\keyword{string}
\keyword{the}
\keyword{to}
\keyword{true}
\keyword{valid.}
\keyword{values}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character optional}

\item{\code{description}}{character optional}

\item{\code{id}}{integer optional}

\item{\code{localInstallDbName}}{character optional}

\item{\code{webUri}}{character optional}

\item{\code{ftpUri}}{character optional}

\item{\code{type}}{\link{DatabaseType} optional}

\item{\code{databaseSupplier}}{\link{Contact} optional}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExternalDatabase-new}{\code{ExternalDatabase$new()}}
\item \href{#method-ExternalDatabase-toJSON}{\code{ExternalDatabase$toJSON()}}
\item \href{#method-ExternalDatabase-fromJSON}{\code{ExternalDatabase$fromJSON()}}
\item \href{#method-ExternalDatabase-toJSONString}{\code{ExternalDatabase$toJSONString()}}
\item \href{#method-ExternalDatabase-fromJSONString}{\code{ExternalDatabase$fromJSONString()}}
\item \href{#method-ExternalDatabase-validateJSON}{\code{ExternalDatabase$validateJSON()}}
\item \href{#method-ExternalDatabase-toString}{\code{ExternalDatabase$toString()}}
\item \href{#method-ExternalDatabase-isValid}{\code{ExternalDatabase$isValid()}}
\item \href{#method-ExternalDatabase-getInvalidFields}{\code{ExternalDatabase$getInvalidFields()}}
\item \href{#method-ExternalDatabase-print}{\code{ExternalDatabase$print()}}
\item \href{#method-ExternalDatabase-clone}{\code{ExternalDatabase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalDatabase-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalDatabase-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ExternalDatabase class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalDatabase$new(
  name = NULL,
  description = NULL,
  id = NULL,
  localInstallDbName = NULL,
  webUri = NULL,
  ftpUri = NULL,
  type = NULL,
  databaseSupplier = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}

\item{\code{description}}{description}

\item{\code{id}}{id}

\item{\code{localInstallDbName}}{localInstallDbName}

\item{\code{webUri}}{webUri}

\item{\code{ftpUri}}{ftpUri}

\item{\code{type}}{type}

\item{\code{databaseSupplier}}{databaseSupplier}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalDatabase-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalDatabase-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalDatabase$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ExternalDatabase in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalDatabase-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalDatabase-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ExternalDatabase
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalDatabase$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ExternalDatabase
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalDatabase-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalDatabase-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalDatabase$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ExternalDatabase in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalDatabase-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalDatabase-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ExternalDatabase
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalDatabase$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ExternalDatabase
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalDatabase-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalDatabase-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ExternalDatabase and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalDatabase$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalDatabase-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalDatabase-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalDatabase$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ExternalDatabase
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalDatabase-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalDatabase-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalDatabase$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalDatabase-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalDatabase-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalDatabase$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalDatabase-print"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalDatabase-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalDatabase$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalDatabase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalDatabase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalDatabase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
