% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openapi_generated.R
\docType{class}
\name{ExternalDatabaseValueObject}
\alias{ExternalDatabaseValueObject}
\title{ExternalDatabaseValueObject}
\format{
An \code{R6Class} generator object
}
\description{
ExternalDatabaseValueObject Class

ExternalDatabaseValueObject Class
}
\details{
Create a new ExternalDatabaseValueObject
}
\keyword{(JSON}
\keyword{(if}
\keyword{Deserialize}
\keyword{ExternalDatabaseValueObject}
\keyword{Initialize}
\keyword{JSON}
\keyword{Return}
\keyword{To}
\keyword{Validate}
\keyword{a}
\keyword{all}
\keyword{an}
\keyword{any).}
\keyword{are}
\keyword{class.}
\keyword{fields}
\keyword{format)}
\keyword{if}
\keyword{in}
\keyword{input}
\keyword{instance}
\keyword{internal}
\keyword{into}
\keyword{invalid}
\keyword{list}
\keyword{new}
\keyword{of}
\keyword{respect}
\keyword{string}
\keyword{the}
\keyword{to}
\keyword{true}
\keyword{valid.}
\keyword{values}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character optional}

\item{\code{id}}{integer optional}

\item{\code{checked}}{character optional}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExternalDatabaseValueObject-new}{\code{ExternalDatabaseValueObject$new()}}
\item \href{#method-ExternalDatabaseValueObject-toJSON}{\code{ExternalDatabaseValueObject$toJSON()}}
\item \href{#method-ExternalDatabaseValueObject-fromJSON}{\code{ExternalDatabaseValueObject$fromJSON()}}
\item \href{#method-ExternalDatabaseValueObject-toJSONString}{\code{ExternalDatabaseValueObject$toJSONString()}}
\item \href{#method-ExternalDatabaseValueObject-fromJSONString}{\code{ExternalDatabaseValueObject$fromJSONString()}}
\item \href{#method-ExternalDatabaseValueObject-validateJSON}{\code{ExternalDatabaseValueObject$validateJSON()}}
\item \href{#method-ExternalDatabaseValueObject-toString}{\code{ExternalDatabaseValueObject$toString()}}
\item \href{#method-ExternalDatabaseValueObject-isValid}{\code{ExternalDatabaseValueObject$isValid()}}
\item \href{#method-ExternalDatabaseValueObject-getInvalidFields}{\code{ExternalDatabaseValueObject$getInvalidFields()}}
\item \href{#method-ExternalDatabaseValueObject-print}{\code{ExternalDatabaseValueObject$print()}}
\item \href{#method-ExternalDatabaseValueObject-clone}{\code{ExternalDatabaseValueObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalDatabaseValueObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalDatabaseValueObject-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ExternalDatabaseValueObject class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalDatabaseValueObject$new(name = NULL, id = NULL, checked = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}

\item{\code{id}}{id}

\item{\code{checked}}{checked}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalDatabaseValueObject-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalDatabaseValueObject-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalDatabaseValueObject$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ExternalDatabaseValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalDatabaseValueObject-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalDatabaseValueObject-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ExternalDatabaseValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalDatabaseValueObject$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ExternalDatabaseValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalDatabaseValueObject-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalDatabaseValueObject-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalDatabaseValueObject$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ExternalDatabaseValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalDatabaseValueObject-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalDatabaseValueObject-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ExternalDatabaseValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalDatabaseValueObject$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ExternalDatabaseValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalDatabaseValueObject-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalDatabaseValueObject-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ExternalDatabaseValueObject and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalDatabaseValueObject$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalDatabaseValueObject-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalDatabaseValueObject-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalDatabaseValueObject$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ExternalDatabaseValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalDatabaseValueObject-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalDatabaseValueObject-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalDatabaseValueObject$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalDatabaseValueObject-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalDatabaseValueObject-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalDatabaseValueObject$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalDatabaseValueObject-print"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalDatabaseValueObject-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalDatabaseValueObject$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExternalDatabaseValueObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExternalDatabaseValueObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExternalDatabaseValueObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
