% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{filter_properties}
\alias{filter_properties}
\title{Return all supported filter properties}
\usage{
filter_properties()
}
\value{
A list of data.tables that contain supported properties and their data
types
}
\description{
Some functions such as \code{\link{get_datasets}} and \code{\link{get_platforms_by_ids}}
include a filter argument that allows creation of more complex queries. This
function returns a list of supported properties to be used in those filters
}
\examples{
filter_properties()

}
\keyword{misc}
