% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{forget_gemma_memoised}
\alias{forget_gemma_memoised}
\title{Clear gemma.R cache}
\usage{
forget_gemma_memoised()
}
\value{
TRUE to indicate cache was cleared.
}
\description{
Forget past results from memoised calls to the Gemma API (ie. using functions with memoised = \code{TRUE})
}
\examples{
forget_gemma_memoised()
}
\keyword{misc}
