% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable.R
\name{gemma_kable}
\alias{gemma_kable}
\title{Create printable tables out of gemma.R outputs}
\usage{
gemma_kable(table)
}
\arguments{
\item{table}{A data.table or data.frame outputted by a gemma.R function}
}
\description{
Creates a \code{\link[knitr]{kable}} where certain columns are automatically
shortened to better fit a document.
}
