% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_caching.R
\name{gemma_memoise}
\alias{gemma_memoise}
\title{Enable and disable memoisation of gemma.R functions}
\usage{
gemma_memoise(memoised = NULL)
}
\arguments{
\item{memoised}{boolean. If TRUE memoisation will be enabled}
}
\value{
Option set for memoisation
}
\description{
Enable and disable memoisation of gemma.R functions
}
\keyword{internal}
