% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_convenience.R
\name{gemmaCall}
\alias{gemmaCall}
\title{Custom gemma call}
\usage{
gemmaCall(call, ..., json = TRUE)
}
\arguments{
\item{call}{Gemma API endpoint.}

\item{...}{parameters included in the call}

\item{json}{If \code{TRUE} will parse the content as a list}
}
\value{
A list if \code{json = TRUE} and an httr response if \code{FALSE}
}
\description{
A minimal function to create custom calls. Can be used to acquire unimplemented
endpoints and/or raw output without any processing. Refer to the
\href{https://gemma.msl.ubc.ca/resources/restapidocs/}{API documentation}.
}
\examples{
# get singular value decomposition for the dataset
gemmaCall('datasets/{dataset}/svd',dataset = 1)
}
\keyword{misc}
