% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\docType{class}
\name{Gene}
\alias{Gene}
\title{Gene}
\format{
An \code{R6Class} generator object
}
\description{
Gene Class

Gene Class
}
\details{
Create a new Gene
}
\keyword{(JSON}
\keyword{(if}
\keyword{Deserialize}
\keyword{Gene}
\keyword{Initialize}
\keyword{JSON}
\keyword{Return}
\keyword{To}
\keyword{Validate}
\keyword{a}
\keyword{all}
\keyword{an}
\keyword{any).}
\keyword{are}
\keyword{class.}
\keyword{fields}
\keyword{format)}
\keyword{if}
\keyword{in}
\keyword{input}
\keyword{instance}
\keyword{internal}
\keyword{into}
\keyword{invalid}
\keyword{list}
\keyword{new}
\keyword{of}
\keyword{respect}
\keyword{string}
\keyword{the}
\keyword{to}
\keyword{true}
\keyword{valid.}
\keyword{values}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character \link{optional}}

\item{\code{description}}{character \link{optional}}

\item{\code{id}}{integer \link{optional}}

\item{\code{previousNcbiId}}{character \link{optional}}

\item{\code{physicalLocation}}{\link{PhysicalLocation} \link{optional}}

\item{\code{officialSymbol}}{character \link{optional}}

\item{\code{officialName}}{character \link{optional}}

\item{\code{ncbiGeneId}}{integer \link{optional}}

\item{\code{ensemblId}}{character \link{optional}}

\item{\code{products}}{list(\link{GeneProduct}) \link{optional}}

\item{\code{aliases}}{list(\link{GeneAlias}) \link{optional}}

\item{\code{taxon}}{\link{Taxon} \link{optional}}

\item{\code{accessions}}{list(\link{DatabaseEntry}) \link{optional}}

\item{\code{multifunctionality}}{\link{Multifunctionality} \link{optional}}

\item{\code{phenotypeAssociations}}{list(\link{PhenotypeAssociation}) \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Gene-new}{\code{Gene$new()}}
\item \href{#method-Gene-toJSON}{\code{Gene$toJSON()}}
\item \href{#method-Gene-fromJSON}{\code{Gene$fromJSON()}}
\item \href{#method-Gene-toJSONString}{\code{Gene$toJSONString()}}
\item \href{#method-Gene-fromJSONString}{\code{Gene$fromJSONString()}}
\item \href{#method-Gene-validateJSON}{\code{Gene$validateJSON()}}
\item \href{#method-Gene-toString}{\code{Gene$toString()}}
\item \href{#method-Gene-isValid}{\code{Gene$isValid()}}
\item \href{#method-Gene-getInvalidFields}{\code{Gene$getInvalidFields()}}
\item \href{#method-Gene-print}{\code{Gene$print()}}
\item \href{#method-Gene-clone}{\code{Gene$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gene-new"></a>}}
\if{latex}{\out{\hypertarget{method-Gene-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Gene class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gene$new(
  name = NULL,
  description = NULL,
  id = NULL,
  previousNcbiId = NULL,
  physicalLocation = NULL,
  officialSymbol = NULL,
  officialName = NULL,
  ncbiGeneId = NULL,
  ensemblId = NULL,
  products = NULL,
  aliases = NULL,
  taxon = NULL,
  accessions = NULL,
  multifunctionality = NULL,
  phenotypeAssociations = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}

\item{\code{description}}{description}

\item{\code{id}}{id}

\item{\code{previousNcbiId}}{previousNcbiId}

\item{\code{physicalLocation}}{physicalLocation}

\item{\code{officialSymbol}}{officialSymbol}

\item{\code{officialName}}{officialName}

\item{\code{ncbiGeneId}}{ncbiGeneId}

\item{\code{ensemblId}}{ensemblId}

\item{\code{products}}{products}

\item{\code{aliases}}{aliases}

\item{\code{taxon}}{taxon}

\item{\code{accessions}}{accessions}

\item{\code{multifunctionality}}{multifunctionality}

\item{\code{phenotypeAssociations}}{phenotypeAssociations}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gene-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Gene-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gene$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Gene in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gene-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Gene-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Gene
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gene$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Gene
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gene-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Gene-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gene$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Gene in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gene-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Gene-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Gene
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gene$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Gene
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gene-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Gene-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Gene and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gene$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gene-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Gene-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gene$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Gene
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gene-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Gene-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gene$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gene-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Gene-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gene$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gene-print"></a>}}
\if{latex}{\out{\hypertarget{method-Gene-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gene$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gene-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Gene-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gene$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
