% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openapi_generated.R
\docType{class}
\name{GeneSetValueObject}
\alias{GeneSetValueObject}
\title{GeneSetValueObject}
\format{
An \code{R6Class} generator object
}
\description{
GeneSetValueObject Class

GeneSetValueObject Class
}
\details{
Create a new GeneSetValueObject
}
\keyword{(JSON}
\keyword{(if}
\keyword{Deserialize}
\keyword{GeneSetValueObject}
\keyword{Initialize}
\keyword{JSON}
\keyword{Return}
\keyword{To}
\keyword{Validate}
\keyword{a}
\keyword{all}
\keyword{an}
\keyword{any).}
\keyword{are}
\keyword{class.}
\keyword{fields}
\keyword{format)}
\keyword{if}
\keyword{in}
\keyword{input}
\keyword{instance}
\keyword{internal}
\keyword{into}
\keyword{invalid}
\keyword{list}
\keyword{new}
\keyword{of}
\keyword{respect}
\keyword{string}
\keyword{the}
\keyword{to}
\keyword{true}
\keyword{valid.}
\keyword{values}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer optional}

\item{\code{description}}{character optional}

\item{\code{geneIds}}{list(integer) optional}

\item{\code{isPublic}}{character optional}

\item{\code{isShared}}{character optional}

\item{\code{name}}{character optional}

\item{\code{size}}{integer optional}

\item{\code{taxonId}}{integer optional}

\item{\code{taxonName}}{character optional}

\item{\code{userOwned}}{character optional}

\item{\code{userCanWrite}}{character optional}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GeneSetValueObject-new}{\code{GeneSetValueObject$new()}}
\item \href{#method-GeneSetValueObject-toJSON}{\code{GeneSetValueObject$toJSON()}}
\item \href{#method-GeneSetValueObject-fromJSON}{\code{GeneSetValueObject$fromJSON()}}
\item \href{#method-GeneSetValueObject-toJSONString}{\code{GeneSetValueObject$toJSONString()}}
\item \href{#method-GeneSetValueObject-fromJSONString}{\code{GeneSetValueObject$fromJSONString()}}
\item \href{#method-GeneSetValueObject-validateJSON}{\code{GeneSetValueObject$validateJSON()}}
\item \href{#method-GeneSetValueObject-toString}{\code{GeneSetValueObject$toString()}}
\item \href{#method-GeneSetValueObject-isValid}{\code{GeneSetValueObject$isValid()}}
\item \href{#method-GeneSetValueObject-getInvalidFields}{\code{GeneSetValueObject$getInvalidFields()}}
\item \href{#method-GeneSetValueObject-print}{\code{GeneSetValueObject$print()}}
\item \href{#method-GeneSetValueObject-clone}{\code{GeneSetValueObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneSetValueObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-GeneSetValueObject-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new GeneSetValueObject class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneSetValueObject$new(
  id = NULL,
  description = NULL,
  geneIds = NULL,
  isPublic = NULL,
  isShared = NULL,
  name = NULL,
  size = NULL,
  taxonId = NULL,
  taxonName = NULL,
  userOwned = NULL,
  userCanWrite = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{description}}{description}

\item{\code{geneIds}}{geneIds}

\item{\code{isPublic}}{isPublic}

\item{\code{isShared}}{isShared}

\item{\code{name}}{name}

\item{\code{size}}{size}

\item{\code{taxonId}}{taxonId}

\item{\code{taxonName}}{taxonName}

\item{\code{userOwned}}{userOwned}

\item{\code{userCanWrite}}{userCanWrite}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneSetValueObject-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GeneSetValueObject-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneSetValueObject$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
GeneSetValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneSetValueObject-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GeneSetValueObject-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of GeneSetValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneSetValueObject$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of GeneSetValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneSetValueObject-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GeneSetValueObject-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneSetValueObject$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
GeneSetValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneSetValueObject-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GeneSetValueObject-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of GeneSetValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneSetValueObject$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of GeneSetValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneSetValueObject-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GeneSetValueObject-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to GeneSetValueObject and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneSetValueObject$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneSetValueObject-toString"></a>}}
\if{latex}{\out{\hypertarget{method-GeneSetValueObject-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneSetValueObject$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of GeneSetValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneSetValueObject-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-GeneSetValueObject-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneSetValueObject$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneSetValueObject-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-GeneSetValueObject-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneSetValueObject$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneSetValueObject-print"></a>}}
\if{latex}{\out{\hypertarget{method-GeneSetValueObject-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneSetValueObject$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneSetValueObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GeneSetValueObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneSetValueObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
