% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{get_dataset_object}
\alias{get_dataset_object}
\title{Compile gene expression data and metadata}
\usage{
get_dataset_object(
  datasets,
  genes = NULL,
  keepNonSpecific = FALSE,
  consolidate = NA_character_,
  filter = FALSE,
  metaType = "text",
  type = "se",
  memoised = getOption("gemma.memoised", FALSE)
)
}
\arguments{
\item{datasets}{A numerical dataset identifier or a dataset short name}

\item{genes}{An ensembl gene identifier which typically starts with ensg or an ncbi gene identifier or an official gene symbol approved by hgnc}

\item{keepNonSpecific}{logical. \code{FALSE} by default. If \code{TRUE}, results
from probesets that are not specific to the gene will also be returned.}

\item{consolidate}{An option for gene expression level consolidation. If empty,
will return every probe for the genes. "pickmax" to
pick the probe with the highest expression, "pickvar" to pick the prove with
the highest variance and "average" for returning the average expression}

\item{filter}{The filtered version corresponds to what is used in most Gemma analyses, removing some probes/elements. Unfiltered includes all elements.}

\item{type}{"se"for a SummarizedExperiment or "eset" for Expression Set. We recommend using
SummarizedExperiments which are more recent. See the Summarized experiment
\href{https://bioconductor.org/packages/release/bioc/vignettes/SummarizedExperiment/inst/doc/SummarizedExperiment.html}{vignette}
or the ExpressionSet \href{https://bioconductor.org/packages/release/bioc/vignettes/Biobase/inst/doc/ExpressionSetIntroduction.pdf}{vignette}
for more details.}

\item{memoised}{Whether or not to save to cache for future calls with the
same inputs and use the result saved in cache if a result is already saved.
Doing \code{options(gemma.memoised = TRUE)} will ensure that the cache is always
used. Use \code{\link{forget_gemma_memoised}} to clear the cache.}
}
\value{
A SummarizedExperiment, ExpressionSet or tibble containing metadata and expression data for the queried dataset.
}
\description{
Return an annotated Bioconductor-compatible
data structure or a long form tibble of the queried dataset, including
expression data and the experimental design.
}
\examples{
get_dataset_object("GSE2018")
}
\keyword{dataset}
