% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{get_platform_annotations}
\alias{get_platform_annotations}
\title{Retrieve Platform Annotations by Gemma}
\usage{
get_platform_annotations(
  platform,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)
}
\arguments{
\item{platform}{A platform numerical identifier or a platform short name}

\item{raw}{\code{TRUE} to receive results as-is from Gemma, or \code{FALSE} to enable
parsing. Raw results usually contain additional fields and flags that are
omitted in the parsed results.}

\item{memoised}{Whether or not to save to cache for future calls with the
same inputs and use the result saved in cache if a result is already saved.
Doing \code{options(gemma.memoised = TRUE)} will ensure that the cache is always
used. Use \code{\link{forget_gemma_memoised}} to clear the cache.}

\item{file}{The name of a file to save the results to, or \code{NULL} to not write
results to a file. If \code{raw == TRUE}, the output will be the raw endpoint from the
API, likely a JSON or a gzip file. Otherwise, it will be a RDS file.}

\item{overwrite}{Whether or not to overwrite if a file exists at the specified
filename.}
}
\value{
A table of annotations
\itemize{
\item \code{ElementName}: Probeset names provided by the platform.
Gene symbols for generic annotations typically used for RNA-seq experiments.
\item \code{GeneSymbols}: Genes that were found to be aligned to
the probe sequence. Note that it is possible for probes to be
non-specific. Alignment to multiple genes are indicated with gene
symbols separated by "|"s
\item \code{GeneNames}: Name of the gene
\item \code{GOTerms}: GO Terms associated with the genes.
\item \code{GemmaIDs} and \code{NCBIids}: respective IDs for the genes.
}
}
\description{
Gets Gemma's platform annotations including mappings of microarray probes to genes.
}
\examples{
head(get_platform_annotations("GPL96"))
head(get_platform_annotations("Generic_human_ncbiIds"))
}
\keyword{platform}
