% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{get_taxon_datasets}
\alias{get_taxon_datasets}
\title{Retrieve the datasets for a given taxon}
\usage{
get_taxon_datasets(
  taxon,
  offset = 0L,
  limit = 20,
  sort = "+id",
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE),
  attributes = getOption("gemma.attributes", TRUE)
)
}
\arguments{
\item{taxon}{can either be Taxon ID, Taxon NCBI ID, or one of its string identifiers: scientific name, common name.
It is recommended to use Taxon ID for efficiency.
Please note, that not all taxa have all the possible identifiers available.
Use the \code{\link{get_taxa_by_ids}} function to retrieve the necessary information. For convenience, below is a list of officially supported taxa:
\tabular{rllr}{
\strong{ID} \tab \strong{Comm.name} \tab \strong{Scient.name}    \tab \strong{NcbiID}\cr
1            \tab human               \tab Homo sapiens             \tab 9606            \cr
2            \tab mouse               \tab Mus musculus             \tab 10090           \cr
3            \tab rat                 \tab Rattus norvegicus        \tab 10116           \cr
11           \tab yeast               \tab Saccharomyces cerevisiae \tab 4932            \cr
12           \tab zebrafish           \tab Danio rerio              \tab 7955            \cr
13           \tab fly                 \tab Drosophila melanogaster  \tab 7227            \cr
14           \tab worm                \tab Caenorhabditis elegans   \tab 6239
}}

\item{offset}{The offset of the first retrieved result.}

\item{limit}{Optional, defaults to 20. Limits the result to specified amount
of objects. Has a maximum value of 100. Use together with \code{offset} and
the \code{totalElements} \link[base:attributes]{attribute} in the output to
compile all data if needed.}

\item{sort}{Order results by the given property and direction. The '+' sign indicate ascending order whereas the '-' indicate descending.}

\item{raw}{\code{TRUE} to receive results as-is from Gemma, or \code{FALSE} to enable
parsing. Raw results usually contain additional fields and flags that are
omitted in the parsed results.}

\item{memoised}{Whether or not to save to cache for future calls with the
same inputs and use the result saved in cache if a result is already saved.
Doing \code{options(gemma.memoised = TRUE)} will ensure that the cache is always
used. Use \code{\link{forget_gemma_memoised}} to clear the cache.}

\item{file}{The name of a file to save the results to, or \code{NULL} to not write
results to a file. If \code{raw == TRUE}, the output will be a JSON file. Otherwise,
it will be a RDS file.}

\item{overwrite}{Whether or not to overwrite if a file exists at the specified
filename.}

\item{attributes}{If \code{TRUE} additional information from the call will be added
into the output object's attributes such as offset and available elements.}
}
\value{
"A data table with information about the datasets associated with
the queried taxon. A \verb{404 error} if the given identifier does not map
to any object."
}
\description{
Retrieve the datasets for a given taxon
}
\examples{
get_taxon_datasets("human")
}
\keyword{taxon}
