% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{getAnnotation}
\alias{getAnnotation}
\title{Get Gemma annotations}
\usage{
getAnnotation(
  platform,
  annotType = c("bioProcess", "noParents", "allParents"),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", F)
)
}
\arguments{
\item{platform}{A platform identifier @seealso getPlatforms}

\item{annotType}{Which GO terms should the output include}

\item{file}{Where to save the annotation file to, or empty to just load into memory}

\item{overwrite}{Whether or not to overwrite an existing file}
}
\value{
A table of annotations
}
\description{
Gets Gemma's platform annotation files that can be accessed from https://gemma.msl.ubc.ca/annots/
}
\keyword{platform}
