% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{getBioconductor}
\alias{getBioconductor}
\title{Get Bioconductor data structures}
\usage{
getBioconductor(type, dataset, filter = TRUE)
}
\arguments{
\item{type}{"SummarizedExperiment" or "ExpressionSet". We recommend using
SummarizedExperiments if you are querying RNA-Seq data. See the Summarized experiment
\href{https://bioconductor.org/packages/release/bioc/vignettes/SummarizedExperiment/inst/doc/SummarizedExperiment.html}{vignette}
or the ExpressionSet \href{https://bioconductor.org/packages/release/bioc/vignettes/Biobase/inst/doc/ExpressionSetIntroduction.pdf}{vignette}
for more details.}

\item{dataset}{A dataset identifier.}

\item{filter}{The filtered vertsion corresponds to what is used in most Gemma analyses, removing some probes/elements. Unfiltered includes all elements.}
}
\value{
An annotated \code{\link[SummarizedExperiment]{SummarizedExperiment-class}} or \code{\link[Biobase]{ExpressionSet}} for the queried dataset.
}
\description{
Combines various endpoint calls to return an annotated Bioconductor-compatible
data structure of the queried dataset, including expression data,
the experimental design and experiment metadata.
}
\examples{
getBioconductor("ExpressionSet", "GSE2018", filter = TRUE)
}
\keyword{dataset}
