% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{getDatasetDEA}
\alias{getDatasetDEA}
\title{getDatasetDEA}
\usage{
getDatasetDEA(
  dataset = NA_character_,
  offset = 0L,
  limit = 20L,
  raw = FALSE,
  async = FALSE,
  memoised = FALSE
)
}
\description{
getDatasetDEA
}
