% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{getDatasetPCA}
\alias{getDatasetPCA}
\title{getDatasetPCA
Retrieves the expression levels most correlated with the given principal component}
\usage{
getDatasetPCA(
  dataset = NA_character_,
  component = 1L,
  limit = 100L,
  keepNonSpecific = FALSE,
  consolidate = NA_character_,
  raw = FALSE,
  async = FALSE,
  memoised = FALSE,
  file = NA_character_,
  overwrite = FALSE
)
}
\arguments{
\item{dataset}{<p class='description-frow'>Required, part of the URL path.</p><p>Can either be the dataset ID or its short name (e.g. <code>GSE1234</code>).</p><p>Retrieval by ID is more efficient.</p><p>Only datasets that user has access to will be available</p>}

\item{component}{<p class='description-frow'>Required, defaults to <code>empty</code>.</p><p>The pca component to limit the results to.</p>}

\item{limit}{<p class='description-frow'>Optional, defaults to <code>20</code>.</p><p>Limits the result to specified amount of objects. Use 0 for no limit.</p>}

\item{keepNonSpecific}{<p class='description-frow'>Optional, defaults to <code>false</code>.</p><p>   If set to <code>false</code>, the response will only include elements that map exclusively to each queried gene</p><p>   If set to <code>true</code>, the response will include all elements that map to each queried gene, even if they   also map to other genes.</p>}

\item{consolidate}{<p class='description-frow'>Optional, defaults to <code>empty</code>.</p><p>   What action to take when there is more than one element per gene in a dataset.</p><p>   The choices are:   <ul>       <li><code>[empty]</code> - will list all vectors separately</li>       <li><code>pickmax</code> - only return the vector that has the highest expression (mean over all its bioAssays)</li>       <li><code>pickvar</code> - only return the vector with highest variance of expression across its bioAssays</li>       <li><code>average</code> - create a new vector that will average the bioAssay values from all vectors</li>   </ul></p>}

\item{raw}{<p><code>FALSE</code> to receive results as-is from Gemma, or <code>TRUE</code> to enable parsing.</p>}

\item{async}{<p><code>TRUE</code> to run the API query on a separate worker, or <code>FALSE</code> to run synchronously. See the <code>async</code> package for details.</p>}

\item{memoised}{<p>Whether or not to cache results so future requests for the same data will be faster. Use <code>forgetGemmaMemoised</code> to clear the cache.</p>}

\item{file}{<p>The name of a file to save the results to, or <code>NULL</code> to not write results to a file. If <code>raw == TRUE</code>, the output will be a JSON file. Otherwise, it will be a RDS file.</p>}

\item{overwrite}{<p>Whether or not to overwrite if a file exists at the specified filename.</p>}
}
\value{
<p>   The expression levels for each given experiment (experiment expression levels value object) of genes that are   most correlated with the given principal component.</p><p>A <code>404 error</code> if the given identifier does not map to any object.</p></p>
}
\description{
getDatasetPCA
Retrieves the expression levels most correlated with the given principal component
}
