% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{getDatasetResultSets}
\alias{getDatasetResultSets}
\title{getDatasetResultSets}
\usage{
getDatasetResultSets(
  dataset = NA_character_,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)
}
\arguments{
\item{dataset}{Required, part of the URL path.
Can either be the dataset ID or its short name (e.g. \code{GSE1234}).
Retrieval by ID is more efficient.
Only datasets that user has access to will be available}

\item{raw}{\code{TRUE} to receive results as-is from Gemma, or \code{FALSE} to enable
parsing.}

\item{memoised}{Whether or not to cache results so future requests for the same data
will be faster. Use \code{forgetGemmaMemoised} to clear the cache.}

\item{file}{The name of a file to save the results to, or \code{NULL} to not write
results to a file. If \code{raw == TRUE}, the output will be a JSON file.
Otherwise, it will be a RDS file.}

\item{overwrite}{Whether or not to overwrite if a file exists at the specified filename.}
}
\value{
Varies
}
\description{
foobar
}
\examples{
getDatasetResultSets("GSE2018")
}
\keyword{dataset}
