% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{getDatasetsInfo}
\alias{getDatasetsInfo}
\title{Datasets}
\usage{
getDatasetsInfo(
  datasets = NA_character_,
  offset = 0L,
  limit = 20L,
  sort = "+id",
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)
}
\arguments{
\item{datasets}{Optional, defaults to \code{empty}.
Limits the result to entities with given identifiers.
A list of identifiers, separated by commas (e.g:
\verb{GSE2871,GSE2869,GSE2868}). Identifiers can either be the Dataset ID or
its short name. Retrieval by ID is more efficient.
Only datasets that user has access to will be available.
Do not combine different identifiers in one query.}

\item{offset}{Optional, defaults to \code{0}.
Skips the specified amount of objects when retrieving them from the
database.}

\item{limit}{Optional, defaults to 20. Limits the result to specified amount of
objects.}

\item{sort}{Optional, defaults to \code{+id}.
Sets the ordering property and direction.
Format is \verb{[+,-][property name]}. E.g. \code{-accession} will translate to
descending ordering by the 'Accession' property.
Note that this does [not guarantee the order of the returned
entities!]{.description-imp} This is merely a signal to how the data
should be pre-sorted before the limit and offset are applied.
Nested properties are also supported (recursively).
E.g: \code{+curationDetails.lastTroubledEvent.date}
When
using in scripts, remember to URL-encode the '+' plus character (see
the compiled URL below).}

\item{raw}{\code{TRUE} to receive results as-is from Gemma, or \code{FALSE} to enable
parsing. Raw results usually contain additional fields and flags that
are omitted in the parsed results.}

\item{memoised}{Whether or not to cache results so future requests for the same data
will be faster. Use \code{forgetGemmaMemoised} to clear the cache.}

\item{file}{The name of a file to save the results to, or \code{NULL} to not write
results to a file. If \code{raw == TRUE}, the output will be a JSON file.
Otherwise, it will be a RDS file.}

\item{overwrite}{Whether or not to overwrite if a file exists at the specified filename.}
}
\value{
An array of value objects representing the objects that matched the
query.
Empty array if no objects matched.
A successful response may contain a sub-element with 'Geeq'
information, which aims to provide a unified metric to measure
experiments by the quality of their data, and their suitability for use
in Gemma. You can \href{https://pavlidislab.github.io/Gemma/geeq.html}{read more about the geeq properties here}.
}
\description{
Lists datasets filtered and organized by given parameters
}
\examples{
getDatasetsInfo("GSE2018")
getDatasetsInfo(c("GSE2018", "GSE2872"))
}
\keyword{dataset}
