% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{getDiffExData}
\alias{getDiffExData}
\title{Get differential expression data}
\usage{
getDiffExData(dataset, diffExSet)
}
\arguments{
\item{dataset}{The experiment ID (integer)}

\item{diffExSet}{The result set ID for the differential expression analysis we want (integer) @seealso getDatasetDEA}
}
\value{
A list of \code{data.table}s that were read in, typically the first one is analysis results (elements, p-values) and the second is full data (elements, FC, t- and p-values)
}
\description{
This occurs asynchronously as data will not always be available upon request. If the data is not available,
the function will continue to run and poll for updates every 3 seconds until it becomes available.
}
