% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{getDiffExpr}
\alias{getDiffExpr}
\title{Get Differential Expression ResultSet}
\usage{
getDiffExpr(dataset)
}
\arguments{
\item{dataset}{A dataset identifier.}
}
\value{
A \code{\link[SummarizedExperiment]{SummarizedExperiment}} or \code{\link[Biobase]{ExpressionSet}} of the queried resultSet.
}
\description{
Combines the expression and design matrix of the queried dataset into one
tibble for easy visualization and exploration with \code{\link[ggplot2]{ggplot}} and the rest of the tidyverse.
}
\examples{
getDiffExpr("GSE2018")
}
\keyword{dataset}
