% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{getExpressionSet}
\alias{getExpressionSet}
\title{Get ExpressionSet}
\usage{
getExpressionSet(dataset, filter)
}
\arguments{
\item{dataset}{A dataset identifier}

\item{filter}{If true, call returns filtered expression data.}
}
\value{
An ExpressionSet instance with an expression matrix annotated with metadata
}
\description{
Combines various endpoint calls to return a Bioconductor ExpressionSet
}
