% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{getGeneEvidence}
\alias{getGeneEvidence}
\title{getGeneEvidence}
\usage{
getGeneEvidence(
  gene = NA_character_,
  raw = FALSE,
  async = FALSE,
  memoised = FALSE,
  file = NA_character_,
  overwrite = FALSE
)
}
\description{
getGeneEvidence
}
