% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{getGeneLocation}
\alias{getGeneLocation}
\title{Gene locations}
\usage{
getGeneLocation(
  gene,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)
}
\arguments{
\item{gene}{Required, part of the URL path.
Can either be the NCBI ID (\code{1859}), Ensembl ID (\code{ENSG00000157540}) or
official symbol (\code{DYRK1A}) of the gene.
NCBI ID is the most efficient (and guaranteed to be unique) identifier.
Official
symbol represents a gene homologue for a random taxon, unless used in a
specific taxon (see Taxon Endpoints).}

\item{raw}{\code{TRUE} to receive results as-is from Gemma, or \code{FALSE} to enable
parsing. Raw results usually contain additional fields and flags that
are omitted in the parsed results.}

\item{memoised}{Whether or not to save to cache for future calls with the same inputs
and use the result saved in cache if a result is already saved. Doing
\code{options(gemma.memoised = TRUE)} will ensure you are always using the
cache. Use \code{forgetGemmaMemoised} to clear the cache.}

\item{file}{The name of a file to save the results to, or \code{NULL} to not write
results to a file. If \code{raw == TRUE}, the output will be a JSON file.
Otherwise, it will be a RDS file.}

\item{overwrite}{Whether or not to overwrite if a file exists at the specified filename.}
}
\value{
An array of locations (physical location value objects) of the given
gene or gene homologues.
A \verb{404 error} if the given identifier does not map to any object.
}
\description{
Retrieves the physical location of the given gene
}
\examples{
getGeneLocation("DYRK1A")
}
\keyword{gene}
