% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{getGeneLocationOnTaxon}
\alias{getGeneLocationOnTaxon}
\title{getGeneLocationOnTaxon}
\usage{
getGeneLocationOnTaxon(
  taxon = NA_character_,
  gene = NA_character_,
  raw = FALSE,
  async = FALSE,
  memoised = FALSE,
  file = NA_character_,
  overwrite = FALSE
)
}
\description{
getGeneLocationOnTaxon
}
