% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{getGeneOnTaxon}
\alias{getGeneOnTaxon}
\title{getGeneOnTaxon}
\usage{
getGeneOnTaxon(
  taxon = NA_character_,
  gene = NA_character_,
  raw = FALSE,
  async = FALSE,
  memoised = FALSE
)
}
\description{
getGeneOnTaxon
}
