% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{getGenesAtLocation}
\alias{getGenesAtLocation}
\title{getGenesAtLocation
Finds genes overlapping a given region.}
\usage{
getGenesAtLocation(
  taxon = NA_character_,
  chromosome = NA_character_,
  strand = "+",
  start = NA_integer_,
  size = NA_integer_,
  raw = FALSE,
  async = FALSE,
  memoised = FALSE,
  file = NA_character_,
  overwrite = FALSE
)
}
\arguments{
\item{taxon}{<p class='description-frow'>Not required, part of the URL path.</p><p>   can either be Taxon ID, Taxon NCBI ID, or one of its string identifiers:   scientific name, common name</p><p>It is recommended to use Taxon ID for efficiency.</p><p>Please note, that not all taxa have all the possible identifiers available.</p><p>Use the 'All Taxa' endpoint to retrieve the necessary information. For convenience, below is a list of officially supported taxa: </p><table><tr><th> ID </th><th>Comm.name</th><th>Scient.name</th><th>NcbiID</th></tr><tr><td>  1 </td><td> human           </td><td> Homo sapiens              </td><td>    9606 </td></tr><tr><td>  2 </td><td> mouse           </td><td> Mus musculus              </td><td>   10090 </td></tr><tr><td>  3 </td><td> rat             </td><td> Rattus norvegicus         </td><td>   10116 </td></tr><tr><td> 11 </td><td> yeast           </td><td> Saccharomyces cerevisiae  </td><td>    4932 </td></tr><tr><td> 12 </td><td> zebrafish       </td><td> Danio rerio               </td><td>    7955 </td></tr><tr><td> 13 </td><td> fly             </td><td> Drosophila melanogaster   </td><td>    7227 </td></tr><tr><td> 14 </td><td> worm            </td><td> Caenorhabditis elegans    </td><td>    6239 </td></tr></table>}

\item{chromosome}{<p class='description-frow'>Required, defaults to <code>empty</code>.</p><p>The chromosome of the query location. Eg: <code>3</code>, <code>21</code>, <code>X</code></p>}

\item{strand}{<p class='description-frow'>Optional, defaults to <code>+</code>.</p><p>Can either be <code>+</code> or <code>-</code>.</p><p class='description-imp'>   <span class='glyphicon glyphicon-th-large glyphicon-exclamation-sign'></span>   This is a WIP parameter and does currently not do anything</p><p class='description-imp'>   <span class='glyphicon glyphicon-th-large glyphicon-exclamation-sign'></span>   When using in scripts, remember to URL-encode the '+' plus character (see the compiled URL below).</p>}

\item{start}{<p class='description-frow'>Required, defaults to <code>empty</code>.</p><p>Number of the start nucleotide of the desired region.</p>}

\item{size}{<p class='description-frow'>Required, defaults to <code>empty</code>.</p><p>Amount of nucleotides in the desired region (i.e. the length of the region).</p>}

\item{raw}{<p><code>FALSE</code> to receive results as-is from Gemma, or <code>TRUE</code> to enable parsing.</p>}

\item{async}{<p><code>TRUE</code> to run the API query on a separate worker, or <code>FALSE</code> to run synchronously. See the <code>async</code> package for details.</p>}

\item{memoised}{<p>Whether or not to cache results so future requests for the same data will be faster. Use <code>forgetGemmaMemoised</code> to clear the cache.</p>}

\item{file}{<p>The name of a file to save the results to, or <code>NULL</code> to not write results to a file. If <code>raw == TRUE</code>, the output will be a JSON file. Otherwise, it will be a RDS file.</p>}

\item{overwrite}{<p>Whether or not to overwrite if a file exists at the specified filename.</p>}
}
\value{
<p> An array of genes (gene value objects) that overlap the given region.</p><p>A <code>404 error</code> if the given identifier does not map to any object.</p><p>A <code>400 error</code> if required parameters are missing.</p>
}
\description{
getGenesAtLocation
Finds genes overlapping a given region.
}
