% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{getGenomeVersions}
\alias{getGenomeVersions}
\title{Get genome versions}
\usage{
getGenomeVersions(memoised = getOption("gemma.memoised", FALSE))
}
\arguments{
\item{memoised}{Whether or not to save to cache for future calls with the same inputs
and use the result saved in cache if a result is already saved. Doing
\code{options(gemma.memoised = TRUE)} will ensure that the catche is always used.
Use \code{forgetGemmaMemoised} to clear the cache.}
}
\value{
A data frame
}
\description{
Returns the genome version used within Gemma
}
\examples{
getGenomeVersions()
}
