% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{getPlatformAnnotation}
\alias{getPlatformAnnotation}
\title{Gemma platform annotations}
\usage{
getPlatformAnnotation(
  platform,
  annotType = c("bioProcess", "noParents", "allParents"),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  unzip = FALSE
)
}
\arguments{
\item{platform}{A platform identifier @seealso getPlatforms}

\item{annotType}{Which GO terms should the output include}

\item{file}{Where to save the annotation file to, or empty to just load into memory}

\item{overwrite}{Whether or not to overwrite an existing file}

\item{memoised}{Whether or not to save to cache for future calls with the same inputs
and use the result saved in cache if a result is already saved. Doing
\code{options(gemma.memoised = TRUE)} will ensure that the cache is always used.
Use \code{forgetGemmaMemoised} to clear the cache.}

\item{unzip}{Whether or not to unzip the file (if @param file is not empty)}
}
\value{
A table of annotations
}
\description{
Gets Gemma's platform annotation files that can be accessed from https://gemma.msl.ubc.ca/annots/
}
\examples{
dat <- getPlatformAnnotation("GPL96")
str(dat)
}
\keyword{platform}
