% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{getPlatformElementGenes}
\alias{getPlatformElementGenes}
\title{getPlatformElementGenes
Retrieves the genes on the given platform element}
\usage{
getPlatformElementGenes(
  platform = NA_character_,
  element = NA_character_,
  offset = 0L,
  limit = 20L,
  raw = FALSE,
  async = FALSE,
  memoised = FALSE,
  file = NA_character_,
  overwrite = FALSE
)
}
\arguments{
\item{platform}{<p class='description-frow'>Required, part of the URL path.</p><p>Can either be the platform ID or its short name (e.g: <code>GPL1355</code>)</p><p>Retrieval by ID is more efficient. </p><p>Only platforms that user has access to will be available.</p>}

\item{element}{<p class='description-frow'>Optional, defaults to <code>empty</code>.</p><p>Limits the result to entities with given identifiers.</p><p>A list of identifiers, separated by commas (e.g: <code>AFFX_Rat_beta-actin_M_at, AFFX_Rat_Hexokinase_M_at</code>).</p><p>Can either be probes name or IDs. <br/><p>Do not combine different identifiers in one query.</p><p class='description-imp'>   <span class='glyphicon glyphicon-th-large glyphicon-exclamation-sign'></span>   When using in scripts, remember to URL-encode any forward slashes in the probe name (see the compiled URL below).</p>}

\item{offset}{<p class='description-frow'>Optional, defaults to <code>0</code>.</p><p>Skips the specified amount of objects when retrieving them from the database.</p>}

\item{limit}{<p class='description-frow'>Optional, defaults to <code>20</code>.</p><p>Limits the result to specified amount of objects. Use 0 for no limit.</p>}

\item{raw}{<p><code>FALSE</code> to receive results as-is from Gemma, or <code>TRUE</code> to enable parsing.</p>}

\item{async}{<p><code>TRUE</code> to run the API query on a separate worker, or <code>FALSE</code> to run synchronously. See the <code>async</code> package for details.</p>}

\item{memoised}{<p>Whether or not to cache results so future requests for the same data will be faster. Use <code>forgetGemmaMemoised</code> to clear the cache.</p>}

\item{file}{<p>The name of a file to save the results to, or <code>NULL</code> to not write results to a file. If <code>raw == TRUE</code>, the output will be a JSON file. Otherwise, it will be a RDS file.</p>}

\item{overwrite}{<p>Whether or not to overwrite if a file exists at the specified filename.</p>}
}
\value{
<p> An array of genes (gene value objects) aligned with the the given platform element.</p><p> All identifiers must be valid. </p><p>A <code>404 error</code> if the given identifier does not map to any object.</p></p>
}
\description{
getPlatformElementGenes
Retrieves the genes on the given platform element
}
