% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{getPlatforms}
\alias{getPlatforms}
\title{getPlatforms}
\usage{
getPlatforms(
  platforms = NA_character_,
  filter = NA_character_,
  offset = 0L,
  limit = 20L,
  sort = "+id",
  raw = TRUE,
  async = FALSE,
  memoised = FALSE
)
}
\description{
getPlatforms
}
