% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_allEndpoints.R
\name{getPlatformsInfo}
\alias{getPlatformsInfo}
\title{Platforms}
\usage{
getPlatformsInfo(
  platforms = NA_character_,
  offset = 0L,
  limit = 20L,
  sort = "+id",
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE),
  attributes = getOption("gemma.attributes", TRUE)
)
}
\arguments{
\item{platforms}{Optional, defaults to \code{empty}.
Limits the result to entities with given identifiers.
A list of identifiers, separated by commas (e.g:
\verb{GPL96,GPL1355,GPL1261}). Identifiers can either be the Platform ID or
its short name. Retrieval by ID is more efficient.
Only platforms that user has access to will be available.
Do not combine different identifiers in one query.}

\item{offset}{Optional, defaults to \code{0}.
Skips the specified amount of objects when retrieving them from the
database.}

\item{limit}{Optional, defaults to 20. Limits the result to specified amount of objects.}

\item{sort}{Optional, defaults to \code{+id}.
Sets the ordering property and direction.
Format is \verb{[+,-][property name]}. E.g. \code{-accession} will translate to
descending ordering by the 'Accession' property.
Note that this does [not guarantee the order of the returned
entities!]{.description-imp} This is merely a signal to how the data
should be pre-sorted before the limit and offset are applied.
Nested properties are also supported (recursively).
E.g: \code{+curationDetails.lastTroubledEvent.date}
When
using in scripts, remember to URL-encode the '+' plus character (see
the compiled URL below).}

\item{raw}{\code{TRUE} to receive results as-is from Gemma, or \code{FALSE} to enable
parsing. Raw results usually contain additional fields and flags that are
omitted in the parsed results.}

\item{memoised}{Whether or not to save to cache for future calls with the
same inputs and use the result saved in cache if a result is already saved.
Doing \code{options(gemma.memoised = TRUE)} will ensure that the cache is always
used. Use \\code{\\link{forgetGemmaMemoised}} to clear the cache.}

\item{file}{The name of a file to save the results to, or \code{NULL} to not write
results to a file. If \code{raw == TRUE}, the output will be a JSON file. Otherwise,
it will be a RDS file.}

\item{overwrite}{Whether or not to overwrite if a file exists at the specified
filename.}

\item{attributes}{If \code{TRUE} additional information from the call will be added
into the output object's attributes such as offset and available elements.}
}
\value{
A data table with information about the queried platform(s).
}
\description{
List platforms filtered and organized by given parameters
}
\examples{
getPlatformsInfo("GPL1355")
getPlatformsInfo(c("GPL1355", "GPL96"))
}
\keyword{platform}
