% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_result_sets_database_entries_parameter_inner.R
\docType{class}
\name{GetResultSetsDatabaseEntriesParameterInner}
\alias{GetResultSetsDatabaseEntriesParameterInner}
\title{GetResultSetsDatabaseEntriesParameterInner}
\format{
An \code{R6Class} generator object
}
\description{
GetResultSetsDatabaseEntriesParameterInner Class
}
\keyword{Deserialize}
\keyword{GetResultSetsDatabaseEntriesParameterInner}
\keyword{GetResultSetsDatabaseEntriesParameterInner.}
\keyword{JSON}
\keyword{JSON.}
\keyword{Returns}
\keyword{Serialize}
\keyword{Validate}
\keyword{an}
\keyword{input}
\keyword{instance}
\keyword{instance.}
\keyword{internal}
\keyword{into}
\keyword{of}
\keyword{representation}
\keyword{respect}
\keyword{string}
\keyword{string.}
\keyword{the}
\keyword{to}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new GetResultSetsDatabaseEntriesParameterInner.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetResultSetsDatabaseEntriesParameterInner-new}{\code{GetResultSetsDatabaseEntriesParameterInner$new()}}
\item \href{#method-GetResultSetsDatabaseEntriesParameterInner-fromJSONString}{\code{GetResultSetsDatabaseEntriesParameterInner$fromJSONString()}}
\item \href{#method-GetResultSetsDatabaseEntriesParameterInner-fromJSON}{\code{GetResultSetsDatabaseEntriesParameterInner$fromJSON()}}
\item \href{#method-GetResultSetsDatabaseEntriesParameterInner-toJSONString}{\code{GetResultSetsDatabaseEntriesParameterInner$toJSONString()}}
\item \href{#method-GetResultSetsDatabaseEntriesParameterInner-toJSON}{\code{GetResultSetsDatabaseEntriesParameterInner$toJSON()}}
\item \href{#method-GetResultSetsDatabaseEntriesParameterInner-validateJSON}{\code{GetResultSetsDatabaseEntriesParameterInner$validateJSON()}}
\item \href{#method-GetResultSetsDatabaseEntriesParameterInner-toString}{\code{GetResultSetsDatabaseEntriesParameterInner$toString()}}
\item \href{#method-GetResultSetsDatabaseEntriesParameterInner-print}{\code{GetResultSetsDatabaseEntriesParameterInner$print()}}
\item \href{#method-GetResultSetsDatabaseEntriesParameterInner-clone}{\code{GetResultSetsDatabaseEntriesParameterInner$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetResultSetsDatabaseEntriesParameterInner-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetResultSetsDatabaseEntriesParameterInner-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new GetResultSetsDatabaseEntriesParameterInner.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetResultSetsDatabaseEntriesParameterInner$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "character", "integer"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetResultSetsDatabaseEntriesParameterInner-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GetResultSetsDatabaseEntriesParameterInner-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of GetResultSetsDatabaseEntriesParameterInner.
An alias to the method \code{fromJSON} .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetResultSetsDatabaseEntriesParameterInner$fromJSONString(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of GetResultSetsDatabaseEntriesParameterInner.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetResultSetsDatabaseEntriesParameterInner-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GetResultSetsDatabaseEntriesParameterInner-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of GetResultSetsDatabaseEntriesParameterInner.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetResultSetsDatabaseEntriesParameterInner$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of GetResultSetsDatabaseEntriesParameterInner.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetResultSetsDatabaseEntriesParameterInner-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GetResultSetsDatabaseEntriesParameterInner-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize GetResultSetsDatabaseEntriesParameterInner to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetResultSetsDatabaseEntriesParameterInner$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the GetResultSetsDatabaseEntriesParameterInner.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetResultSetsDatabaseEntriesParameterInner-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GetResultSetsDatabaseEntriesParameterInner-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize GetResultSetsDatabaseEntriesParameterInner to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetResultSetsDatabaseEntriesParameterInner$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the GetResultSetsDatabaseEntriesParameterInner.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetResultSetsDatabaseEntriesParameterInner-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GetResultSetsDatabaseEntriesParameterInner-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to GetResultSetsDatabaseEntriesParameterInner and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetResultSetsDatabaseEntriesParameterInner$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetResultSetsDatabaseEntriesParameterInner-toString"></a>}}
\if{latex}{\out{\hypertarget{method-GetResultSetsDatabaseEntriesParameterInner-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetResultSetsDatabaseEntriesParameterInner$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetResultSetsDatabaseEntriesParameterInner-print"></a>}}
\if{latex}{\out{\hypertarget{method-GetResultSetsDatabaseEntriesParameterInner-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetResultSetsDatabaseEntriesParameterInner$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetResultSetsDatabaseEntriesParameterInner-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetResultSetsDatabaseEntriesParameterInner-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetResultSetsDatabaseEntriesParameterInner$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
