% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openapi_generated.R
\docType{class}
\name{GetResultSetsDatasetsParameterInner}
\alias{GetResultSetsDatasetsParameterInner}
\title{GetResultSetsDatasetsParameterInner}
\format{
An \code{R6Class} generator object
}
\description{
GetResultSetsDatasetsParameterInner Class
}
\keyword{Deserialize}
\keyword{GetResultSetsDatasetsParameterInner}
\keyword{GetResultSetsDatasetsParameterInner.}
\keyword{JSON}
\keyword{JSON.}
\keyword{Returns}
\keyword{Serialize}
\keyword{Validate}
\keyword{an}
\keyword{input}
\keyword{instance}
\keyword{instance.}
\keyword{internal}
\keyword{into}
\keyword{of}
\keyword{representation}
\keyword{respect}
\keyword{string}
\keyword{string.}
\keyword{the}
\keyword{to}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new GetResultSetsDatasetsParameterInner.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GetResultSetsDatasetsParameterInner-new}{\code{GetResultSetsDatasetsParameterInner$new()}}
\item \href{#method-GetResultSetsDatasetsParameterInner-fromJSONString}{\code{GetResultSetsDatasetsParameterInner$fromJSONString()}}
\item \href{#method-GetResultSetsDatasetsParameterInner-fromJSON}{\code{GetResultSetsDatasetsParameterInner$fromJSON()}}
\item \href{#method-GetResultSetsDatasetsParameterInner-toJSONString}{\code{GetResultSetsDatasetsParameterInner$toJSONString()}}
\item \href{#method-GetResultSetsDatasetsParameterInner-toJSON}{\code{GetResultSetsDatasetsParameterInner$toJSON()}}
\item \href{#method-GetResultSetsDatasetsParameterInner-validateJSON}{\code{GetResultSetsDatasetsParameterInner$validateJSON()}}
\item \href{#method-GetResultSetsDatasetsParameterInner-toString}{\code{GetResultSetsDatasetsParameterInner$toString()}}
\item \href{#method-GetResultSetsDatasetsParameterInner-print}{\code{GetResultSetsDatasetsParameterInner$print()}}
\item \href{#method-GetResultSetsDatasetsParameterInner-clone}{\code{GetResultSetsDatasetsParameterInner$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetResultSetsDatasetsParameterInner-new"></a>}}
\if{latex}{\out{\hypertarget{method-GetResultSetsDatasetsParameterInner-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new GetResultSetsDatasetsParameterInner.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetResultSetsDatasetsParameterInner$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "character", "integer"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetResultSetsDatasetsParameterInner-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GetResultSetsDatasetsParameterInner-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of GetResultSetsDatasetsParameterInner.
An alias to the method \code{fromJSON} .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetResultSetsDatasetsParameterInner$fromJSONString(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of GetResultSetsDatasetsParameterInner.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetResultSetsDatasetsParameterInner-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GetResultSetsDatasetsParameterInner-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of GetResultSetsDatasetsParameterInner.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetResultSetsDatasetsParameterInner$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of GetResultSetsDatasetsParameterInner.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetResultSetsDatasetsParameterInner-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GetResultSetsDatasetsParameterInner-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize GetResultSetsDatasetsParameterInner to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetResultSetsDatasetsParameterInner$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the GetResultSetsDatasetsParameterInner.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetResultSetsDatasetsParameterInner-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GetResultSetsDatasetsParameterInner-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize GetResultSetsDatasetsParameterInner to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetResultSetsDatasetsParameterInner$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the GetResultSetsDatasetsParameterInner.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetResultSetsDatasetsParameterInner-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GetResultSetsDatasetsParameterInner-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to GetResultSetsDatasetsParameterInner and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetResultSetsDatasetsParameterInner$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetResultSetsDatasetsParameterInner-toString"></a>}}
\if{latex}{\out{\hypertarget{method-GetResultSetsDatasetsParameterInner-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetResultSetsDatasetsParameterInner$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetResultSetsDatasetsParameterInner-print"></a>}}
\if{latex}{\out{\hypertarget{method-GetResultSetsDatasetsParameterInner-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetResultSetsDatasetsParameterInner$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetResultSetsDatasetsParameterInner-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GetResultSetsDatasetsParameterInner-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GetResultSetsDatasetsParameterInner$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
