% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{getTidyDataset}
\alias{getTidyDataset}
\title{Tidy expression and design}
\usage{
getTidyDataset(dataset, filter = FALSE)
}
\arguments{
\item{dataset}{A dataset identifier.}

\item{filter}{The filtered version corresponds to what is used in most Gemma analyses, removing some probes/elements. Unfiltered includes all elements.}
}
\value{
A tibble that combines the expression and design matrices.
}
\description{
Combines the expression and design matrices of the queried dataset into a
tibble for easy visualization and exploration with ggplot and the rest of the tidyverse.
}
\examples{
getTidyDataset("GSE2018")
}
\keyword{dataset}
