% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiable_value_object.R
\docType{class}
\name{IdentifiableValueObject}
\alias{IdentifiableValueObject}
\title{IdentifiableValueObject}
\format{
An \code{R6Class} generator object
}
\description{
IdentifiableValueObject Class

IdentifiableValueObject Class
}
\details{
Create a new IdentifiableValueObject
}
\keyword{(JSON}
\keyword{(if}
\keyword{Deserialize}
\keyword{IdentifiableValueObject}
\keyword{Initialize}
\keyword{JSON}
\keyword{Return}
\keyword{To}
\keyword{Validate}
\keyword{a}
\keyword{all}
\keyword{an}
\keyword{any).}
\keyword{are}
\keyword{class.}
\keyword{fields}
\keyword{format)}
\keyword{if}
\keyword{in}
\keyword{input}
\keyword{instance}
\keyword{internal}
\keyword{into}
\keyword{invalid}
\keyword{list}
\keyword{new}
\keyword{of}
\keyword{respect}
\keyword{string}
\keyword{the}
\keyword{to}
\keyword{true}
\keyword{valid.}
\keyword{values}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IdentifiableValueObject-new}{\code{IdentifiableValueObject$new()}}
\item \href{#method-IdentifiableValueObject-toJSON}{\code{IdentifiableValueObject$toJSON()}}
\item \href{#method-IdentifiableValueObject-fromJSON}{\code{IdentifiableValueObject$fromJSON()}}
\item \href{#method-IdentifiableValueObject-toJSONString}{\code{IdentifiableValueObject$toJSONString()}}
\item \href{#method-IdentifiableValueObject-fromJSONString}{\code{IdentifiableValueObject$fromJSONString()}}
\item \href{#method-IdentifiableValueObject-validateJSON}{\code{IdentifiableValueObject$validateJSON()}}
\item \href{#method-IdentifiableValueObject-toString}{\code{IdentifiableValueObject$toString()}}
\item \href{#method-IdentifiableValueObject-isValid}{\code{IdentifiableValueObject$isValid()}}
\item \href{#method-IdentifiableValueObject-getInvalidFields}{\code{IdentifiableValueObject$getInvalidFields()}}
\item \href{#method-IdentifiableValueObject-print}{\code{IdentifiableValueObject$print()}}
\item \href{#method-IdentifiableValueObject-clone}{\code{IdentifiableValueObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentifiableValueObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-IdentifiableValueObject-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new IdentifiableValueObject class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdentifiableValueObject$new(id = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentifiableValueObject-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-IdentifiableValueObject-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdentifiableValueObject$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
IdentifiableValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentifiableValueObject-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-IdentifiableValueObject-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of IdentifiableValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdentifiableValueObject$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of IdentifiableValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentifiableValueObject-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-IdentifiableValueObject-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdentifiableValueObject$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
IdentifiableValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentifiableValueObject-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-IdentifiableValueObject-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of IdentifiableValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdentifiableValueObject$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of IdentifiableValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentifiableValueObject-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-IdentifiableValueObject-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to IdentifiableValueObject and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdentifiableValueObject$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentifiableValueObject-toString"></a>}}
\if{latex}{\out{\hypertarget{method-IdentifiableValueObject-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdentifiableValueObject$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of IdentifiableValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentifiableValueObject-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-IdentifiableValueObject-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdentifiableValueObject$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentifiableValueObject-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-IdentifiableValueObject-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdentifiableValueObject$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentifiableValueObject-print"></a>}}
\if{latex}{\out{\hypertarget{method-IdentifiableValueObject-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdentifiableValueObject$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentifiableValueObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IdentifiableValueObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdentifiableValueObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
