% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subProcessors.R
\name{isEmpty}
\alias{isEmpty}
\title{Check for empty arguments}
\usage{
isEmpty(x)
}
\arguments{
\item{x}{A parameter that can be NA or NULL when empty}
}
\description{
Gemma functions accept typed NAs instead of simple NULLS. I believe this was
done as a non-standard-in-R way of specifying data types for the users. Usually
this is fine but it makes checking for emptyness a bit annoying since a user
can still provide NULLs to make them empty
}
\keyword{internal}
