% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openapi_generated.R
\docType{class}
\name{Keyword}
\alias{Keyword}
\title{Keyword}
\format{
An \code{R6Class} generator object
}
\description{
Keyword Class

Keyword Class
}
\details{
Create a new Keyword
}
\keyword{(JSON}
\keyword{(if}
\keyword{Deserialize}
\keyword{Initialize}
\keyword{JSON}
\keyword{Keyword}
\keyword{Return}
\keyword{To}
\keyword{Validate}
\keyword{a}
\keyword{all}
\keyword{an}
\keyword{any).}
\keyword{are}
\keyword{class.}
\keyword{fields}
\keyword{format)}
\keyword{if}
\keyword{in}
\keyword{input}
\keyword{instance}
\keyword{internal}
\keyword{into}
\keyword{invalid}
\keyword{list}
\keyword{new}
\keyword{of}
\keyword{respect}
\keyword{string}
\keyword{the}
\keyword{to}
\keyword{true}
\keyword{valid.}
\keyword{values}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer optional}

\item{\code{isMajorTopic}}{character optional}

\item{\code{term}}{character optional}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Keyword-new}{\code{Keyword$new()}}
\item \href{#method-Keyword-toJSON}{\code{Keyword$toJSON()}}
\item \href{#method-Keyword-fromJSON}{\code{Keyword$fromJSON()}}
\item \href{#method-Keyword-toJSONString}{\code{Keyword$toJSONString()}}
\item \href{#method-Keyword-fromJSONString}{\code{Keyword$fromJSONString()}}
\item \href{#method-Keyword-validateJSON}{\code{Keyword$validateJSON()}}
\item \href{#method-Keyword-toString}{\code{Keyword$toString()}}
\item \href{#method-Keyword-isValid}{\code{Keyword$isValid()}}
\item \href{#method-Keyword-getInvalidFields}{\code{Keyword$getInvalidFields()}}
\item \href{#method-Keyword-print}{\code{Keyword$print()}}
\item \href{#method-Keyword-clone}{\code{Keyword$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Keyword-new"></a>}}
\if{latex}{\out{\hypertarget{method-Keyword-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Keyword class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Keyword$new(id = NULL, isMajorTopic = NULL, term = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{isMajorTopic}}{isMajorTopic}

\item{\code{term}}{term}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Keyword-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Keyword-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Keyword$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Keyword in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Keyword-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Keyword-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Keyword
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Keyword$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Keyword
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Keyword-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Keyword-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Keyword$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Keyword in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Keyword-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Keyword-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Keyword
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Keyword$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Keyword
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Keyword-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Keyword-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Keyword and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Keyword$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Keyword-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Keyword-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Keyword$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Keyword
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Keyword-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Keyword-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Keyword$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Keyword-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Keyword-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Keyword$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Keyword-print"></a>}}
\if{latex}{\out{\hypertarget{method-Keyword-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Keyword$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Keyword-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Keyword-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Keyword$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
