% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{make_design}
\alias{make_design}
\title{Make simplified design frames}
\usage{
make_design(samples, metaType = "text")
}
\arguments{
\item{samples}{An output from get_dataset_samples. The output should not be raw}

\item{metaType}{Type of metadata to include in the output. "text", "uri" or "both"}
}
\value{
A data.frame including the design table for the dataset
}
\description{
Using on the output of \code{\link{get_dataset_samples}}, this function creates
a simplified design table, granting one column to each experimental variable
}
\examples{
get_dataset_samples('GSE46416') \%>\% make_design('both') \%>\% class

}
\keyword{misc}
