% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_value_object.R
\docType{class}
\name{MeasurementValueObject}
\alias{MeasurementValueObject}
\title{MeasurementValueObject}
\format{
An \code{R6Class} generator object
}
\description{
MeasurementValueObject Class

MeasurementValueObject Class
}
\details{
Create a new MeasurementValueObject
}
\keyword{(JSON}
\keyword{(if}
\keyword{Deserialize}
\keyword{Initialize}
\keyword{JSON}
\keyword{MeasurementValueObject}
\keyword{Return}
\keyword{To}
\keyword{Validate}
\keyword{a}
\keyword{all}
\keyword{an}
\keyword{any).}
\keyword{are}
\keyword{class.}
\keyword{fields}
\keyword{format)}
\keyword{if}
\keyword{in}
\keyword{input}
\keyword{instance}
\keyword{internal}
\keyword{into}
\keyword{invalid}
\keyword{list}
\keyword{new}
\keyword{of}
\keyword{respect}
\keyword{string}
\keyword{the}
\keyword{to}
\keyword{true}
\keyword{valid.}
\keyword{values}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer optional}

\item{\code{value}}{character optional}

\item{\code{unit}}{character optional}

\item{\code{unitId}}{integer optional}

\item{\code{type}}{character optional}

\item{\code{representation}}{character optional}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MeasurementValueObject-new}{\code{MeasurementValueObject$new()}}
\item \href{#method-MeasurementValueObject-toJSON}{\code{MeasurementValueObject$toJSON()}}
\item \href{#method-MeasurementValueObject-fromJSON}{\code{MeasurementValueObject$fromJSON()}}
\item \href{#method-MeasurementValueObject-toJSONString}{\code{MeasurementValueObject$toJSONString()}}
\item \href{#method-MeasurementValueObject-fromJSONString}{\code{MeasurementValueObject$fromJSONString()}}
\item \href{#method-MeasurementValueObject-validateJSON}{\code{MeasurementValueObject$validateJSON()}}
\item \href{#method-MeasurementValueObject-toString}{\code{MeasurementValueObject$toString()}}
\item \href{#method-MeasurementValueObject-isValid}{\code{MeasurementValueObject$isValid()}}
\item \href{#method-MeasurementValueObject-getInvalidFields}{\code{MeasurementValueObject$getInvalidFields()}}
\item \href{#method-MeasurementValueObject-print}{\code{MeasurementValueObject$print()}}
\item \href{#method-MeasurementValueObject-clone}{\code{MeasurementValueObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasurementValueObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-MeasurementValueObject-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new MeasurementValueObject class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasurementValueObject$new(
  id = NULL,
  value = NULL,
  unit = NULL,
  unitId = NULL,
  type = NULL,
  representation = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{value}}{value}

\item{\code{unit}}{unit}

\item{\code{unitId}}{unitId}

\item{\code{type}}{type}

\item{\code{representation}}{representation}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasurementValueObject-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MeasurementValueObject-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasurementValueObject$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
MeasurementValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasurementValueObject-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MeasurementValueObject-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of MeasurementValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasurementValueObject$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of MeasurementValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasurementValueObject-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-MeasurementValueObject-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasurementValueObject$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
MeasurementValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasurementValueObject-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-MeasurementValueObject-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of MeasurementValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasurementValueObject$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of MeasurementValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasurementValueObject-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MeasurementValueObject-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to MeasurementValueObject and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasurementValueObject$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasurementValueObject-toString"></a>}}
\if{latex}{\out{\hypertarget{method-MeasurementValueObject-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasurementValueObject$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of MeasurementValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasurementValueObject-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-MeasurementValueObject-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasurementValueObject$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasurementValueObject-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-MeasurementValueObject-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasurementValueObject$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasurementValueObject-print"></a>}}
\if{latex}{\out{\hypertarget{method-MeasurementValueObject-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasurementValueObject$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasurementValueObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MeasurementValueObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasurementValueObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
