% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{memgetDatasetDE}
\alias{memgetDatasetDE}
\title{Memoise getDatasetDE}
\usage{
memgetDatasetDE(
  datasets = NA_character_,
  keepNonSpecific = FALSE,
  diffExSet = NA_integer_,
  threshold = 100,
  limit = 100L,
  consolidate = NA_character_,
  raw = getOption("gemma.raw", FALSE),
  async = getOption("gemma.async", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)
}
\description{
Memoise getDatasetDE
}
\keyword{internal}
