% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{memgetDatasetDesign}
\alias{memgetDatasetDesign}
\title{Memoise getDatasetDesign}
\usage{
memgetDatasetDesign(
  dataset = NA_character_,
  raw = FALSE,
  async = FALSE,
  memoised = FALSE,
  file = NA_character_,
  overwrite = FALSE
)
}
\description{
Memoise getDatasetDesign
}
