% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{memgetDatasetPCA}
\alias{memgetDatasetPCA}
\title{Memoise getDatasetPCA}
\usage{
memgetDatasetPCA(
  datasets = NA_character_,
  component = 1L,
  limit = 100L,
  keepNonSpecific = FALSE,
  consolidate = NA_character_,
  raw = getOption("gemma.raw", FALSE),
  async = getOption("gemma.async", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)
}
\description{
Memoise getDatasetPCA
}
\keyword{internal}
