% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{memgetDatasets}
\alias{memgetDatasets}
\title{Memoise getDatasets}
\usage{
memgetDatasets(
  datasets = NA_character_,
  filter = NA_character_,
  offset = 0L,
  limit = 20L,
  sort = "+id",
  raw = FALSE,
  async = FALSE,
  memoised = FALSE
)
}
\description{
Memoise getDatasets
}
