% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{memgetDiffExpr}
\alias{memgetDiffExpr}
\title{Memoise getDiffExpr}
\usage{
memgetDiffExpr(
  dataset = NA_character_,
  offset = 0L,
  keepNonSpecific = FALSE,
  threshold = 100,
  limit = 100L,
  consolidate = NA_character_,
  raw = FALSE,
  async = FALSE,
  memoised = FALSE
)
}
\description{
Memoise getDiffExpr
}
