% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{memgetGeneEvidence}
\alias{memgetGeneEvidence}
\title{Memoise getGeneEvidence}
\usage{
memgetGeneEvidence(
  gene = NA_character_,
  raw = FALSE,
  async = FALSE,
  memoised = FALSE
)
}
\description{
Memoise getGeneEvidence
}
