% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{memgetGeneLocation}
\alias{memgetGeneLocation}
\title{Memoise getGeneLocation}
\usage{
memgetGeneLocation(
  gene = NA_character_,
  raw = FALSE,
  async = FALSE,
  memoised = FALSE
)
}
\description{
Memoise getGeneLocation
}
