% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{memgetTaxonPhenotypes}
\alias{memgetTaxonPhenotypes}
\title{Memoise getTaxonPhenotypes}
\usage{
memgetTaxonPhenotypes(
  taxon = NA_character_,
  editableOnly = FALSE,
  tree = FALSE,
  raw = getOption("gemma.raw", FALSE),
  async = getOption("gemma.async", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)
}
\description{
Memoise getTaxonPhenotypes
}
\keyword{internal}
