% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonArguments.R
\name{memoise}
\alias{memoise}
\title{Memoise doc}
\arguments{
\item{memoised}{Whether or not to save to cache for future calls with the
same inputs and use the result saved in cache if a result is already saved.
Doing \code{options(gemma.memoised = TRUE)} will ensure that the cache is always
used. Use \code{\link{forget_gemma_memoised}} to clear the cache.}
}
\description{
Memoise doc
}
