% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{memsearchDatasets}
\alias{memsearchDatasets}
\title{Memoise searchDatasets}
\usage{
memsearchDatasets(
  query = NA_character_,
  taxon = NA_character_,
  filter = NA_character_,
  offset = 0L,
  limit = 0L,
  sort = "+id",
  raw = getOption("gemma.raw", FALSE),
  async = getOption("gemma.async", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)
}
\description{
Memoise searchDatasets
}
\keyword{internal}
