% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifunctionality.R
\docType{class}
\name{Multifunctionality}
\alias{Multifunctionality}
\title{Multifunctionality}
\format{
An \code{R6Class} generator object
}
\description{
Multifunctionality Class

Multifunctionality Class
}
\details{
Create a new Multifunctionality
}
\keyword{(JSON}
\keyword{(if}
\keyword{Deserialize}
\keyword{Initialize}
\keyword{JSON}
\keyword{Multifunctionality}
\keyword{Return}
\keyword{To}
\keyword{Validate}
\keyword{a}
\keyword{all}
\keyword{an}
\keyword{any).}
\keyword{are}
\keyword{class.}
\keyword{fields}
\keyword{format)}
\keyword{if}
\keyword{in}
\keyword{input}
\keyword{instance}
\keyword{internal}
\keyword{into}
\keyword{invalid}
\keyword{list}
\keyword{new}
\keyword{of}
\keyword{respect}
\keyword{string}
\keyword{the}
\keyword{to}
\keyword{true}
\keyword{valid.}
\keyword{values}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{score}}{numeric \link{optional}}

\item{\code{rank}}{numeric \link{optional}}

\item{\code{numGoTerms}}{integer \link{optional}}

\item{\code{id}}{integer \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Multifunctionality-new}{\code{Multifunctionality$new()}}
\item \href{#method-Multifunctionality-toJSON}{\code{Multifunctionality$toJSON()}}
\item \href{#method-Multifunctionality-fromJSON}{\code{Multifunctionality$fromJSON()}}
\item \href{#method-Multifunctionality-toJSONString}{\code{Multifunctionality$toJSONString()}}
\item \href{#method-Multifunctionality-fromJSONString}{\code{Multifunctionality$fromJSONString()}}
\item \href{#method-Multifunctionality-validateJSON}{\code{Multifunctionality$validateJSON()}}
\item \href{#method-Multifunctionality-toString}{\code{Multifunctionality$toString()}}
\item \href{#method-Multifunctionality-isValid}{\code{Multifunctionality$isValid()}}
\item \href{#method-Multifunctionality-getInvalidFields}{\code{Multifunctionality$getInvalidFields()}}
\item \href{#method-Multifunctionality-print}{\code{Multifunctionality$print()}}
\item \href{#method-Multifunctionality-clone}{\code{Multifunctionality$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Multifunctionality-new"></a>}}
\if{latex}{\out{\hypertarget{method-Multifunctionality-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Multifunctionality class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Multifunctionality$new(
  score = NULL,
  rank = NULL,
  numGoTerms = NULL,
  id = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{score}}{score}

\item{\code{rank}}{rank}

\item{\code{numGoTerms}}{numGoTerms}

\item{\code{id}}{id}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Multifunctionality-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Multifunctionality-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Multifunctionality$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Multifunctionality in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Multifunctionality-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Multifunctionality-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Multifunctionality
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Multifunctionality$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Multifunctionality
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Multifunctionality-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Multifunctionality-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Multifunctionality$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Multifunctionality in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Multifunctionality-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Multifunctionality-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Multifunctionality
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Multifunctionality$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Multifunctionality
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Multifunctionality-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Multifunctionality-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Multifunctionality and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Multifunctionality$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Multifunctionality-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Multifunctionality-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Multifunctionality$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Multifunctionality
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Multifunctionality-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Multifunctionality-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Multifunctionality$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Multifunctionality-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Multifunctionality-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Multifunctionality$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Multifunctionality-print"></a>}}
\if{latex}{\out{\hypertarget{method-Multifunctionality-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Multifunctionality$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Multifunctionality-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Multifunctionality-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Multifunctionality$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
