% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subProcessors.R
\name{nullCheck}
\alias{nullCheck}
\title{Avoid NULLS as data.table columns}
\usage{
nullCheck(x, natype = NA)
}
\arguments{
\item{x}{A value that might be null}

\item{natype}{What to fill in when data is unavailable}
}
\value{
x as is or natypee
}
\description{
Avoid NULLS as data.table columns
}
\keyword{internal}
