% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/paginated_response_data_object_composite_sequence_value_object.R
\docType{class}
\name{PaginatedResponseDataObjectCompositeSequenceValueObject}
\alias{PaginatedResponseDataObjectCompositeSequenceValueObject}
\title{PaginatedResponseDataObjectCompositeSequenceValueObject}
\format{
An \code{R6Class} generator object
}
\description{
PaginatedResponseDataObjectCompositeSequenceValueObject Class

PaginatedResponseDataObjectCompositeSequenceValueObject Class
}
\details{
Create a new PaginatedResponseDataObjectCompositeSequenceValueObject
}
\keyword{(JSON}
\keyword{(if}
\keyword{Deserialize}
\keyword{Initialize}
\keyword{JSON}
\keyword{PaginatedResponseDataObjectCompositeSequenceValueObject}
\keyword{Return}
\keyword{To}
\keyword{Validate}
\keyword{a}
\keyword{all}
\keyword{an}
\keyword{any).}
\keyword{are}
\keyword{class.}
\keyword{fields}
\keyword{format)}
\keyword{if}
\keyword{in}
\keyword{input}
\keyword{instance}
\keyword{internal}
\keyword{into}
\keyword{invalid}
\keyword{list}
\keyword{new}
\keyword{of}
\keyword{respect}
\keyword{string}
\keyword{the}
\keyword{to}
\keyword{true}
\keyword{valid.}
\keyword{values}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list(\link{CompositeSequenceValueObject}) optional}

\item{\code{offset}}{integer optional}

\item{\code{limit}}{integer optional}

\item{\code{sort}}{\link{SortValueObject} optional}

\item{\code{totalElements}}{integer optional}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PaginatedResponseDataObjectCompositeSequenceValueObject-new}{\code{PaginatedResponseDataObjectCompositeSequenceValueObject$new()}}
\item \href{#method-PaginatedResponseDataObjectCompositeSequenceValueObject-toJSON}{\code{PaginatedResponseDataObjectCompositeSequenceValueObject$toJSON()}}
\item \href{#method-PaginatedResponseDataObjectCompositeSequenceValueObject-fromJSON}{\code{PaginatedResponseDataObjectCompositeSequenceValueObject$fromJSON()}}
\item \href{#method-PaginatedResponseDataObjectCompositeSequenceValueObject-toJSONString}{\code{PaginatedResponseDataObjectCompositeSequenceValueObject$toJSONString()}}
\item \href{#method-PaginatedResponseDataObjectCompositeSequenceValueObject-fromJSONString}{\code{PaginatedResponseDataObjectCompositeSequenceValueObject$fromJSONString()}}
\item \href{#method-PaginatedResponseDataObjectCompositeSequenceValueObject-validateJSON}{\code{PaginatedResponseDataObjectCompositeSequenceValueObject$validateJSON()}}
\item \href{#method-PaginatedResponseDataObjectCompositeSequenceValueObject-toString}{\code{PaginatedResponseDataObjectCompositeSequenceValueObject$toString()}}
\item \href{#method-PaginatedResponseDataObjectCompositeSequenceValueObject-isValid}{\code{PaginatedResponseDataObjectCompositeSequenceValueObject$isValid()}}
\item \href{#method-PaginatedResponseDataObjectCompositeSequenceValueObject-getInvalidFields}{\code{PaginatedResponseDataObjectCompositeSequenceValueObject$getInvalidFields()}}
\item \href{#method-PaginatedResponseDataObjectCompositeSequenceValueObject-print}{\code{PaginatedResponseDataObjectCompositeSequenceValueObject$print()}}
\item \href{#method-PaginatedResponseDataObjectCompositeSequenceValueObject-clone}{\code{PaginatedResponseDataObjectCompositeSequenceValueObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResponseDataObjectCompositeSequenceValueObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResponseDataObjectCompositeSequenceValueObject-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PaginatedResponseDataObjectCompositeSequenceValueObject class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResponseDataObjectCompositeSequenceValueObject$new(
  data = NULL,
  offset = NULL,
  limit = NULL,
  sort = NULL,
  totalElements = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{offset}}{offset}

\item{\code{limit}}{limit}

\item{\code{sort}}{sort}

\item{\code{totalElements}}{totalElements}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResponseDataObjectCompositeSequenceValueObject-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResponseDataObjectCompositeSequenceValueObject-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResponseDataObjectCompositeSequenceValueObject$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResponseDataObjectCompositeSequenceValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResponseDataObjectCompositeSequenceValueObject-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResponseDataObjectCompositeSequenceValueObject-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PaginatedResponseDataObjectCompositeSequenceValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResponseDataObjectCompositeSequenceValueObject$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResponseDataObjectCompositeSequenceValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResponseDataObjectCompositeSequenceValueObject-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResponseDataObjectCompositeSequenceValueObject-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResponseDataObjectCompositeSequenceValueObject$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResponseDataObjectCompositeSequenceValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResponseDataObjectCompositeSequenceValueObject-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResponseDataObjectCompositeSequenceValueObject-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PaginatedResponseDataObjectCompositeSequenceValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResponseDataObjectCompositeSequenceValueObject$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResponseDataObjectCompositeSequenceValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResponseDataObjectCompositeSequenceValueObject-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResponseDataObjectCompositeSequenceValueObject-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PaginatedResponseDataObjectCompositeSequenceValueObject and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResponseDataObjectCompositeSequenceValueObject$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResponseDataObjectCompositeSequenceValueObject-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResponseDataObjectCompositeSequenceValueObject-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResponseDataObjectCompositeSequenceValueObject$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PaginatedResponseDataObjectCompositeSequenceValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResponseDataObjectCompositeSequenceValueObject-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResponseDataObjectCompositeSequenceValueObject-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResponseDataObjectCompositeSequenceValueObject$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResponseDataObjectCompositeSequenceValueObject-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResponseDataObjectCompositeSequenceValueObject-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResponseDataObjectCompositeSequenceValueObject$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResponseDataObjectCompositeSequenceValueObject-print"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResponseDataObjectCompositeSequenceValueObject-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResponseDataObjectCompositeSequenceValueObject$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResponseDataObjectCompositeSequenceValueObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResponseDataObjectCompositeSequenceValueObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResponseDataObjectCompositeSequenceValueObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
