% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/paginated_response_data_object_expression_experiment_value_object.R
\docType{class}
\name{PaginatedResponseDataObjectExpressionExperimentValueObject}
\alias{PaginatedResponseDataObjectExpressionExperimentValueObject}
\title{PaginatedResponseDataObjectExpressionExperimentValueObject}
\format{
An \code{R6Class} generator object
}
\description{
PaginatedResponseDataObjectExpressionExperimentValueObject Class

PaginatedResponseDataObjectExpressionExperimentValueObject Class
}
\details{
Create a new PaginatedResponseDataObjectExpressionExperimentValueObject
}
\keyword{(JSON}
\keyword{(if}
\keyword{Deserialize}
\keyword{Initialize}
\keyword{JSON}
\keyword{PaginatedResponseDataObjectExpressionExperimentValueObject}
\keyword{Return}
\keyword{To}
\keyword{Validate}
\keyword{a}
\keyword{all}
\keyword{an}
\keyword{any).}
\keyword{are}
\keyword{class.}
\keyword{fields}
\keyword{format)}
\keyword{if}
\keyword{in}
\keyword{input}
\keyword{instance}
\keyword{internal}
\keyword{into}
\keyword{invalid}
\keyword{list}
\keyword{new}
\keyword{of}
\keyword{respect}
\keyword{string}
\keyword{the}
\keyword{to}
\keyword{true}
\keyword{valid.}
\keyword{values}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list(\link{ExpressionExperimentValueObject}) optional}

\item{\code{offset}}{integer optional}

\item{\code{limit}}{integer optional}

\item{\code{sort}}{\link{SortValueObject} optional}

\item{\code{totalElements}}{integer optional}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PaginatedResponseDataObjectExpressionExperimentValueObject-new}{\code{PaginatedResponseDataObjectExpressionExperimentValueObject$new()}}
\item \href{#method-PaginatedResponseDataObjectExpressionExperimentValueObject-toJSON}{\code{PaginatedResponseDataObjectExpressionExperimentValueObject$toJSON()}}
\item \href{#method-PaginatedResponseDataObjectExpressionExperimentValueObject-fromJSON}{\code{PaginatedResponseDataObjectExpressionExperimentValueObject$fromJSON()}}
\item \href{#method-PaginatedResponseDataObjectExpressionExperimentValueObject-toJSONString}{\code{PaginatedResponseDataObjectExpressionExperimentValueObject$toJSONString()}}
\item \href{#method-PaginatedResponseDataObjectExpressionExperimentValueObject-fromJSONString}{\code{PaginatedResponseDataObjectExpressionExperimentValueObject$fromJSONString()}}
\item \href{#method-PaginatedResponseDataObjectExpressionExperimentValueObject-validateJSON}{\code{PaginatedResponseDataObjectExpressionExperimentValueObject$validateJSON()}}
\item \href{#method-PaginatedResponseDataObjectExpressionExperimentValueObject-toString}{\code{PaginatedResponseDataObjectExpressionExperimentValueObject$toString()}}
\item \href{#method-PaginatedResponseDataObjectExpressionExperimentValueObject-isValid}{\code{PaginatedResponseDataObjectExpressionExperimentValueObject$isValid()}}
\item \href{#method-PaginatedResponseDataObjectExpressionExperimentValueObject-getInvalidFields}{\code{PaginatedResponseDataObjectExpressionExperimentValueObject$getInvalidFields()}}
\item \href{#method-PaginatedResponseDataObjectExpressionExperimentValueObject-print}{\code{PaginatedResponseDataObjectExpressionExperimentValueObject$print()}}
\item \href{#method-PaginatedResponseDataObjectExpressionExperimentValueObject-clone}{\code{PaginatedResponseDataObjectExpressionExperimentValueObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResponseDataObjectExpressionExperimentValueObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResponseDataObjectExpressionExperimentValueObject-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PaginatedResponseDataObjectExpressionExperimentValueObject class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResponseDataObjectExpressionExperimentValueObject$new(
  data = NULL,
  offset = NULL,
  limit = NULL,
  sort = NULL,
  totalElements = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{offset}}{offset}

\item{\code{limit}}{limit}

\item{\code{sort}}{sort}

\item{\code{totalElements}}{totalElements}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResponseDataObjectExpressionExperimentValueObject-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResponseDataObjectExpressionExperimentValueObject-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResponseDataObjectExpressionExperimentValueObject$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResponseDataObjectExpressionExperimentValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResponseDataObjectExpressionExperimentValueObject-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResponseDataObjectExpressionExperimentValueObject-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of PaginatedResponseDataObjectExpressionExperimentValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResponseDataObjectExpressionExperimentValueObject$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResponseDataObjectExpressionExperimentValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResponseDataObjectExpressionExperimentValueObject-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResponseDataObjectExpressionExperimentValueObject-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResponseDataObjectExpressionExperimentValueObject$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PaginatedResponseDataObjectExpressionExperimentValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResponseDataObjectExpressionExperimentValueObject-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResponseDataObjectExpressionExperimentValueObject-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of PaginatedResponseDataObjectExpressionExperimentValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResponseDataObjectExpressionExperimentValueObject$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of PaginatedResponseDataObjectExpressionExperimentValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResponseDataObjectExpressionExperimentValueObject-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResponseDataObjectExpressionExperimentValueObject-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to PaginatedResponseDataObjectExpressionExperimentValueObject and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResponseDataObjectExpressionExperimentValueObject$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResponseDataObjectExpressionExperimentValueObject-toString"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResponseDataObjectExpressionExperimentValueObject-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResponseDataObjectExpressionExperimentValueObject$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of PaginatedResponseDataObjectExpressionExperimentValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResponseDataObjectExpressionExperimentValueObject-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResponseDataObjectExpressionExperimentValueObject-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResponseDataObjectExpressionExperimentValueObject$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResponseDataObjectExpressionExperimentValueObject-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResponseDataObjectExpressionExperimentValueObject-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResponseDataObjectExpressionExperimentValueObject$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResponseDataObjectExpressionExperimentValueObject-print"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResponseDataObjectExpressionExperimentValueObject-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResponseDataObjectExpressionExperimentValueObject$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PaginatedResponseDataObjectExpressionExperimentValueObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PaginatedResponseDataObjectExpressionExperimentValueObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaginatedResponseDataObjectExpressionExperimentValueObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
