% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{process_search}
\alias{process_search}
\title{Returns the ids of the found results}
\usage{
process_search(d)
}
\value{
A data.table or a list of resultObjects
}
\description{
Returns the ids of the found results
}
\keyword{internal}
