% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processAnnotationFile}
\alias{processAnnotationFile}
\title{Processes a platform annotation file}
\usage{
processAnnotationFile(content)
}
\arguments{
\item{content}{The content from an \code{http_get} request}
}
\value{
A table of annotations
\itemize{
\item \code{ElementName}: Probeset names provided by the platform.
Gene symbols for generic annotations typically used for RNA-seq experiments.
\item \code{GeneSymbols}: Genes that were found to be aligned to
the probe sequence. Note that it is possible for probes to be
non-specific. Alignment to multiple genes are indicated with gene
symbols separated by "|"s
\item \code{GeneNames}: Name of the gene
\item \code{GOTerms}: GO Terms associated with the genes.
\item \code{GemmaIDs} and \code{NCBIids}: respective IDs for the genes.
}
}
\description{
Processes a platform annotation file
}
\keyword{internal}
