% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processAnnotationTerms}
\alias{processAnnotationTerms}
\title{Processes JSON as related annotation terms}
\usage{
processAnnotationTerms(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A data table with the annotation terms related to the queried term.
A list if \code{raw = TRUE}. A \code{400 error} if required parameters are missing.

The fields of the output data.table are:

\itemize{
\item \code{value.name}: Annotation term
\item \code{value.URI}: URI for the value.name
}
}
\description{
Processes JSON as related annotation terms
}
\keyword{internal}
