% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processCharacteristicBasicValueObject}
\alias{processCharacteristicBasicValueObject}
\title{Processes JSON as a factor}
\usage{
processCharacteristicBasicValueObject(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A processed data.table
}
\description{
Processes JSON as a factor
}
\keyword{internal}
