% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subProcessors.R
\name{processCharacteristicValueObject}
\alias{processCharacteristicValueObject}
\title{Processes JSON as a factor}
\usage{
processCharacteristicValueObject(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A processed data.table
}
\description{
Processes JSON as a factor
}
\keyword{internal}
