% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processCoexpression}
\alias{processCoexpression}
\title{Processes JSON as coexpression data}
\usage{
processCoexpression(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A processed data.table
}
\description{
Processes JSON as coexpression data
}
