% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_processors.R
\name{processDatasetResultSets}
\alias{processDatasetResultSets}
\title{Processes JSON as a datasets result set}
\usage{
processDatasetResultSets(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A processed data.table
}
\description{
Processes JSON as a datasets result set
}
\keyword{internal}
