% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_processors.R
\name{processDatasets}
\alias{processDatasets}
\title{Processes JSON as a vector of datasets}
\usage{
processDatasets(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A processed data.table
}
\description{
Processes JSON as a vector of datasets
}
\keyword{internal}
