% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processDEcontrasts}
\alias{processDEcontrasts}
\title{Replaces factor ids by the factors strings in DE table columns}
\usage{
processDEcontrasts(rs, rsID)
}
\arguments{
\item{rs}{The resultSet matrix to process}
}
\value{
A processed matrix
}
\description{
Replaces factor ids by the factors strings in DE table columns
}
\keyword{internal}
