% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processDEMatrix}
\alias{processDEMatrix}
\title{Processes differential expression matrix}
\usage{
processDEMatrix(m)
}
\arguments{
\item{m}{The differential expression matrix to process}
}
\value{
A processed matrix
}
\description{
Processes differential expression matrix
}
\keyword{internal}
