% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processDesignMatrix}
\alias{processDesignMatrix}
\title{Processes design matrix}
\usage{
processDesignMatrix(m)
}
\arguments{
\item{m}{The design matrix to process}
}
\value{
A processed matrix
}
\description{
Processes design matrix
}
\keyword{internal}
