% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processDifferentialExpressionAnalysisResultByGeneValueObject_tsv}
\alias{processDifferentialExpressionAnalysisResultByGeneValueObject_tsv}
\title{processDifferentialExpressionAnalysisResultByGeneValueObject_tsv}
\usage{
processDifferentialExpressionAnalysisResultByGeneValueObject_tsv(content)
}
\value{
A data.table containing differential expression results. This table
is stripped down some relevant information for speed of execution. Details about
the contrasts can be accessesed via \code{\link{get_result_sets}} function

The fields of the output data.table are:

\itemize{
\item \code{result.ID}: Result set ID of the differential expression analysis.
May represent multiple factors in a single model.
\item \code{contrast.ID}: Id of the specific contrast factor. Together with the result.ID
they uniquely represent a given contrast.
\item \code{experiment.ID}: Id of the source experiment
\item \code{factor.coefficient}: Model coefficient calculated for the specific contrast factor
\item \code{factor.logfc}: Log 2 fold change calculated for the specific contrast factor
\item \code{factor.pvalue}: p values calculated for the specific contrast factor
}
}
\description{
processDifferentialExpressionAnalysisResultByGeneValueObject_tsv
}
\keyword{internal}
