% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processExpression}
\alias{processExpression}
\title{Processes JSON as expression data}
\usage{
processExpression(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A processed data.table
}
\description{
Processes JSON as expression data
}
