% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processExpressionMatrix}
\alias{processExpressionMatrix}
\title{Processes expression matrix}
\usage{
processExpressionMatrix(m)
}
\arguments{
\item{m}{The matrix to process}
}
\value{
A processed matrix
}
\description{
Processes expression matrix
}
\keyword{internal}
