% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processFile}
\alias{processFile}
\title{Processes gzip file with expression data}
\usage{
processFile(response)
}
\arguments{
\item{d}{The file to process}
}
\value{
A processed data.table
}
\description{
Processes gzip file with expression data
}
