% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processGemmaFactor}
\alias{processGemmaFactor}
\title{Processes JSON as a factor}
\usage{
processGemmaFactor(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A processed data.table
}
\description{
Processes JSON as a factor
}
\keyword{internal}
