% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processGeneEvidence}
\alias{processGeneEvidence}
\title{Processes JSON as a vector of gene evidence (which also contains genes)}
\usage{
processGeneEvidence(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A processed data.table
}
\description{
Processes JSON as a vector of gene evidence (which also contains genes)
}
