% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processGenes}
\alias{processGenes}
\title{Processes JSON as a vector of genes}
\usage{
processGenes(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A data table with information about the querried gene(s)
A list if \code{raw = TRUE}.

The fields of the output data.table are:

\itemize{
\item \code{gene.Symbol}: Symbol for the gene
\item \code{gene.Ensembl}: Ensembl ID for the gene
\item \code{gene.NCBI}: NCBI id for the gene
\item \code{gene.Name}: Name of the gene
\item \code{taxon.name}: Name of the taxon of origin
\item \code{taxon.ID}: Gemma ID for the taxon
\item \code{taxon.Scientific}: Scientific name for the taxon
}
}
\description{
Processes JSON as a vector of genes
}
\keyword{internal}
