% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_processors.R
\name{processGO}
\alias{processGO}
\title{Processes JSON as GO terms}
\usage{
processGO(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A processed data.table
}
\description{
Processes JSON as GO terms
}
\keyword{internal}
