% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processPhenotypes}
\alias{processPhenotypes}
\title{Processes JSON as a vector of phenotypes}
\usage{
processPhenotypes(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A processed data.table
}
\description{
Processes JSON as a vector of phenotypes
}
