% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_processors.R
\name{processPlatforms}
\alias{processPlatforms}
\title{Processes JSON as a vector of platforms}
\usage{
processPlatforms(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A processed data.table
}
\description{
Processes JSON as a vector of platforms
}
\keyword{internal}
