% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processQuantitationTypeValueObject}
\alias{processQuantitationTypeValueObject}
\title{processQuantitationTypeValueObject}
\usage{
processQuantitationTypeValueObject(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A data.table containing the quantitation types

The fields of the output data.table are:

\itemize{
\item \code{id}: If of the quantitation type. Any raw quantitation type
can be accessed by \code{\link{get_dataset_raw_expression}} function using
this id.
\item \code{name}: Name of the quantitation type
\item \code{description}: Description of the quantitation type
\item \code{type}: Type of the quantitation type. Either raw or processed.
Each dataset will have one processed quantitation type which is the data
returned using \code{\link{get_dataset_processed_expression}}
\item \code{preferred}: The preferred raw quantitation type. This version
is used in generation of the processed data within gemma.
\item \code{recomputed}: If TRUE this quantitation type is generated by
recomputing raw data files Gemma had access to.
}
}
\description{
processQuantitationTypeValueObject
}
\keyword{internal}
