% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processResultSetFactors}
\alias{processResultSetFactors}
\title{Processes JSON as a result set}
\usage{
processResultSetFactors(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A processed data.table
}
\description{
Processes JSON as a result set
}
\keyword{internal}
