% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processSamples}
\alias{processSamples}
\title{Processes JSON as a vector of samples}
\usage{
processSamples(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A data table with information about the samples of the queried dataset. A list if
\code{raw = TRUE}. A \code{404 error} if the given identifier does not map to any object.

The fields of the output data.table are:
\itemize{
\item \code{sample.Name}: Internal name given to the sample.
\item \code{sample.ID}: Internal ID of the sample
\item \code{sample.Description}: Free text description of the sample
\item \code{sample.Outlier}: Whether or not the sample is marked as an outlier
\item \code{sample.Accession}: Accession ID of the sample in it's original database
\item \code{sample.Database}: Database of origin for the sample
\item \code{sample.Characteristics}: Characteristics of the sample. This field is a data table
\item \code{sample.FactorValues}: Experimental factor values of the sample. This field is a data table
}
}
\description{
Processes JSON as a vector of samples
}
\keyword{internal}
