% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processSVD}
\alias{processSVD}
\title{Processes JSON as an experiment's SVD}
\usage{
processSVD(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A processed data.table
}
\description{
Processes JSON as an experiment's SVD
}
