% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_processors.R
\name{processTaxon}
\alias{processTaxon}
\title{Processes JSON as a vector of taxa}
\usage{
processTaxon(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A processed data.table
}
\description{
Processes JSON as a vector of taxa
}
\keyword{internal}
