% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openapi_generated.R
\docType{class}
\name{QuantitationType}
\alias{QuantitationType}
\title{QuantitationType}
\format{
An \code{R6Class} generator object

An \code{R6Class} generator object
}
\description{
QuantitationType Class

QuantitationType Class

QuantitationType Class

QuantitationType Class
}
\details{
Create a new QuantitationType

Create a new QuantitationType
}
\keyword{(JSON}
\keyword{(if}
\keyword{Deserialize}
\keyword{Initialize}
\keyword{JSON}
\keyword{Print}
\keyword{QuantitationType}
\keyword{Return}
\keyword{To}
\keyword{Validate}
\keyword{a}
\keyword{all}
\keyword{an}
\keyword{any).}
\keyword{are}
\keyword{class.}
\keyword{fields}
\keyword{format)}
\keyword{if}
\keyword{in}
\keyword{input}
\keyword{instance}
\keyword{internal}
\keyword{into}
\keyword{invalid}
\keyword{list}
\keyword{new}
\keyword{object}
\keyword{of}
\keyword{respect}
\keyword{string}
\keyword{the}
\keyword{to}
\keyword{true}
\keyword{valid.}
\keyword{values}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character optional}

\item{\code{description}}{character optional}

\item{\code{id}}{integer optional}

\item{\code{isBackground}}{character optional}

\item{\code{isBackgroundSubtracted}}{character optional}

\item{\code{isBatchCorrected}}{character optional}

\item{\code{isMaskedPreferred}}{character optional}

\item{\code{isNormalized}}{character optional}

\item{\code{isPreferred}}{character optional}

\item{\code{isRecomputedFromRawData}}{character optional}

\item{\code{isRatio}}{character optional}

\item{\code{generalType}}{\link{GeneralType} optional}

\item{\code{representation}}{\link{PrimitiveType} optional}

\item{\code{scale}}{\link{ScaleType} optional}

\item{\code{type}}{\link{StandardQuantitationType} optional}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QuantitationType-new}{\code{QuantitationType$new()}}
\item \href{#method-QuantitationType-toJSON}{\code{QuantitationType$toJSON()}}
\item \href{#method-QuantitationType-fromJSON}{\code{QuantitationType$fromJSON()}}
\item \href{#method-QuantitationType-toJSONString}{\code{QuantitationType$toJSONString()}}
\item \href{#method-QuantitationType-fromJSONString}{\code{QuantitationType$fromJSONString()}}
\item \href{#method-QuantitationType-validateJSON}{\code{QuantitationType$validateJSON()}}
\item \href{#method-QuantitationType-toString}{\code{QuantitationType$toString()}}
\item \href{#method-QuantitationType-isValid}{\code{QuantitationType$isValid()}}
\item \href{#method-QuantitationType-getInvalidFields}{\code{QuantitationType$getInvalidFields()}}
\item \href{#method-QuantitationType-print}{\code{QuantitationType$print()}}
\item \href{#method-QuantitationType-clone}{\code{QuantitationType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantitationType-new"></a>}}
\if{latex}{\out{\hypertarget{method-QuantitationType-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new QuantitationType class.


To JSON String


Deserialize JSON string into an instance of QuantitationType


To JSON String


Deserialize JSON string into an instance of QuantitationType


Validate JSON input with respect to QuantitationType and throw an exception if invalid


To string (JSON format)


Return true if the values in all fields are valid.


Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantitationType$new(
  name = NULL,
  description = NULL,
  id = NULL,
  isBackground = NULL,
  isBackgroundSubtracted = NULL,
  isBatchCorrected = NULL,
  isMaskedPreferred = NULL,
  isNormalized = NULL,
  isPreferred = NULL,
  isRecomputedFromRawData = NULL,
  isRatio = NULL,
  generalType = NULL,
  representation = NULL,
  scale = NULL,
  type = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}

\item{\code{description}}{description}

\item{\code{id}}{id}

\item{\code{isBackground}}{isBackground}

\item{\code{isBackgroundSubtracted}}{isBackgroundSubtracted}

\item{\code{isBatchCorrected}}{isBatchCorrected}

\item{\code{isMaskedPreferred}}{isMaskedPreferred}

\item{\code{isNormalized}}{isNormalized}

\item{\code{isPreferred}}{isPreferred}

\item{\code{isRecomputedFromRawData}}{isRecomputedFromRawData}

\item{\code{isRatio}}{isRatio}

\item{\code{generalType}}{generalType}

\item{\code{representation}}{representation}

\item{\code{scale}}{scale}

\item{\code{type}}{type}

\item{\code{...}}{Other optional arguments.}

\item{\code{input_json}}{the JSON input}

\item{\code{input_json}}{the JSON input}

\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
QuantitationType in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantitationType-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QuantitationType-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantitationType$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantitationType-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QuantitationType-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantitationType$fromJSON(input_json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantitationType-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-QuantitationType-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantitationType$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantitationType-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-QuantitationType-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantitationType$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantitationType-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QuantitationType-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantitationType$validateJSON(input)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantitationType-toString"></a>}}
\if{latex}{\out{\hypertarget{method-QuantitationType-toString}{}}}
\subsection{Method \code{toString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantitationType$toString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantitationType-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-QuantitationType-isValid}{}}}
\subsection{Method \code{isValid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantitationType$isValid()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantitationType-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-QuantitationType-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantitationType$getInvalidFields()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantitationType-print"></a>}}
\if{latex}{\out{\hypertarget{method-QuantitationType-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantitationType$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantitationType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QuantitationType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantitationType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character optional}

\item{\code{description}}{character optional}

\item{\code{id}}{integer optional}

\item{\code{isBackground}}{character optional}

\item{\code{isBackgroundSubtracted}}{character optional}

\item{\code{isBatchCorrected}}{character optional}

\item{\code{isMaskedPreferred}}{character optional}

\item{\code{isNormalized}}{character optional}

\item{\code{isPreferred}}{character optional}

\item{\code{isRecomputedFromRawData}}{character optional}

\item{\code{isRatio}}{character optional}

\item{\code{generalType}}{\link{GeneralType} optional}

\item{\code{representation}}{\link{PrimitiveType} optional}

\item{\code{scale}}{\link{ScaleType} optional}

\item{\code{type}}{\link{StandardQuantitationType} optional}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QuantitationType-new}{\code{QuantitationType$new()}}
\item \href{#method-QuantitationType-toJSON}{\code{QuantitationType$toJSON()}}
\item \href{#method-QuantitationType-fromJSON}{\code{QuantitationType$fromJSON()}}
\item \href{#method-QuantitationType-toJSONString}{\code{QuantitationType$toJSONString()}}
\item \href{#method-QuantitationType-fromJSONString}{\code{QuantitationType$fromJSONString()}}
\item \href{#method-QuantitationType-validateJSON}{\code{QuantitationType$validateJSON()}}
\item \href{#method-QuantitationType-toString}{\code{QuantitationType$toString()}}
\item \href{#method-QuantitationType-isValid}{\code{QuantitationType$isValid()}}
\item \href{#method-QuantitationType-getInvalidFields}{\code{QuantitationType$getInvalidFields()}}
\item \href{#method-QuantitationType-print}{\code{QuantitationType$print()}}
\item \href{#method-QuantitationType-clone}{\code{QuantitationType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantitationType-new"></a>}}
\if{latex}{\out{\hypertarget{method-QuantitationType-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new QuantitationType class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantitationType$new(
  name = NULL,
  description = NULL,
  id = NULL,
  isBackground = NULL,
  isBackgroundSubtracted = NULL,
  isBatchCorrected = NULL,
  isMaskedPreferred = NULL,
  isNormalized = NULL,
  isPreferred = NULL,
  isRecomputedFromRawData = NULL,
  isRatio = NULL,
  generalType = NULL,
  representation = NULL,
  scale = NULL,
  type = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}

\item{\code{description}}{description}

\item{\code{id}}{id}

\item{\code{isBackground}}{isBackground}

\item{\code{isBackgroundSubtracted}}{isBackgroundSubtracted}

\item{\code{isBatchCorrected}}{isBatchCorrected}

\item{\code{isMaskedPreferred}}{isMaskedPreferred}

\item{\code{isNormalized}}{isNormalized}

\item{\code{isPreferred}}{isPreferred}

\item{\code{isRecomputedFromRawData}}{isRecomputedFromRawData}

\item{\code{isRatio}}{isRatio}

\item{\code{generalType}}{generalType}

\item{\code{representation}}{representation}

\item{\code{scale}}{scale}

\item{\code{type}}{type}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantitationType-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QuantitationType-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantitationType$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
QuantitationType in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantitationType-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QuantitationType-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of QuantitationType
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantitationType$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of QuantitationType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantitationType-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-QuantitationType-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantitationType$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
QuantitationType in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantitationType-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-QuantitationType-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of QuantitationType
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantitationType$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of QuantitationType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantitationType-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QuantitationType-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to QuantitationType and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantitationType$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantitationType-toString"></a>}}
\if{latex}{\out{\hypertarget{method-QuantitationType-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantitationType$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of QuantitationType
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantitationType-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-QuantitationType-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantitationType$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantitationType-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-QuantitationType-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantitationType$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantitationType-print"></a>}}
\if{latex}{\out{\hypertarget{method-QuantitationType-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantitationType$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuantitationType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QuantitationType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantitationType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
