% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openapi_generated.R
\docType{class}
\name{ResponseDataObjectListBioAssayValueObject}
\alias{ResponseDataObjectListBioAssayValueObject}
\title{ResponseDataObjectListBioAssayValueObject}
\format{
An \code{R6Class} generator object
}
\description{
ResponseDataObjectListBioAssayValueObject Class

ResponseDataObjectListBioAssayValueObject Class
}
\details{
Create a new ResponseDataObjectListBioAssayValueObject
}
\keyword{(JSON}
\keyword{(if}
\keyword{Deserialize}
\keyword{Initialize}
\keyword{JSON}
\keyword{ResponseDataObjectListBioAssayValueObject}
\keyword{Return}
\keyword{To}
\keyword{Validate}
\keyword{a}
\keyword{all}
\keyword{an}
\keyword{any).}
\keyword{are}
\keyword{class.}
\keyword{fields}
\keyword{format)}
\keyword{if}
\keyword{in}
\keyword{input}
\keyword{instance}
\keyword{internal}
\keyword{into}
\keyword{invalid}
\keyword{list}
\keyword{new}
\keyword{of}
\keyword{respect}
\keyword{string}
\keyword{the}
\keyword{to}
\keyword{true}
\keyword{valid.}
\keyword{values}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list(\link{BioAssayValueObject}) optional}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResponseDataObjectListBioAssayValueObject-new}{\code{ResponseDataObjectListBioAssayValueObject$new()}}
\item \href{#method-ResponseDataObjectListBioAssayValueObject-toJSON}{\code{ResponseDataObjectListBioAssayValueObject$toJSON()}}
\item \href{#method-ResponseDataObjectListBioAssayValueObject-fromJSON}{\code{ResponseDataObjectListBioAssayValueObject$fromJSON()}}
\item \href{#method-ResponseDataObjectListBioAssayValueObject-toJSONString}{\code{ResponseDataObjectListBioAssayValueObject$toJSONString()}}
\item \href{#method-ResponseDataObjectListBioAssayValueObject-fromJSONString}{\code{ResponseDataObjectListBioAssayValueObject$fromJSONString()}}
\item \href{#method-ResponseDataObjectListBioAssayValueObject-validateJSON}{\code{ResponseDataObjectListBioAssayValueObject$validateJSON()}}
\item \href{#method-ResponseDataObjectListBioAssayValueObject-toString}{\code{ResponseDataObjectListBioAssayValueObject$toString()}}
\item \href{#method-ResponseDataObjectListBioAssayValueObject-isValid}{\code{ResponseDataObjectListBioAssayValueObject$isValid()}}
\item \href{#method-ResponseDataObjectListBioAssayValueObject-getInvalidFields}{\code{ResponseDataObjectListBioAssayValueObject$getInvalidFields()}}
\item \href{#method-ResponseDataObjectListBioAssayValueObject-print}{\code{ResponseDataObjectListBioAssayValueObject$print()}}
\item \href{#method-ResponseDataObjectListBioAssayValueObject-clone}{\code{ResponseDataObjectListBioAssayValueObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListBioAssayValueObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListBioAssayValueObject-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ResponseDataObjectListBioAssayValueObject class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListBioAssayValueObject$new(data = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListBioAssayValueObject-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListBioAssayValueObject-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListBioAssayValueObject$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ResponseDataObjectListBioAssayValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListBioAssayValueObject-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListBioAssayValueObject-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ResponseDataObjectListBioAssayValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListBioAssayValueObject$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ResponseDataObjectListBioAssayValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListBioAssayValueObject-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListBioAssayValueObject-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListBioAssayValueObject$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ResponseDataObjectListBioAssayValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListBioAssayValueObject-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListBioAssayValueObject-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ResponseDataObjectListBioAssayValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListBioAssayValueObject$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ResponseDataObjectListBioAssayValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListBioAssayValueObject-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListBioAssayValueObject-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ResponseDataObjectListBioAssayValueObject and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListBioAssayValueObject$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListBioAssayValueObject-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListBioAssayValueObject-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListBioAssayValueObject$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ResponseDataObjectListBioAssayValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListBioAssayValueObject-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListBioAssayValueObject-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListBioAssayValueObject$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListBioAssayValueObject-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListBioAssayValueObject-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListBioAssayValueObject$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListBioAssayValueObject-print"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListBioAssayValueObject-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListBioAssayValueObject$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListBioAssayValueObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListBioAssayValueObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListBioAssayValueObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
