% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openapi_generated.R
\docType{class}
\name{ResponseDataObjectListDifferentialExpressionAnalysisValueObject}
\alias{ResponseDataObjectListDifferentialExpressionAnalysisValueObject}
\title{ResponseDataObjectListDifferentialExpressionAnalysisValueObject}
\format{
An \code{R6Class} generator object
}
\description{
ResponseDataObjectListDifferentialExpressionAnalysisValueObject Class

ResponseDataObjectListDifferentialExpressionAnalysisValueObject Class
}
\details{
Create a new ResponseDataObjectListDifferentialExpressionAnalysisValueObject
}
\keyword{(JSON}
\keyword{(if}
\keyword{Deserialize}
\keyword{Initialize}
\keyword{JSON}
\keyword{ResponseDataObjectListDifferentialExpressionAnalysisValueObject}
\keyword{Return}
\keyword{To}
\keyword{Validate}
\keyword{a}
\keyword{all}
\keyword{an}
\keyword{any).}
\keyword{are}
\keyword{class.}
\keyword{fields}
\keyword{format)}
\keyword{if}
\keyword{in}
\keyword{input}
\keyword{instance}
\keyword{internal}
\keyword{into}
\keyword{invalid}
\keyword{list}
\keyword{new}
\keyword{of}
\keyword{respect}
\keyword{string}
\keyword{the}
\keyword{to}
\keyword{true}
\keyword{valid.}
\keyword{values}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list(\link{DifferentialExpressionAnalysisValueObject}) optional}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-new}{\code{ResponseDataObjectListDifferentialExpressionAnalysisValueObject$new()}}
\item \href{#method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-toJSON}{\code{ResponseDataObjectListDifferentialExpressionAnalysisValueObject$toJSON()}}
\item \href{#method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-fromJSON}{\code{ResponseDataObjectListDifferentialExpressionAnalysisValueObject$fromJSON()}}
\item \href{#method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-toJSONString}{\code{ResponseDataObjectListDifferentialExpressionAnalysisValueObject$toJSONString()}}
\item \href{#method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-fromJSONString}{\code{ResponseDataObjectListDifferentialExpressionAnalysisValueObject$fromJSONString()}}
\item \href{#method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-validateJSON}{\code{ResponseDataObjectListDifferentialExpressionAnalysisValueObject$validateJSON()}}
\item \href{#method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-toString}{\code{ResponseDataObjectListDifferentialExpressionAnalysisValueObject$toString()}}
\item \href{#method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-isValid}{\code{ResponseDataObjectListDifferentialExpressionAnalysisValueObject$isValid()}}
\item \href{#method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-getInvalidFields}{\code{ResponseDataObjectListDifferentialExpressionAnalysisValueObject$getInvalidFields()}}
\item \href{#method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-print}{\code{ResponseDataObjectListDifferentialExpressionAnalysisValueObject$print()}}
\item \href{#method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-clone}{\code{ResponseDataObjectListDifferentialExpressionAnalysisValueObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ResponseDataObjectListDifferentialExpressionAnalysisValueObject class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListDifferentialExpressionAnalysisValueObject$new(
  data = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListDifferentialExpressionAnalysisValueObject$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ResponseDataObjectListDifferentialExpressionAnalysisValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ResponseDataObjectListDifferentialExpressionAnalysisValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListDifferentialExpressionAnalysisValueObject$fromJSON(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ResponseDataObjectListDifferentialExpressionAnalysisValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListDifferentialExpressionAnalysisValueObject$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ResponseDataObjectListDifferentialExpressionAnalysisValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ResponseDataObjectListDifferentialExpressionAnalysisValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListDifferentialExpressionAnalysisValueObject$fromJSONString(
  input_json
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ResponseDataObjectListDifferentialExpressionAnalysisValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ResponseDataObjectListDifferentialExpressionAnalysisValueObject and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListDifferentialExpressionAnalysisValueObject$validateJSON(
  input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListDifferentialExpressionAnalysisValueObject$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ResponseDataObjectListDifferentialExpressionAnalysisValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListDifferentialExpressionAnalysisValueObject$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListDifferentialExpressionAnalysisValueObject$getInvalidFields(
  
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-print"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListDifferentialExpressionAnalysisValueObject$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListDifferentialExpressionAnalysisValueObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListDifferentialExpressionAnalysisValueObject$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
