% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openapi_generated.R
\docType{class}
\name{ResponseDataObjectListPhysicalLocationValueObject}
\alias{ResponseDataObjectListPhysicalLocationValueObject}
\title{ResponseDataObjectListPhysicalLocationValueObject}
\format{
An \code{R6Class} generator object

An \code{R6Class} generator object
}
\description{
ResponseDataObjectListPhysicalLocationValueObject Class

ResponseDataObjectListPhysicalLocationValueObject Class

ResponseDataObjectListPhysicalLocationValueObject Class

ResponseDataObjectListPhysicalLocationValueObject Class
}
\details{
Create a new ResponseDataObjectListPhysicalLocationValueObject

Create a new ResponseDataObjectListPhysicalLocationValueObject
}
\keyword{(JSON}
\keyword{(if}
\keyword{Deserialize}
\keyword{Initialize}
\keyword{JSON}
\keyword{Print}
\keyword{ResponseDataObjectListPhysicalLocationValueObject}
\keyword{Return}
\keyword{To}
\keyword{Validate}
\keyword{a}
\keyword{all}
\keyword{an}
\keyword{any).}
\keyword{are}
\keyword{class.}
\keyword{fields}
\keyword{format)}
\keyword{if}
\keyword{in}
\keyword{input}
\keyword{instance}
\keyword{internal}
\keyword{into}
\keyword{invalid}
\keyword{list}
\keyword{new}
\keyword{object}
\keyword{of}
\keyword{respect}
\keyword{string}
\keyword{the}
\keyword{to}
\keyword{true}
\keyword{valid.}
\keyword{values}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list(\link{PhysicalLocationValueObject}) optional}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResponseDataObjectListPhysicalLocationValueObject-new}{\code{ResponseDataObjectListPhysicalLocationValueObject$new()}}
\item \href{#method-ResponseDataObjectListPhysicalLocationValueObject-toJSON}{\code{ResponseDataObjectListPhysicalLocationValueObject$toJSON()}}
\item \href{#method-ResponseDataObjectListPhysicalLocationValueObject-fromJSON}{\code{ResponseDataObjectListPhysicalLocationValueObject$fromJSON()}}
\item \href{#method-ResponseDataObjectListPhysicalLocationValueObject-toJSONString}{\code{ResponseDataObjectListPhysicalLocationValueObject$toJSONString()}}
\item \href{#method-ResponseDataObjectListPhysicalLocationValueObject-fromJSONString}{\code{ResponseDataObjectListPhysicalLocationValueObject$fromJSONString()}}
\item \href{#method-ResponseDataObjectListPhysicalLocationValueObject-validateJSON}{\code{ResponseDataObjectListPhysicalLocationValueObject$validateJSON()}}
\item \href{#method-ResponseDataObjectListPhysicalLocationValueObject-toString}{\code{ResponseDataObjectListPhysicalLocationValueObject$toString()}}
\item \href{#method-ResponseDataObjectListPhysicalLocationValueObject-isValid}{\code{ResponseDataObjectListPhysicalLocationValueObject$isValid()}}
\item \href{#method-ResponseDataObjectListPhysicalLocationValueObject-getInvalidFields}{\code{ResponseDataObjectListPhysicalLocationValueObject$getInvalidFields()}}
\item \href{#method-ResponseDataObjectListPhysicalLocationValueObject-print}{\code{ResponseDataObjectListPhysicalLocationValueObject$print()}}
\item \href{#method-ResponseDataObjectListPhysicalLocationValueObject-clone}{\code{ResponseDataObjectListPhysicalLocationValueObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListPhysicalLocationValueObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListPhysicalLocationValueObject-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ResponseDataObjectListPhysicalLocationValueObject class.


To JSON String


Deserialize JSON string into an instance of ResponseDataObjectListPhysicalLocationValueObject


To JSON String


Deserialize JSON string into an instance of ResponseDataObjectListPhysicalLocationValueObject


Validate JSON input with respect to ResponseDataObjectListPhysicalLocationValueObject and throw an exception if invalid


To string (JSON format)


Return true if the values in all fields are valid.


Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListPhysicalLocationValueObject$new(data = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{...}}{Other optional arguments.}

\item{\code{input_json}}{the JSON input}

\item{\code{input_json}}{the JSON input}

\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ResponseDataObjectListPhysicalLocationValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListPhysicalLocationValueObject-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListPhysicalLocationValueObject-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListPhysicalLocationValueObject$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListPhysicalLocationValueObject-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListPhysicalLocationValueObject-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListPhysicalLocationValueObject$fromJSON(input_json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListPhysicalLocationValueObject-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListPhysicalLocationValueObject-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListPhysicalLocationValueObject$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListPhysicalLocationValueObject-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListPhysicalLocationValueObject-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListPhysicalLocationValueObject$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListPhysicalLocationValueObject-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListPhysicalLocationValueObject-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListPhysicalLocationValueObject$validateJSON(input)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListPhysicalLocationValueObject-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListPhysicalLocationValueObject-toString}{}}}
\subsection{Method \code{toString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListPhysicalLocationValueObject$toString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListPhysicalLocationValueObject-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListPhysicalLocationValueObject-isValid}{}}}
\subsection{Method \code{isValid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListPhysicalLocationValueObject$isValid()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListPhysicalLocationValueObject-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListPhysicalLocationValueObject-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListPhysicalLocationValueObject$getInvalidFields()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListPhysicalLocationValueObject-print"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListPhysicalLocationValueObject-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListPhysicalLocationValueObject$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListPhysicalLocationValueObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListPhysicalLocationValueObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListPhysicalLocationValueObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{list(\link{PhysicalLocationValueObject}) optional}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResponseDataObjectListPhysicalLocationValueObject-new}{\code{ResponseDataObjectListPhysicalLocationValueObject$new()}}
\item \href{#method-ResponseDataObjectListPhysicalLocationValueObject-toJSON}{\code{ResponseDataObjectListPhysicalLocationValueObject$toJSON()}}
\item \href{#method-ResponseDataObjectListPhysicalLocationValueObject-fromJSON}{\code{ResponseDataObjectListPhysicalLocationValueObject$fromJSON()}}
\item \href{#method-ResponseDataObjectListPhysicalLocationValueObject-toJSONString}{\code{ResponseDataObjectListPhysicalLocationValueObject$toJSONString()}}
\item \href{#method-ResponseDataObjectListPhysicalLocationValueObject-fromJSONString}{\code{ResponseDataObjectListPhysicalLocationValueObject$fromJSONString()}}
\item \href{#method-ResponseDataObjectListPhysicalLocationValueObject-validateJSON}{\code{ResponseDataObjectListPhysicalLocationValueObject$validateJSON()}}
\item \href{#method-ResponseDataObjectListPhysicalLocationValueObject-toString}{\code{ResponseDataObjectListPhysicalLocationValueObject$toString()}}
\item \href{#method-ResponseDataObjectListPhysicalLocationValueObject-isValid}{\code{ResponseDataObjectListPhysicalLocationValueObject$isValid()}}
\item \href{#method-ResponseDataObjectListPhysicalLocationValueObject-getInvalidFields}{\code{ResponseDataObjectListPhysicalLocationValueObject$getInvalidFields()}}
\item \href{#method-ResponseDataObjectListPhysicalLocationValueObject-print}{\code{ResponseDataObjectListPhysicalLocationValueObject$print()}}
\item \href{#method-ResponseDataObjectListPhysicalLocationValueObject-clone}{\code{ResponseDataObjectListPhysicalLocationValueObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListPhysicalLocationValueObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListPhysicalLocationValueObject-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ResponseDataObjectListPhysicalLocationValueObject class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListPhysicalLocationValueObject$new(data = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListPhysicalLocationValueObject-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListPhysicalLocationValueObject-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListPhysicalLocationValueObject$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ResponseDataObjectListPhysicalLocationValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListPhysicalLocationValueObject-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListPhysicalLocationValueObject-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ResponseDataObjectListPhysicalLocationValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListPhysicalLocationValueObject$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ResponseDataObjectListPhysicalLocationValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListPhysicalLocationValueObject-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListPhysicalLocationValueObject-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListPhysicalLocationValueObject$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ResponseDataObjectListPhysicalLocationValueObject in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListPhysicalLocationValueObject-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListPhysicalLocationValueObject-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ResponseDataObjectListPhysicalLocationValueObject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListPhysicalLocationValueObject$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ResponseDataObjectListPhysicalLocationValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListPhysicalLocationValueObject-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListPhysicalLocationValueObject-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ResponseDataObjectListPhysicalLocationValueObject and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListPhysicalLocationValueObject$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListPhysicalLocationValueObject-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListPhysicalLocationValueObject-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListPhysicalLocationValueObject$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ResponseDataObjectListPhysicalLocationValueObject
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListPhysicalLocationValueObject-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListPhysicalLocationValueObject-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListPhysicalLocationValueObject$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListPhysicalLocationValueObject-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListPhysicalLocationValueObject-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListPhysicalLocationValueObject$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListPhysicalLocationValueObject-print"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListPhysicalLocationValueObject-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListPhysicalLocationValueObject$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResponseDataObjectListPhysicalLocationValueObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResponseDataObjectListPhysicalLocationValueObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResponseDataObjectListPhysicalLocationValueObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
