% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{search_datasets}
\alias{search_datasets}
\title{Retrieve datasets within a given taxa associated to an annotation tags search}
\usage{
search_datasets(
  query,
  taxon = NA_character_,
  offset = 0L,
  limit = 20L,
  sort = "+id",
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE),
  attributes = getOption("gemma.attributes", TRUE)
)
}
\arguments{
\item{query}{The search query. Either plain text ('traumatic'), or an ontology
term (UBERON_0002048). Datasets that contain the given string in their short
or full name will also be matched. Can be multiple identifiers separated by commas.}

\item{taxon}{A numerical taxon identifier or an ncbi taxon identifier or a taxon identifier that matches either its scientific or common name}

\item{offset}{The offset of the first retrieved result.}

\item{limit}{Optional, defaults to 20. Limits the result to specified amount
of objects. Has a maximum value of 100. Use together with \code{offset} and
the \code{totalElements} \link[base:attributes]{attribute} in the output to
compile all data if needed.}

\item{sort}{Order results by the given property and direction. The '+' sign indicate ascending order whereas the '-' indicate descending.}

\item{raw}{\code{TRUE} to receive results as-is from Gemma, or \code{FALSE} to enable
parsing. Raw results usually contain additional fields and flags that are
omitted in the parsed results.}

\item{memoised}{Whether or not to save to cache for future calls with the
same inputs and use the result saved in cache if a result is already saved.
Doing \code{options(gemma.memoised = TRUE)} will ensure that the cache is always
used. Use \code{\link{forget_gemma_memoised}} to clear the cache.}

\item{file}{The name of a file to save the results to, or \code{NULL} to not write
results to a file. If \code{raw == TRUE}, the output will be a JSON file. Otherwise,
it will be a RDS file.}

\item{overwrite}{Whether or not to overwrite if a file exists at the specified
filename.}

\item{attributes}{If \code{TRUE} additional information from the call will be added
into the output object's attributes such as offset and available elements.}
}
\value{
A data table with information about matching datasets. Returns an
empty list if no datasets found. Lists dataset (expression experiment value
objects) that are annotated with the given ontology terms, or, in case of
plaintext query, experiments that contain the given words (name, short name,
accession, tags) If an ontology term URI is given, the results will also
include datasets that are associated with the descendants of the term. The
search only only checks the annotations value, not the category (which is
also an ontology term).
}
\description{
Retrieve datasets within a given taxa associated to an annotation tags search
}
\examples{
search_datasets("bipolar")
}
\keyword{dataset}
