% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{searchDatasets}
\alias{searchDatasets}
\title{searchDatasets}
\usage{
searchDatasets(
  taxon = "",
  query = NA_character_,
  filter = NA_character_,
  offset = 0L,
  limit = 0L,
  sort = "+id",
  raw = FALSE,
  async = FALSE,
  memoised = FALSE
)
}
\description{
searchDatasets
}
