% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{setGemmaPath}
\alias{setGemmaPath}
\title{Set gemma path}
\usage{
setGemmaPath(path)
}
\arguments{
\item{path}{"dev", "prod" or a link to use to access gemma API}
}
\value{
Link to Gemma API
}
\description{
Set gemma path
}
\keyword{internal}
