% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{setGemmaUser}
\alias{setGemmaUser}
\title{Authentication}
\usage{
setGemmaUser(username = NULL, password = NULL)
}
\arguments{
\item{username}{Your username (or empty, if logging out)}

\item{password}{Your password (or empty, if logging out)}
}
\description{
Allows the user to access information that requires logging in to Gemma. To log out, run \code{setGemmaUser} without specifying the username or password.
}
\keyword{internal}
