% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{subset_factorValues}
\alias{subset_factorValues}
\title{Get a subset of an array of factorValues}
\usage{
subset_factorValues(
  factorValues,
  factorValue = NULL,
  differential_expressions = NULL,
  resultSet = NULL,
  contrast = NULL
)
}
\arguments{
\item{factorValue}{unimplemented}
}
\value{
a boolean vector, samples representing the resultSet and/or the contrast
are set to TRUE
}
\description{
Get a subset of an array of factorValues
}
\keyword{internal}
