% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openapi_generated.R
\docType{class}
\name{Taxon}
\alias{Taxon}
\title{Taxon}
\format{
An \code{R6Class} generator object
}
\description{
Taxon Class

Taxon Class
}
\details{
Create a new Taxon
}
\keyword{(JSON}
\keyword{(if}
\keyword{Deserialize}
\keyword{Initialize}
\keyword{JSON}
\keyword{Return}
\keyword{Taxon}
\keyword{To}
\keyword{Validate}
\keyword{a}
\keyword{all}
\keyword{an}
\keyword{any).}
\keyword{are}
\keyword{class.}
\keyword{fields}
\keyword{format)}
\keyword{if}
\keyword{in}
\keyword{input}
\keyword{instance}
\keyword{internal}
\keyword{into}
\keyword{invalid}
\keyword{list}
\keyword{new}
\keyword{of}
\keyword{respect}
\keyword{string}
\keyword{the}
\keyword{to}
\keyword{true}
\keyword{valid.}
\keyword{values}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{scientificName}}{character optional}

\item{\code{commonName}}{character optional}

\item{\code{ncbiId}}{integer optional}

\item{\code{isGenesUsable}}{character optional}

\item{\code{secondaryNcbiId}}{integer optional}

\item{\code{id}}{integer optional}

\item{\code{externalDatabase}}{\link{ExternalDatabase} optional}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Taxon-new}{\code{Taxon$new()}}
\item \href{#method-Taxon-toJSON}{\code{Taxon$toJSON()}}
\item \href{#method-Taxon-fromJSON}{\code{Taxon$fromJSON()}}
\item \href{#method-Taxon-toJSONString}{\code{Taxon$toJSONString()}}
\item \href{#method-Taxon-fromJSONString}{\code{Taxon$fromJSONString()}}
\item \href{#method-Taxon-validateJSON}{\code{Taxon$validateJSON()}}
\item \href{#method-Taxon-toString}{\code{Taxon$toString()}}
\item \href{#method-Taxon-isValid}{\code{Taxon$isValid()}}
\item \href{#method-Taxon-getInvalidFields}{\code{Taxon$getInvalidFields()}}
\item \href{#method-Taxon-print}{\code{Taxon$print()}}
\item \href{#method-Taxon-clone}{\code{Taxon$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Taxon-new"></a>}}
\if{latex}{\out{\hypertarget{method-Taxon-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Taxon class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Taxon$new(
  scientificName = NULL,
  commonName = NULL,
  ncbiId = NULL,
  isGenesUsable = NULL,
  secondaryNcbiId = NULL,
  id = NULL,
  externalDatabase = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scientificName}}{scientificName}

\item{\code{commonName}}{commonName}

\item{\code{ncbiId}}{ncbiId}

\item{\code{isGenesUsable}}{isGenesUsable}

\item{\code{secondaryNcbiId}}{secondaryNcbiId}

\item{\code{id}}{id}

\item{\code{externalDatabase}}{externalDatabase}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Taxon-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Taxon-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Taxon$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Taxon in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Taxon-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Taxon-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Taxon
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Taxon$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Taxon
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Taxon-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Taxon-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Taxon$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Taxon in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Taxon-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Taxon-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Taxon
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Taxon$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Taxon
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Taxon-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Taxon-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Taxon and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Taxon$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Taxon-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Taxon-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Taxon$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Taxon
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Taxon-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Taxon-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Taxon$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Taxon-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Taxon-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Taxon$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Taxon-print"></a>}}
\if{latex}{\out{\hypertarget{method-Taxon-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Taxon$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Taxon-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Taxon-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Taxon$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
