% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_prime_distance_method.R
\docType{class}
\name{ThreePrimeDistanceMethod}
\alias{ThreePrimeDistanceMethod}
\title{ThreePrimeDistanceMethod}
\format{
An \code{R6Class} generator object
}
\description{
ThreePrimeDistanceMethod Class

ThreePrimeDistanceMethod Class
}
\details{
Create a new ThreePrimeDistanceMethod
}
\keyword{(JSON}
\keyword{(if}
\keyword{Deserialize}
\keyword{Initialize}
\keyword{JSON}
\keyword{Return}
\keyword{ThreePrimeDistanceMethod}
\keyword{To}
\keyword{Validate}
\keyword{a}
\keyword{all}
\keyword{an}
\keyword{any).}
\keyword{are}
\keyword{class.}
\keyword{fields}
\keyword{format)}
\keyword{if}
\keyword{in}
\keyword{input}
\keyword{instance}
\keyword{internal}
\keyword{into}
\keyword{invalid}
\keyword{list}
\keyword{new}
\keyword{of}
\keyword{respect}
\keyword{string}
\keyword{the}
\keyword{to}
\keyword{true}
\keyword{valid.}
\keyword{values}
\keyword{with}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{character optional}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ThreePrimeDistanceMethod-new}{\code{ThreePrimeDistanceMethod$new()}}
\item \href{#method-ThreePrimeDistanceMethod-toJSON}{\code{ThreePrimeDistanceMethod$toJSON()}}
\item \href{#method-ThreePrimeDistanceMethod-fromJSON}{\code{ThreePrimeDistanceMethod$fromJSON()}}
\item \href{#method-ThreePrimeDistanceMethod-toJSONString}{\code{ThreePrimeDistanceMethod$toJSONString()}}
\item \href{#method-ThreePrimeDistanceMethod-fromJSONString}{\code{ThreePrimeDistanceMethod$fromJSONString()}}
\item \href{#method-ThreePrimeDistanceMethod-validateJSON}{\code{ThreePrimeDistanceMethod$validateJSON()}}
\item \href{#method-ThreePrimeDistanceMethod-toString}{\code{ThreePrimeDistanceMethod$toString()}}
\item \href{#method-ThreePrimeDistanceMethod-isValid}{\code{ThreePrimeDistanceMethod$isValid()}}
\item \href{#method-ThreePrimeDistanceMethod-getInvalidFields}{\code{ThreePrimeDistanceMethod$getInvalidFields()}}
\item \href{#method-ThreePrimeDistanceMethod-print}{\code{ThreePrimeDistanceMethod$print()}}
\item \href{#method-ThreePrimeDistanceMethod-clone}{\code{ThreePrimeDistanceMethod$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThreePrimeDistanceMethod-new"></a>}}
\if{latex}{\out{\hypertarget{method-ThreePrimeDistanceMethod-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ThreePrimeDistanceMethod class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThreePrimeDistanceMethod$new(value = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{value}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThreePrimeDistanceMethod-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ThreePrimeDistanceMethod-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThreePrimeDistanceMethod$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ThreePrimeDistanceMethod in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThreePrimeDistanceMethod-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ThreePrimeDistanceMethod-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ThreePrimeDistanceMethod
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThreePrimeDistanceMethod$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ThreePrimeDistanceMethod
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThreePrimeDistanceMethod-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ThreePrimeDistanceMethod-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThreePrimeDistanceMethod$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ThreePrimeDistanceMethod in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThreePrimeDistanceMethod-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ThreePrimeDistanceMethod-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ThreePrimeDistanceMethod
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThreePrimeDistanceMethod$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ThreePrimeDistanceMethod
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThreePrimeDistanceMethod-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ThreePrimeDistanceMethod-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ThreePrimeDistanceMethod and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThreePrimeDistanceMethod$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThreePrimeDistanceMethod-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ThreePrimeDistanceMethod-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThreePrimeDistanceMethod$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ThreePrimeDistanceMethod
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThreePrimeDistanceMethod-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ThreePrimeDistanceMethod-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThreePrimeDistanceMethod$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThreePrimeDistanceMethod-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ThreePrimeDistanceMethod-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThreePrimeDistanceMethod$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThreePrimeDistanceMethod-print"></a>}}
\if{latex}{\out{\hypertarget{method-ThreePrimeDistanceMethod-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThreePrimeDistanceMethod$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ThreePrimeDistanceMethod-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ThreePrimeDistanceMethod-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ThreePrimeDistanceMethod$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
