% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validateBoolean}
\alias{validateBoolean}
\title{Validate a boolean value}
\usage{
validateBoolean(name, ...)
}
\arguments{
\item{name}{The variable name}

\item{...}{Any boolean types}
}
\value{
The validated boolean as a character string (true or false), or stop with an error message
}
\description{
Validate a boolean value
}
