% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validateConsolidate}
\alias{validateConsolidate}
\title{Validate a consolidate entry (one of pickmax, pickvar, average or missing)}
\usage{
validateConsolidate(name, ...)
}
\arguments{
\item{name}{The variable name}

\item{...}{Any consolidate entries}
}
\value{
The validated consolidate entries, or stop with an error message
}
\description{
Validate a consolidate entry (one of pickmax, pickvar, average or missing)
}
\keyword{internal}
