% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validateFilter}
\alias{validateFilter}
\title{Validate a filter}
\usage{
validateFilter(name, ...)
}
\arguments{
\item{name}{The variable name}

\item{...}{Any filters}
}
\value{
The validated filters, or stop with an error message
}
\description{
Validate a filter
}
\keyword{internal}
