% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validateLimit}
\alias{validateLimit}
\title{Validate a limit value}
\usage{
validateLimit(name, ...)
}
\arguments{
\item{name}{The variable name}

\item{...}{Any possible integers}
}
\value{
The validated integers, or stop with an error message
}
\description{
Validate a limit value
}
\keyword{internal}
