% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_validators.R
\name{validateOptionalID}
\alias{validateOptionalID}
\title{Validate identifiers (ie. gene ID, platform ID, etc.) that are homogeneous (either all numerics or all not)}
\usage{
validateOptionalID(name, ...)
}
\arguments{
\item{name}{The variable name}

\item{...}{Any identifiers}
}
\value{
The validated identifiers, or stop with an error message
}
\description{
Validate identifiers (ie. gene ID, platform ID, etc.) that are homogeneous (either all numerics or all not)
}
\keyword{internal}
