% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validateOptionalQuery}
\alias{validateOptionalQuery}
\title{Validate am optional query}
\usage{
validateOptionalQuery(name, ...)
}
\arguments{
\item{name}{The variable name}

\item{...}{Any queries}
}
\value{
The validated queries
}
\description{
Validate am optional query
}
\keyword{internal}
