% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validateOptionalTaxon}
\alias{validateOptionalTaxon}
\title{Validate a taxon using the acceptable taxa entries}
\usage{
validateOptionalTaxon(name, ...)
}
\arguments{
\item{name}{The variable name}

\item{...}{Any taxa to validate}
}
\value{
The validated taxon, or stop with an error message
}
\description{
Validate a taxon using the acceptable taxa entries
}
\keyword{internal}
