% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validatePositiveInteger}
\alias{validatePositiveInteger}
\title{Validate a non-negative integer value}
\usage{
validatePositiveInteger(name, ...)
}
\arguments{
\item{name}{The variable name}

\item{...}{Any possible integers}
}
\value{
The validated integers, or stop with an error message
}
\description{
Validate a non-negative integer value
}
\keyword{internal}
