% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validatePositiveReal}
\alias{validatePositiveReal}
\title{Validate a non-negative number}
\usage{
validatePositiveReal(name, ...)
}
\arguments{
\item{name}{The variable name}

\item{...}{Any possible numbers}
}
\value{
The validated numbers, or stop with an error message
}
\description{
Validate a non-negative number
}
\keyword{internal}
