% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_validators.R
\name{validateQuery}
\alias{validateQuery}
\title{Validate a query}
\usage{
validateQuery(name, ...)
}
\arguments{
\item{name}{The variable name}

\item{...}{Any queries}
}
\value{
The validated queries, or stop with an error message
}
\description{
Validate a query
}
\keyword{internal}
