% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validateResultType}
\alias{validateResultType}
\title{Validate result types}
\usage{
validateResultType(name, ...)
}
\arguments{
\item{name}{The variable name}

\item{...}{result types}
}
\value{
Validated result types. Either returned as they are or they will be
replaced from human readable variants
}
\description{
Validate result types
}
\keyword{internal}
