% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validateSearchQuery}
\alias{validateSearchQuery}
\title{Validate a search query}
\usage{
validateSearchQuery(name, ...)
}
\arguments{
\item{name}{The variable name}

\item{...}{Any query arguments}
}
\value{
The validated query arguments, or stop with an error message
}
\description{
Validate a search query
}
\keyword{internal}
