% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validateSingleID}
\alias{validateSingleID}
\title{Validate a single identifier(ie. gene ID, platform ID, etc.)}
\usage{
validateSingleID(name, ...)
}
\arguments{
\item{name}{The variable name}

\item{...}{An identifier}
}
\value{
The validated identifier, or stop with an error message
}
\description{
Validate a single identifier(ie. gene ID, platform ID, etc.)
}
