% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validateSort}
\alias{validateSort}
\title{Validate a sort argument}
\usage{
validateSort(name, ...)
}
\arguments{
\item{name}{The variable name}

\item{...}{Any sort arguments}
}
\value{
The validated sort arguments, or stop with an error message
}
\description{
Validate a sort argument
}
\keyword{internal}
