% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validateStrand}
\alias{validateStrand}
\title{Validate a strand (ie. + or -)}
\usage{
validateStrand(name, ...)
}
\arguments{
\item{name}{The variable name}

\item{...}{Any strands}
}
\value{
The validated strands, or stop with an error message
}
\description{
Validate a strand (ie. + or -)
}
