% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validateTaxa}
\alias{validateTaxa}
\title{Validate taxa using the acceptable taxa entries}
\usage{
validateTaxa(name, ...)
}
\arguments{
\item{name}{The variable name}

\item{...}{Any taxa to validate}
}
\value{
The validated taxa, or stop with an error message
}
\description{
Validate taxa using the acceptable taxa entries
}
\keyword{internal}
