% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subProcessors.R
\name{checkBounds}
\alias{checkBounds}
\title{Replace missing data with NAs}
\usage{
checkBounds(x, natype = NA)
}
\arguments{
\item{x}{Data}

\item{natype}{type of NA to replace the missing data with}
}
\value{
Data or NA in case of an out of bounds error
}
\description{
Replace missing data with NAs
}
\keyword{internal}
