% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{.body}
\alias{.body}
\title{Prototype function body}
\usage{
.body(
  memoised,
  fname,
  validators,
  endpoint,
  envWhere,
  isFile,
  hasHeader,
  raw,
  overwrite,
  file,
  async,
  .call
)
}
\arguments{
\item{memoised}{Whether or not to memoise results}

\item{fname}{The function name}

\item{validators}{Validators for parameters}

\item{endpoint}{Formatted endpoint URL}

\item{envWhere}{Environment to evaluate in}

\item{isFile}{Whether or not the endpoint is expect to return a file}

\item{hasHeader}{Indicates if the endpoint uses a HTTP header}

\item{raw}{Whether to return JSON (\code{TRUE}) or data.table (\code{FALSE})}

\item{overwrite}{Whether or not to overwrite the file if @param file is specified}

\item{file}{A filename to save results to}

\item{async}{Whether or not to run asynchronously}

\item{.call}{The original function call}
}
\description{
This should not be called directly, but is called from the API functions.
}
\keyword{internal}
