% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{forgetGemmaMemoised}
\alias{forgetGemmaMemoised}
\title{forgetGemmaMemoised}
\usage{
forgetGemmaMemoised()
}
\description{
Forget past results from memoised calls to the Gemma API (ie. using functions with memoised = \code{TRUE})
}
