% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_allEndpoints.R
\name{forgetGemmaMemoised}
\alias{forgetGemmaMemoised}
\title{Clear gemma.R cache}
\usage{
forgetGemmaMemoised()
}
\value{
TRUE to indicate cache was cleared.
}
\description{
Forget past results from memoised calls to the Gemma API (ie. using functions with memoised = \code{TRUE})
}
\examples{
forgetGemmaMemoised()
}
\keyword{misc}
