% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_open_api_expose.R
\docType{data}
\name{gemma_api}
\alias{gemma_api}
\title{Full API wrapper}
\format{
An object of class \code{list} of length 40.
}
\usage{
gemma_api
}
\description{
Exposes the full Gemma API. Auto generated using
\href{https://openapi-generator.tech/}{OpenAPI generator}
Exposed to user to allow access unsupported endpoints
by the main package.
}
\section{Methods}{

\describe{
\strong{ get_api_info } \emph{ Retrieve an object with basic API information }

\itemize{
\item \emph{ @returnType } \link{ResponseDataObjectApiInfoValueObject} \cr

\item status code : 0 | default response

\item return type : ResponseDataObjectApiInfoValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_dataset_annotations } \emph{ Retrieve the annotations analysis of a dataset }

\itemize{
\item \emph{ @param } dataset \link{GetResultSetsDatasetsParameterInner}
\item \emph{ @returnType } \link{ResponseDataObjectSetAnnotationValueObject} \cr

\item status code : 200 |

\item return type : ResponseDataObjectSetAnnotationValueObject
\item response headers :

\tabular{ll}{
}
\item status code : 404 | The dataset does not exist.

\item return type : ResponseErrorObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_dataset_design } \emph{ Retrieve the design of a dataset }

\itemize{
\item \emph{ @param } dataset \link{GetResultSetsDatasetsParameterInner}

\item status code : 200 |

\item return type : data.frame
\item response headers :

\tabular{ll}{
}
\item status code : 404 | The dataset does not exist.

\item return type : ResponseErrorObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_dataset_differential_expression } \emph{ Retrieve the expression levels of a set of datasets subject to a threshold on their differential expressions }

\itemize{
\item \emph{ @param } datasets list( \link{GetResultSetsDatasetsParameterInner} )
\item \emph{ @param } diff_ex_set integer
\item \emph{ @param } threshold numeric
\item \emph{ @param } limit integer
\item \emph{ @param } keep_non_specific character
\item \emph{ @param } consolidate Enum < pickmax, pickvar, average >
\item \emph{ @returnType } \link{ResponseDataObjectListExperimentExpressionLevelsValueObject} \cr

\item status code : 0 | default response

\item return type : ResponseDataObjectListExperimentExpressionLevelsValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_dataset_differential_expression_analyses } \emph{ Retrieve the differential analyses of a dataset }

\itemize{
\item \emph{ @param } dataset \link{GetResultSetsDatasetsParameterInner}
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @returnType } \link{ResponseDataObjectListDifferentialExpressionAnalysisValueObject} \cr

\item status code : 200 |

\item return type : ResponseDataObjectListDifferentialExpressionAnalysisValueObject
\item response headers :

\tabular{ll}{
}
\item status code : 404 | The dataset does not exist.

\item return type : ResponseErrorObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_dataset_differential_expression_analyses_result_sets } \emph{ Retrieve the result sets of all differential analyses of a dataset }

\itemize{
\item \emph{ @param } dataset \link{GetResultSetsDatasetsParameterInner}

\item status code : 302 | If the dataset is found, a redirection to the corresponding getResultSets operation.

\item response headers :

\tabular{ll}{
}
\item status code : 404 | The dataset does not exist.

\item return type : ResponseErrorObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_dataset_expression } \emph{ Retrieve the expression data of a dataset }

\itemize{
\item \emph{ @param } dataset \link{GetResultSetsDatasetsParameterInner}
\item \emph{ @param } filter character

\item status code : 200 |

\item return type : data.frame
\item response headers :

\tabular{ll}{
}
\item status code : 204 | The dataset expression matrix is empty.

\item response headers :

\tabular{ll}{
}
\item status code : 404 | The dataset does not exist.

\item return type : ResponseErrorObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_dataset_expression_for_genes } \emph{ Retrieve the expression data matrix of a set of datasets and genes }

\itemize{
\item \emph{ @param } datasets list( \link{GetResultSetsDatasetsParameterInner} )
\item \emph{ @param } genes list( \link{GetDatasetExpressionForGenesGenesParameterInner} )
\item \emph{ @param } keep_non_specific character
\item \emph{ @param } consolidate Enum < pickmax, pickvar, average >
\item \emph{ @returnType } \link{ResponseDataObjectListExperimentExpressionLevelsValueObject} \cr

\item status code : 0 | default response

\item return type : ResponseDataObjectListExperimentExpressionLevelsValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_dataset_expression_pca } \emph{ Retrieve the principal components (PCA) of a set of datasets }

\itemize{
\item \emph{ @param } datasets list( \link{GetResultSetsDatasetsParameterInner} )
\item \emph{ @param } component integer
\item \emph{ @param } limit integer
\item \emph{ @param } keep_non_specific character
\item \emph{ @param } consolidate Enum < pickmax, pickvar, average >
\item \emph{ @returnType } \link{ResponseDataObjectListExperimentExpressionLevelsValueObject} \cr

\item status code : 0 | default response

\item return type : ResponseDataObjectListExperimentExpressionLevelsValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_dataset_platforms } \emph{ Retrieve the platform of a dataset }

\itemize{
\item \emph{ @param } dataset \link{GetResultSetsDatasetsParameterInner}
\item \emph{ @returnType } \link{ResponseDataObjectListArrayDesignValueObject} \cr

\item status code : 200 |

\item return type : ResponseDataObjectListArrayDesignValueObject
\item response headers :

\tabular{ll}{
}
\item status code : 404 | The dataset does not exist.

\item return type : ResponseErrorObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_dataset_raw_expression } \emph{ Retrieve raw expression data of a dataset }

\itemize{
\item \emph{ @param } dataset \link{GetResultSetsDatasetsParameterInner}

\item status code : 200 |

\item return type : data.frame
\item response headers :

\tabular{ll}{
}
\item status code : 404 | The dataset does not exist.

\item return type : ResponseErrorObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_dataset_samples } \emph{ Retrieve the samples of a dataset }

\itemize{
\item \emph{ @param } dataset \link{GetResultSetsDatasetsParameterInner}
\item \emph{ @returnType } \link{ResponseDataObjectListBioAssayValueObject} \cr

\item status code : 200 |

\item return type : ResponseDataObjectListBioAssayValueObject
\item response headers :

\tabular{ll}{
}
\item status code : 404 | The dataset does not exist.

\item return type : ResponseErrorObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_dataset_svd } \emph{ Retrieve the singular value decomposition (SVD) of a dataset expression data }

\itemize{
\item \emph{ @param } dataset \link{GetResultSetsDatasetsParameterInner}
\item \emph{ @returnType } \link{ResponseDataObjectSimpleSVDValueObject} \cr

\item status code : 200 |

\item return type : ResponseDataObjectSimpleSVDValueObject
\item response headers :

\tabular{ll}{
}
\item status code : 404 | The dataset does not exist.

\item return type : ResponseErrorObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_datasets } \emph{ Retrieve all datasets }

\itemize{
\item \emph{ @param } filter character
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @param } sort character
\item \emph{ @returnType } \link{PaginatedResponseDataObjectExpressionExperimentValueObject} \cr

\item status code : 0 | default response

\item return type : PaginatedResponseDataObjectExpressionExperimentValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_datasets_by_ids } \emph{ Retrieve datasets by their identifiers }

\itemize{
\item \emph{ @param } dataset list( \link{GetResultSetsDatasetsParameterInner} )
\item \emph{ @param } filter character
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @param } sort character
\item \emph{ @returnType } \link{PaginatedResponseDataObjectExpressionExperimentValueObject} \cr

\item status code : 0 | default response

\item return type : PaginatedResponseDataObjectExpressionExperimentValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_gene_gene_coexpression } \emph{ Retrieve the coexpression of two given genes }

\itemize{
\item \emph{ @param } gene \link{GetDatasetExpressionForGenesGenesParameterInner}
\item \emph{ @param } with \link{GetDatasetExpressionForGenesGenesParameterInner}
\item \emph{ @param } limit integer
\item \emph{ @param } stringency integer
\item \emph{ @returnType } \link{ResponseDataObjectListCoexpressionValueObjectExt} \cr

\item status code : 0 | default response

\item return type : ResponseDataObjectListCoexpressionValueObjectExt
\item response headers :

\tabular{ll}{
}
}

\strong{ get_gene_go_terms } \emph{ Retrieve the GO terms associated to a gene }

\itemize{
\item \emph{ @param } gene \link{GetDatasetExpressionForGenesGenesParameterInner}
\item \emph{ @returnType } \link{ResponseDataObjectListGeneOntologyTermValueObject} \cr

\item status code : 0 | default response

\item return type : ResponseDataObjectListGeneOntologyTermValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_gene_locations } \emph{ Retrieve the physical locations of a given gene }

\itemize{
\item \emph{ @param } gene \link{GetDatasetExpressionForGenesGenesParameterInner}
\item \emph{ @returnType } \link{ResponseDataObjectListPhysicalLocationValueObject} \cr

\item status code : 0 | default response

\item return type : ResponseDataObjectListPhysicalLocationValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_gene_locations_in_taxon } \emph{ Retrieve physical locations for a given gene and taxon }

\itemize{
\item \emph{ @param } taxon \link{SearchTaxonDatasetsTaxonParameter}
\item \emph{ @param } gene \link{GetDatasetExpressionForGenesGenesParameterInner}
\item \emph{ @returnType } \link{ResponseDataObjectListPhysicalLocationValueObject} \cr

\item status code : 0 | default response

\item return type : ResponseDataObjectListPhysicalLocationValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_gene_probes } \emph{ Retrieve the probes associated to a genes }

\itemize{
\item \emph{ @param } gene \link{GetDatasetExpressionForGenesGenesParameterInner}
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @returnType } \link{PaginatedResponseDataObjectCompositeSequenceValueObject} \cr

\item status code : 0 | default response

\item return type : PaginatedResponseDataObjectCompositeSequenceValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_genes } \emph{ Retrieve genes matching a gene identifier }

\itemize{
\item \emph{ @param } genes list( \link{GetDatasetExpressionForGenesGenesParameterInner} )
\item \emph{ @returnType } \link{ResponseDataObjectListGeneValueObject} \cr

\item status code : 0 | default response

\item return type : ResponseDataObjectListGeneValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_platform_annotations } \emph{ Retrieve the annotations of a given platform }

\itemize{
\item \emph{ @param } platform \link{GetPlatformAnnotationsPlatformParameter}

\item status code : 200 |

\item return type : data.frame
\item response headers :

\tabular{ll}{
}
}

\strong{ get_platform_datasets } \emph{ Retrieve all experiments within a given platform }

\itemize{
\item \emph{ @param } platform \link{GetPlatformAnnotationsPlatformParameter}
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @returnType } \link{PaginatedResponseDataObjectExpressionExperimentValueObject} \cr

\item status code : 0 | default response

\item return type : PaginatedResponseDataObjectExpressionExperimentValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_platform_element } \emph{ Retrieve the selected composite sequences for a given platform }

\itemize{
\item \emph{ @param } platform \link{GetPlatformAnnotationsPlatformParameter}
\item \emph{ @param } probes list( \link{GetPlatformElementProbesParameterInner} )
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @returnType } \link{PaginatedResponseDataObjectCompositeSequenceValueObject} \cr

\item status code : 0 | default response

\item return type : PaginatedResponseDataObjectCompositeSequenceValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_platform_element_genes } \emph{ Retrieve the genes associated to a probe in a given platform }

\itemize{
\item \emph{ @param } platform \link{GetPlatformAnnotationsPlatformParameter}
\item \emph{ @param } probe \link{GetPlatformElementProbesParameterInner}
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @returnType } \link{PaginatedResponseDataObjectGeneValueObject} \cr

\item status code : 0 | default response

\item return type : PaginatedResponseDataObjectGeneValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_platform_elements } \emph{ Retrieve the composite sequences for a given platform }

\itemize{
\item \emph{ @param } platform \link{GetPlatformAnnotationsPlatformParameter}
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @returnType } \link{PaginatedResponseDataObjectCompositeSequenceValueObject} \cr

\item status code : 0 | default response

\item return type : PaginatedResponseDataObjectCompositeSequenceValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_platforms } \emph{ Retrieve all platforms }

\itemize{
\item \emph{ @param } filter character
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @param } sort character
\item \emph{ @returnType } \link{PaginatedResponseDataObjectArrayDesignValueObject} \cr

\item status code : 0 | default response

\item return type : PaginatedResponseDataObjectArrayDesignValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_platforms_by_ids } \emph{ Retrieve all platforms matching a set of platform identifiers }

\itemize{
\item \emph{ @param } platform list( \link{GetPlatformAnnotationsPlatformParameter} )
\item \emph{ @param } filter character
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @param } sort character
\item \emph{ @returnType } \link{PaginatedResponseDataObjectArrayDesignValueObject} \cr

\item status code : 0 | default response

\item return type : PaginatedResponseDataObjectArrayDesignValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_result_set } \emph{ Retrieve a single analysis result set by its identifier }

\itemize{
\item \emph{ @param } result_set integer
\item \emph{ @returnType } \link{ResponseDataObjectExpressionAnalysisResultSetValueObject} \cr

\item status code : 200 |

\item return type : ResponseDataObjectExpressionAnalysisResultSetValueObject
\item response headers :

\tabular{ll}{
}
\item status code : 404 | The analysis result set could not be found.

\item return type : ResponseErrorObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_result_set_as_tsv } \emph{ Retrieve a single analysis result set by its identifier as a tab-separated values }

\itemize{
\item \emph{ @param } result_set_ integer

\item status code : 200 |

\item return type : data.frame
\item response headers :

\tabular{ll}{
}
\item status code : 404 | The analysis result set could not be found.

\item response headers :

\tabular{ll}{
}
}

\strong{ get_result_sets } \emph{ Retrieve all result sets matching the provided criteria }

\itemize{
\item \emph{ @param } datasets list( \link{GetResultSetsDatasetsParameterInner} )
\item \emph{ @param } database_entries list( \link{GetResultSetsDatabaseEntriesParameterInner} )
\item \emph{ @param } filter character
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @param } sort character
\item \emph{ @returnType } \link{PaginatedResponseDataObjectExpressionAnalysisResultSetValueObject} \cr

\item status code : 0 | default response

\item return type : PaginatedResponseDataObjectExpressionAnalysisResultSetValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_taxa } \emph{ Retrieve all available taxa }

\itemize{
\item \emph{ @returnType } \link{ResponseDataObjectListTaxonValueObject} \cr

\item status code : 0 | default response

\item return type : ResponseDataObjectListTaxonValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_taxa_by_ids } \emph{ Retrieve taxa by their identifiers }

\itemize{
\item \emph{ @param } taxa list( \link{SearchTaxonDatasetsTaxonParameter} )
\item \emph{ @returnType } \link{ResponseDataObjectListTaxonValueObject} \cr

\item status code : 0 | default response

\item return type : ResponseDataObjectListTaxonValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_taxon_datasets } \emph{ Retrieve the datasets for a given taxon }

\itemize{
\item \emph{ @param } taxon \link{SearchTaxonDatasetsTaxonParameter}
\item \emph{ @param } filter character
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @param } sort character
\item \emph{ @returnType } \link{PaginatedResponseDataObjectExpressionExperimentValueObject} \cr

\item status code : 0 | default response

\item return type : PaginatedResponseDataObjectExpressionExperimentValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_taxon_genes } \emph{ Retrieve all genes in a given taxon }

\itemize{
\item \emph{ @param } taxon \link{SearchTaxonDatasetsTaxonParameter}
\item \emph{ @param } gene \link{GetDatasetExpressionForGenesGenesParameterInner}
\item \emph{ @returnType } \link{ResponseDataObjectListGeneValueObject} \cr

\item status code : 0 | default response

\item return type : ResponseDataObjectListGeneValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_taxon_genes_overlapping_chromosome } \emph{ Retrieve genes overlapping a given region in a taxon }

\itemize{
\item \emph{ @param } taxon \link{SearchTaxonDatasetsTaxonParameter}
\item \emph{ @param } chromosome character
\item \emph{ @param } strand character
\item \emph{ @param } start integer
\item \emph{ @param } size integer
\item \emph{ @returnType } \link{ResponseDataObjectListGeneValueObject} \cr

\item status code : 0 | default response

\item return type : ResponseDataObjectListGeneValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ search } \emph{ Search everything in Gemma. }

\itemize{
\item \emph{ @param } query character
\item \emph{ @param } taxon \link{SearchTaxonDatasetsTaxonParameter}
\item \emph{ @param } platform \link{GetPlatformAnnotationsPlatformParameter}
\item \emph{ @param } result_types Enum < ubic.gemma.model.common.description.BibliographicReference, ubic.gemma.model.analysis.expression.ExpressionExperimentSet, ubic.gemma.model.genome.gene.GeneSet, ubic.gemma.model.genome.Gene, ubic.gemma.model.expression.experiment.ExpressionExperiment, ubic.gemma.model.genome.biosequence.BioSequence, ubic.gemma.model.association.phenotype.PhenotypeAssociation, ubic.gemma.model.expression.arrayDesign.ArrayDesign, ubic.gemma.model.expression.designElement.CompositeSequence >
\item \emph{ @param } limit integer
\item \emph{ @returnType } \link{SearchResultsResponseDataObject} \cr

\item status code : 0 | default response

\item return type : SearchResultsResponseDataObject
\item response headers :

\tabular{ll}{
}
}

\strong{ search_annotations } \emph{ Search for annotation tags }

\itemize{
\item \emph{ @param } query list( character )
\item \emph{ @returnType } \link{ResponseDataObjectListAnnotationSearchResultValueObject} \cr

\item status code : 200 |

\item return type : ResponseDataObjectListAnnotationSearchResultValueObject
\item response headers :

\tabular{ll}{
}
\item status code : 400 | The search query is empty.

\item return type : ResponseErrorObject
\item response headers :

\tabular{ll}{
}
}

\strong{ search_datasets } \emph{ Retrieve datasets associated to an annotation tags search }

\itemize{
\item \emph{ @param } query list( character )
\item \emph{ @param } filter character
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @param } sort character
\item \emph{ @returnType } \link{PaginatedResponseDataObjectExpressionExperimentValueObject} \cr

\item status code : 0 | default response

\item return type : PaginatedResponseDataObjectExpressionExperimentValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ search_taxon_datasets } \emph{ Retrieve datasets within a given taxa associated to an annotation tags search }

\itemize{
\item \emph{ @param } taxon \link{SearchTaxonDatasetsTaxonParameter}
\item \emph{ @param } query list( character )
\item \emph{ @param } filter character
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @param } sort character
\item \emph{ @returnType } \link{PaginatedResponseDataObjectExpressionExperimentValueObject} \cr

\item status code : 0 | default response

\item return type : PaginatedResponseDataObjectExpressionExperimentValueObject
\item response headers :

\tabular{ll}{
}
}

}


\describe{
\strong{ get_api_info } \emph{ Retrieve an object with basic API information }

\itemize{
\item \emph{ @returnType } \link{ResponseDataObjectApiInfoValueObject} \cr

\item status code : 0 | default response

\item return type : ResponseDataObjectApiInfoValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_dataset_annotations } \emph{ Retrieve the annotations analysis of a dataset }

\itemize{
\item \emph{ @param } dataset \link{GetResultSetsDatasetsParameterInner}
\item \emph{ @returnType } \link{ResponseDataObjectSetAnnotationValueObject} \cr

\item status code : 200 |

\item return type : ResponseDataObjectSetAnnotationValueObject
\item response headers :

\tabular{ll}{
}
\item status code : 404 | The dataset does not exist.

\item return type : ResponseErrorObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_dataset_design } \emph{ Retrieve the design of a dataset }

\itemize{
\item \emph{ @param } dataset \link{GetResultSetsDatasetsParameterInner}

\item status code : 200 |

\item return type : data.frame
\item response headers :

\tabular{ll}{
}
\item status code : 404 | The dataset does not exist.

\item return type : ResponseErrorObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_dataset_differential_expression } \emph{ Retrieve the expression levels of a set of datasets subject to a threshold on their differential expressions }

\itemize{
\item \emph{ @param } datasets list( \link{GetResultSetsDatasetsParameterInner} )
\item \emph{ @param } diff_ex_set integer
\item \emph{ @param } threshold numeric
\item \emph{ @param } limit integer
\item \emph{ @param } keep_non_specific character
\item \emph{ @param } consolidate Enum < pickmax, pickvar, average >
\item \emph{ @returnType } \link{ResponseDataObjectListExperimentExpressionLevelsValueObject} \cr

\item status code : 0 | default response

\item return type : ResponseDataObjectListExperimentExpressionLevelsValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_dataset_differential_expression_analyses } \emph{ Retrieve the differential analyses of a dataset }

\itemize{
\item \emph{ @param } dataset \link{GetResultSetsDatasetsParameterInner}
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @returnType } \link{ResponseDataObjectListDifferentialExpressionAnalysisValueObject} \cr

\item status code : 200 |

\item return type : ResponseDataObjectListDifferentialExpressionAnalysisValueObject
\item response headers :

\tabular{ll}{
}
\item status code : 404 | The dataset does not exist.

\item return type : ResponseErrorObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_dataset_differential_expression_analyses_result_sets } \emph{ Retrieve the result sets of all differential analyses of a dataset }

\itemize{
\item \emph{ @param } dataset \link{GetResultSetsDatasetsParameterInner}

\item status code : 302 | If the dataset is found, a redirection to the corresponding getResultSets operation.

\item response headers :

\tabular{ll}{
}
\item status code : 404 | The dataset does not exist.

\item return type : ResponseErrorObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_dataset_expression } \emph{ Retrieve the expression data of a dataset }

\itemize{
\item \emph{ @param } dataset \link{GetResultSetsDatasetsParameterInner}
\item \emph{ @param } filter character

\item status code : 200 |

\item return type : data.frame
\item response headers :

\tabular{ll}{
}
\item status code : 204 | The dataset expression matrix is empty.

\item response headers :

\tabular{ll}{
}
\item status code : 404 | The dataset does not exist.

\item return type : ResponseErrorObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_dataset_expression_for_genes } \emph{ Retrieve the expression data matrix of a set of datasets and genes }

\itemize{
\item \emph{ @param } datasets list( \link{GetResultSetsDatasetsParameterInner} )
\item \emph{ @param } genes list( \link{GetDatasetExpressionForGenesGenesParameterInner} )
\item \emph{ @param } keep_non_specific character
\item \emph{ @param } consolidate Enum < pickmax, pickvar, average >
\item \emph{ @returnType } \link{ResponseDataObjectListExperimentExpressionLevelsValueObject} \cr

\item status code : 0 | default response

\item return type : ResponseDataObjectListExperimentExpressionLevelsValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_dataset_expression_pca } \emph{ Retrieve the principal components (PCA) of a set of datasets }

\itemize{
\item \emph{ @param } datasets list( \link{GetResultSetsDatasetsParameterInner} )
\item \emph{ @param } component integer
\item \emph{ @param } limit integer
\item \emph{ @param } keep_non_specific character
\item \emph{ @param } consolidate Enum < pickmax, pickvar, average >
\item \emph{ @returnType } \link{ResponseDataObjectListExperimentExpressionLevelsValueObject} \cr

\item status code : 0 | default response

\item return type : ResponseDataObjectListExperimentExpressionLevelsValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_dataset_platforms } \emph{ Retrieve the platform of a dataset }

\itemize{
\item \emph{ @param } dataset \link{GetResultSetsDatasetsParameterInner}
\item \emph{ @returnType } \link{ResponseDataObjectListArrayDesignValueObject} \cr

\item status code : 200 |

\item return type : ResponseDataObjectListArrayDesignValueObject
\item response headers :

\tabular{ll}{
}
\item status code : 404 | The dataset does not exist.

\item return type : ResponseErrorObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_dataset_raw_expression } \emph{ Retrieve raw expression data of a dataset }

\itemize{
\item \emph{ @param } dataset \link{GetResultSetsDatasetsParameterInner}

\item status code : 200 |

\item return type : data.frame
\item response headers :

\tabular{ll}{
}
\item status code : 404 | The dataset does not exist.

\item return type : ResponseErrorObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_dataset_samples } \emph{ Retrieve the samples of a dataset }

\itemize{
\item \emph{ @param } dataset \link{GetResultSetsDatasetsParameterInner}
\item \emph{ @returnType } \link{ResponseDataObjectListBioAssayValueObject} \cr

\item status code : 200 |

\item return type : ResponseDataObjectListBioAssayValueObject
\item response headers :

\tabular{ll}{
}
\item status code : 404 | The dataset does not exist.

\item return type : ResponseErrorObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_dataset_svd } \emph{ Retrieve the singular value decomposition (SVD) of a dataset expression data }

\itemize{
\item \emph{ @param } dataset \link{GetResultSetsDatasetsParameterInner}
\item \emph{ @returnType } \link{ResponseDataObjectSimpleSVDValueObject} \cr

\item status code : 200 |

\item return type : ResponseDataObjectSimpleSVDValueObject
\item response headers :

\tabular{ll}{
}
\item status code : 404 | The dataset does not exist.

\item return type : ResponseErrorObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_datasets } \emph{ Retrieve all datasets }

\itemize{
\item \emph{ @param } filter character
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @param } sort character
\item \emph{ @returnType } \link{PaginatedResponseDataObjectExpressionExperimentValueObject} \cr

\item status code : 0 | default response

\item return type : PaginatedResponseDataObjectExpressionExperimentValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_datasets_by_ids } \emph{ Retrieve datasets by their identifiers }

\itemize{
\item \emph{ @param } dataset list( \link{GetResultSetsDatasetsParameterInner} )
\item \emph{ @param } filter character
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @param } sort character
\item \emph{ @returnType } \link{PaginatedResponseDataObjectExpressionExperimentValueObject} \cr

\item status code : 0 | default response

\item return type : PaginatedResponseDataObjectExpressionExperimentValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_gene_gene_coexpression } \emph{ Retrieve the coexpression of two given genes }

\itemize{
\item \emph{ @param } gene \link{GetDatasetExpressionForGenesGenesParameterInner}
\item \emph{ @param } with \link{GetDatasetExpressionForGenesGenesParameterInner}
\item \emph{ @param } limit integer
\item \emph{ @param } stringency integer
\item \emph{ @returnType } \link{ResponseDataObjectListCoexpressionValueObjectExt} \cr

\item status code : 0 | default response

\item return type : ResponseDataObjectListCoexpressionValueObjectExt
\item response headers :

\tabular{ll}{
}
}

\strong{ get_gene_go_terms } \emph{ Retrieve the GO terms associated to a gene }

\itemize{
\item \emph{ @param } gene \link{GetDatasetExpressionForGenesGenesParameterInner}
\item \emph{ @returnType } \link{ResponseDataObjectListGeneOntologyTermValueObject} \cr

\item status code : 0 | default response

\item return type : ResponseDataObjectListGeneOntologyTermValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_gene_locations } \emph{ Retrieve the physical locations of a given gene }

\itemize{
\item \emph{ @param } gene \link{GetDatasetExpressionForGenesGenesParameterInner}
\item \emph{ @returnType } \link{ResponseDataObjectListPhysicalLocationValueObject} \cr

\item status code : 0 | default response

\item return type : ResponseDataObjectListPhysicalLocationValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_gene_locations_in_taxon } \emph{ Retrieve physical locations for a given gene and taxon }

\itemize{
\item \emph{ @param } taxon \link{SearchTaxonDatasetsTaxonParameter}
\item \emph{ @param } gene \link{GetDatasetExpressionForGenesGenesParameterInner}
\item \emph{ @returnType } \link{ResponseDataObjectListPhysicalLocationValueObject} \cr

\item status code : 0 | default response

\item return type : ResponseDataObjectListPhysicalLocationValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_gene_probes } \emph{ Retrieve the probes associated to a genes }

\itemize{
\item \emph{ @param } gene \link{GetDatasetExpressionForGenesGenesParameterInner}
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @returnType } \link{PaginatedResponseDataObjectCompositeSequenceValueObject} \cr

\item status code : 0 | default response

\item return type : PaginatedResponseDataObjectCompositeSequenceValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_genes } \emph{ Retrieve genes matching a gene identifier }

\itemize{
\item \emph{ @param } genes list( \link{GetDatasetExpressionForGenesGenesParameterInner} )
\item \emph{ @returnType } \link{ResponseDataObjectListGeneValueObject} \cr

\item status code : 0 | default response

\item return type : ResponseDataObjectListGeneValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_platform_annotations } \emph{ Retrieve the annotations of a given platform }

\itemize{
\item \emph{ @param } platform \link{GetPlatformAnnotationsPlatformParameter}

\item status code : 200 |

\item return type : data.frame
\item response headers :

\tabular{ll}{
}
}

\strong{ get_platform_datasets } \emph{ Retrieve all experiments within a given platform }

\itemize{
\item \emph{ @param } platform \link{GetPlatformAnnotationsPlatformParameter}
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @returnType } \link{PaginatedResponseDataObjectExpressionExperimentValueObject} \cr

\item status code : 0 | default response

\item return type : PaginatedResponseDataObjectExpressionExperimentValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_platform_element } \emph{ Retrieve the selected composite sequences for a given platform }

\itemize{
\item \emph{ @param } platform \link{GetPlatformAnnotationsPlatformParameter}
\item \emph{ @param } probes list( \link{GetPlatformElementProbesParameterInner} )
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @returnType } \link{PaginatedResponseDataObjectCompositeSequenceValueObject} \cr

\item status code : 0 | default response

\item return type : PaginatedResponseDataObjectCompositeSequenceValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_platform_element_genes } \emph{ Retrieve the genes associated to a probe in a given platform }

\itemize{
\item \emph{ @param } platform \link{GetPlatformAnnotationsPlatformParameter}
\item \emph{ @param } probe \link{GetPlatformElementProbesParameterInner}
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @returnType } \link{PaginatedResponseDataObjectGeneValueObject} \cr

\item status code : 0 | default response

\item return type : PaginatedResponseDataObjectGeneValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_platform_elements } \emph{ Retrieve the composite sequences for a given platform }

\itemize{
\item \emph{ @param } platform \link{GetPlatformAnnotationsPlatformParameter}
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @returnType } \link{PaginatedResponseDataObjectCompositeSequenceValueObject} \cr

\item status code : 0 | default response

\item return type : PaginatedResponseDataObjectCompositeSequenceValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_platforms } \emph{ Retrieve all platforms }

\itemize{
\item \emph{ @param } filter character
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @param } sort character
\item \emph{ @returnType } \link{PaginatedResponseDataObjectArrayDesignValueObject} \cr

\item status code : 0 | default response

\item return type : PaginatedResponseDataObjectArrayDesignValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_platforms_by_ids } \emph{ Retrieve all platforms matching a set of platform identifiers }

\itemize{
\item \emph{ @param } platform list( \link{GetPlatformAnnotationsPlatformParameter} )
\item \emph{ @param } filter character
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @param } sort character
\item \emph{ @returnType } \link{PaginatedResponseDataObjectArrayDesignValueObject} \cr

\item status code : 0 | default response

\item return type : PaginatedResponseDataObjectArrayDesignValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_result_set } \emph{ Retrieve a single analysis result set by its identifier }

\itemize{
\item \emph{ @param } result_set integer
\item \emph{ @returnType } \link{ResponseDataObjectExpressionAnalysisResultSetValueObject} \cr

\item status code : 200 |

\item return type : ResponseDataObjectExpressionAnalysisResultSetValueObject
\item response headers :

\tabular{ll}{
}
\item status code : 404 | The analysis result set could not be found.

\item return type : ResponseErrorObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_result_set_as_tsv } \emph{ Retrieve a single analysis result set by its identifier as a tab-separated values }

\itemize{
\item \emph{ @param } result_set_ integer

\item status code : 200 |

\item return type : data.frame
\item response headers :

\tabular{ll}{
}
\item status code : 404 | The analysis result set could not be found.

\item response headers :

\tabular{ll}{
}
}

\strong{ get_result_sets } \emph{ Retrieve all result sets matching the provided criteria }

\itemize{
\item \emph{ @param } datasets list( \link{GetResultSetsDatasetsParameterInner} )
\item \emph{ @param } database_entries list( \link{GetResultSetsDatabaseEntriesParameterInner} )
\item \emph{ @param } filter character
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @param } sort character
\item \emph{ @returnType } \link{PaginatedResponseDataObjectExpressionAnalysisResultSetValueObject} \cr

\item status code : 0 | default response

\item return type : PaginatedResponseDataObjectExpressionAnalysisResultSetValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_taxa } \emph{ Retrieve all available taxa }

\itemize{
\item \emph{ @returnType } \link{ResponseDataObjectListTaxonValueObject} \cr

\item status code : 0 | default response

\item return type : ResponseDataObjectListTaxonValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_taxa_by_ids } \emph{ Retrieve taxa by their identifiers }

\itemize{
\item \emph{ @param } taxa list( \link{SearchTaxonDatasetsTaxonParameter} )
\item \emph{ @returnType } \link{ResponseDataObjectListTaxonValueObject} \cr

\item status code : 0 | default response

\item return type : ResponseDataObjectListTaxonValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_taxon_datasets } \emph{ Retrieve the datasets for a given taxon }

\itemize{
\item \emph{ @param } taxon \link{SearchTaxonDatasetsTaxonParameter}
\item \emph{ @param } filter character
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @param } sort character
\item \emph{ @returnType } \link{PaginatedResponseDataObjectExpressionExperimentValueObject} \cr

\item status code : 0 | default response

\item return type : PaginatedResponseDataObjectExpressionExperimentValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_taxon_genes } \emph{ Retrieve all genes in a given taxon }

\itemize{
\item \emph{ @param } taxon \link{SearchTaxonDatasetsTaxonParameter}
\item \emph{ @param } gene \link{GetDatasetExpressionForGenesGenesParameterInner}
\item \emph{ @returnType } \link{ResponseDataObjectListGeneValueObject} \cr

\item status code : 0 | default response

\item return type : ResponseDataObjectListGeneValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ get_taxon_genes_overlapping_chromosome } \emph{ Retrieve genes overlapping a given region in a taxon }

\itemize{
\item \emph{ @param } taxon \link{SearchTaxonDatasetsTaxonParameter}
\item \emph{ @param } chromosome character
\item \emph{ @param } strand character
\item \emph{ @param } start integer
\item \emph{ @param } size integer
\item \emph{ @returnType } \link{ResponseDataObjectListGeneValueObject} \cr

\item status code : 0 | default response

\item return type : ResponseDataObjectListGeneValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ search } \emph{ Search everything in Gemma. }

\itemize{
\item \emph{ @param } query character
\item \emph{ @param } taxon \link{SearchTaxonDatasetsTaxonParameter}
\item \emph{ @param } platform \link{GetPlatformAnnotationsPlatformParameter}
\item \emph{ @param } result_types Enum < ubic.gemma.model.common.description.BibliographicReference, ubic.gemma.model.analysis.expression.ExpressionExperimentSet, ubic.gemma.model.genome.gene.GeneSet, ubic.gemma.model.genome.Gene, ubic.gemma.model.expression.experiment.ExpressionExperiment, ubic.gemma.model.genome.biosequence.BioSequence, ubic.gemma.model.association.phenotype.PhenotypeAssociation, ubic.gemma.model.expression.arrayDesign.ArrayDesign, ubic.gemma.model.expression.designElement.CompositeSequence >
\item \emph{ @param } limit integer
\item \emph{ @returnType } \link{SearchResultsResponseDataObject} \cr

\item status code : 0 | default response

\item return type : SearchResultsResponseDataObject
\item response headers :

\tabular{ll}{
}
}

\strong{ search_annotations } \emph{ Search for annotation tags }

\itemize{
\item \emph{ @param } query list( character )
\item \emph{ @returnType } \link{ResponseDataObjectListAnnotationSearchResultValueObject} \cr

\item status code : 200 |

\item return type : ResponseDataObjectListAnnotationSearchResultValueObject
\item response headers :

\tabular{ll}{
}
\item status code : 400 | The search query is empty.

\item return type : ResponseErrorObject
\item response headers :

\tabular{ll}{
}
}

\strong{ search_datasets } \emph{ Retrieve datasets associated to an annotation tags search }

\itemize{
\item \emph{ @param } query list( character )
\item \emph{ @param } filter character
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @param } sort character
\item \emph{ @returnType } \link{PaginatedResponseDataObjectExpressionExperimentValueObject} \cr

\item status code : 0 | default response

\item return type : PaginatedResponseDataObjectExpressionExperimentValueObject
\item response headers :

\tabular{ll}{
}
}

\strong{ search_taxon_datasets } \emph{ Retrieve datasets within a given taxa associated to an annotation tags search }

\itemize{
\item \emph{ @param } taxon \link{SearchTaxonDatasetsTaxonParameter}
\item \emph{ @param } query list( character )
\item \emph{ @param } filter character
\item \emph{ @param } offset integer
\item \emph{ @param } limit integer
\item \emph{ @param } sort character
\item \emph{ @returnType } \link{PaginatedResponseDataObjectExpressionExperimentValueObject} \cr

\item status code : 0 | default response

\item return type : PaginatedResponseDataObjectExpressionExperimentValueObject
\item response headers :

\tabular{ll}{
}
}

}
}

\seealso{
\code{\link{gemmaCall}}
}
\keyword{misc}
