% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_caching.R
\name{gemma_memoise}
\alias{gemma_memoise}
\title{Enable and disable memoisation of gemma.R functions}
\usage{
gemma_memoise(
  memoised = FALSE,
  cache = rappdirs::user_cache_dir(appname = "gemmaR")
)
}
\arguments{
\item{memoised}{boolean. If TRUE memoisation will be enabled}

\item{cache}{File path or "cache_in_memory". File path will chose a location
to save the cache files for memoisation. "cache_in_memory" will store the cache
in the current R session}
}
\description{
Enable and disable memoisation of gemma.R functions
}
\keyword{misc}
