% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{geneInfo}
\alias{geneInfo}
\title{geneInfo}
\usage{
geneInfo(
  gene = NA_character_,
  request = NA_character_,
  ...,
  raw = FALSE,
  async = FALSE,
  memoised = FALSE,
  file = NA_character_,
  overwrite = FALSE
)
}
\description{
geneInfo
}
