% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{get_all_pages}
\alias{get_all_pages}
\title{Get all pages of a paginated call}
\usage{
get_all_pages(query, step_size = 100, binder = rbind)
}
\arguments{
\item{query}{Output from a gemma.R function with offset and query argumend}

\item{step_size}{Size of individual calls to the server. 100 is the maximum value}

\item{binder}{Binding function for the calls. If \code{raw = FALSE} use \code{rbind} to
combine the data.tables. If not, use \code{c} to combine lists}
}
\value{
A data.table or a list containing data from all pages.
}
\description{
Given a Gemma.R output with offset and limit arguments,
returns the entire output.
}
