% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{get_child_terms}
\alias{get_child_terms}
\title{Return child terms of a term}
\usage{
get_child_terms(terms)
}
\value{
An array containing descendends of the annotation terms, including
the terms themselves
}
\description{
When querying for ontology terms, Gemma propagates these terms to include
any datasets with their child terms in the results. This function returns
these children for any number of terms, including all children
and the terms itself in the output vector
}
\examples{
get_child_terms("http://purl.obolibrary.org/obo/MONDO_0000408")

}
\keyword{misc}
