% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{get_dataset_differential_expression_analyses}
\alias{get_dataset_differential_expression_analyses}
\title{Retrieve the differential analyses of a dataset}
\usage{
get_dataset_differential_expression_analyses(
  dataset,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE),
  attributes = getOption("gemma.attributes", TRUE)
)
}
\arguments{
\item{dataset}{A numerical dataset identifier or a dataset short name}

\item{raw}{\code{TRUE} to receive results as-is from Gemma, or \code{FALSE} to enable
parsing. Raw results usually contain additional fields and flags that are
omitted in the parsed results.}

\item{memoised}{Whether or not to save to cache for future calls with the
same inputs and use the result saved in cache if a result is already saved.
Doing \code{options(gemma.memoised = TRUE)} will ensure that the cache is always
used. Use \code{\link{forget_gemma_memoised}} to clear the cache.}

\item{file}{The name of a file to save the results to, or \code{NULL} to not write
results to a file. If \code{raw == TRUE}, the output will be a JSON file. Otherwise,
it will be a RDS file.}

\item{overwrite}{Whether or not to overwrite if a file exists at the specified
filename.}

\item{attributes}{If \code{TRUE} additional information from the call will be added
into the output object's attributes such as offset and available elements.}
}
\value{
A data table with information about the differential expression
analysis of the queried dataset. Note that this funciton does not return
differential expression values themselves. Use \code{\link{get_differential_expression_values}}
to get differential expression values (see examples).

The fields of the output data.table are:

\itemize{
\item \code{result.ID}: Result set ID of the differential expression analysis.
May represent multiple factors in a single model.
\item \code{contrast.ID}: Id of the specific contrast factor. Together with the result.ID
they uniquely represent a given contrast.
\item \code{experiment.ID}: Id of the source experiment
\item \code{baseline.category}: Category for the contrast
\item \code{baseline.categoryURI}: URI for the baseline category
\item \code{baseline.factorValue}: Factor value assigned as the baseline in the contrast. Typically represent control samples
\item \code{baseline.factorValueURI}: URI for the baseline.factorValue
\item \code{experimental.factorValue}: Factor value assigned to the experimental group.
\item \code{experimental.factorValueURI}: URI for the experimental.factorValue
\item \code{subsetFactor.subset}:
\item \code{subsetFactor.category}:
\item \code{subsetFactor.categoryURI}:
\item \code{subsetFactor.factorValue}:
\item \code{subsetFactor.factorValueURI}:
\item \code{stats.DE}: Number of differentially expressed genes for the contrast
\item \code{stats.Down}: Number of downregulated genes for the contrast
\item \code{stats.Up}: Number of upregulated genes for the contrast
\item \code{analysis.Threshold}: P value threshold used to determine the number
of differentially expressed genes
\item \code{probes.Analyzed}: Number of probesets represented in the contrast
\item \code{genes.Analyzed}: Number of genes represented in the contrast
\item \code{platform.ID}: Platform id for the contrast
}
}
\description{
Retrieve the differential analyses of a dataset
}
\examples{
result <- get_dataset_differential_expression_analyses("GSE2018")
get_differential_expression_values(resultSet = result$result.ID)
}
\keyword{dataset}
