% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{get_datasets_by_ids}
\alias{get_datasets_by_ids}
\title{Get datasets by ids}
\usage{
get_datasets_by_ids(
  datasets = NA_character_,
  offset = 0L,
  limit = 20L,
  sort = "+id",
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE),
  attributes = getOption("gemma.attributes", TRUE)
)
}
\arguments{
\item{datasets}{Numerical dataset identifiers or dataset short names}

\item{offset}{The offset of the first retrieved result.}

\item{limit}{Optional, defaults to 20. Limits the result to specified amount
of objects. Has a maximum value of 100. Use together with \code{offset} and
the \code{totalElements} \link[base:attributes]{attribute} in the output to
compile all data if needed.}

\item{sort}{Order results by the given property and direction. The '+' sign indicate ascending order whereas the '-' indicate descending.}

\item{raw}{\code{TRUE} to receive results as-is from Gemma, or \code{FALSE} to enable
parsing. Raw results usually contain additional fields and flags that are
omitted in the parsed results.}

\item{memoised}{Whether or not to save to cache for future calls with the
same inputs and use the result saved in cache if a result is already saved.
Doing \code{options(gemma.memoised = TRUE)} will ensure that the cache is always
used. Use \code{\link{forget_gemma_memoised}} to clear the cache.}

\item{file}{The name of a file to save the results to, or \code{NULL} to not write
results to a file. If \code{raw == TRUE}, the output will be a JSON file. Otherwise,
it will be a RDS file.}

\item{overwrite}{Whether or not to overwrite if a file exists at the specified
filename.}

\item{attributes}{If \code{TRUE} additional information from the call will be added
into the output object's attributes such as offset and available elements.}
}
\value{
A data table with information about the queried dataset(s). A list if
\code{raw = TRUE}. Returns an empty list if no datasets matched. A successful
response may contain 'Geeq' information, which aims to provide a unified
metric to measure experiments by the quality of their data, and their
suitability for use in Gemma. You can
read more about the geeq properties \href{https://pavlidislab.github.io/Gemma/geeq.html}{here}.

The fields of the output data.table are:

\itemize{
\item \code{ee.ShortName}: Shortname given to the dataset within Gemma. Often corresponds to accession ID
\item \code{ee.Name}: Full title of the dataset
\item \code{ee.ID}: Internal ID of the dataset.
\item \code{ee.Description}: Description of the dataset
\item \code{ee.Public}: Is the dataset publicly available.
\item \code{ee.Troubled}: Did an automatic process within gemma or a curator mark the dataset as "troubled"
\item \code{ee.Accession}: Accession ID of the dataset in the external database it was taken from
\item \code{ee.Database}: The name of the database where the dataset was taken from
\item \code{ee.URL}: URL of the original database
\item \code{ee.SampleCount}: Number of samples in the dataset
\item \code{ee.batchEffect}: A text field describing whether the dataset has batch effects
\item \code{ee.batchCorrected}: Whether batch correction has been performed on the dataset
\item \code{geeq.qScore}: Data quality score given to the dataset by Gemma.
\item \code{geeq.sScore}: Suitability score given to the dataset by Gemma. Refers to factors like batches, platforms and other aspects of experimental design
\item \code{taxon.Name}: The taxa of the study. In Gemma each study will include a single species. If the original source has samples from multiple species, they will be split into different studies within Gemma
\item \code{taxon.ID}: Internal ID given to the taxon by Gemma
}
}
\description{
Retrieve datasets by their identifiers
}
\examples{
get_datasets_by_ids("GSE2018")
get_datasets_by_ids(c("GSE2018", "GSE2872"))
}
\keyword{dataset}
