% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{get_genes}
\alias{get_genes}
\title{Retrieve genes matching a gene identifier}
\usage{
get_genes(
  genes,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE),
  attributes = getOption("gemma.attributes", TRUE)
)
}
\arguments{
\item{genes}{An ensembl gene identifier which typically starts with ensg or an ncbi gene identifier or an official gene symbol approved by hgnc}

\item{raw}{\code{TRUE} to receive results as-is from Gemma, or \code{FALSE} to enable
parsing. Raw results usually contain additional fields and flags that are
omitted in the parsed results.}

\item{memoised}{Whether or not to save to cache for future calls with the
same inputs and use the result saved in cache if a result is already saved.
Doing \code{options(gemma.memoised = TRUE)} will ensure that the cache is always
used. Use \code{\link{forget_gemma_memoised}} to clear the cache.}

\item{file}{The name of a file to save the results to, or \code{NULL} to not write
results to a file. If \code{raw == TRUE}, the output will be a JSON file. Otherwise,
it will be a RDS file.}

\item{overwrite}{Whether or not to overwrite if a file exists at the specified
filename.}

\item{attributes}{If \code{TRUE} additional information from the call will be added
into the output object's attributes such as offset and available elements.}
}
\value{
A data table with information about the querried gene(s)
A list if \code{raw = TRUE}.

The fields of the output data.table are:

\itemize{
\item \code{gene.Symbol}: Symbol for the gene
\item \code{gene.Ensembl}: Ensembl ID for the gene
\item \code{gene.NCBI}: NCBI id for the gene
\item \code{gene.Name}: Name of the gene
\item \code{taxon.Name}: Name of the species
\item \code{taxon.Scientific}: Scientific name for the taxon
\item \code{taxon.ID}: Internal identifier given to the species by Gemma
\item \code{taxon.NCBI}: NCBI ID of the taxon
\item \code{taxon.Database.Name}: Underlying database used in Gemma for the taxon
\item \code{taxon.Database.ID}: ID of the underyling database used in Gemma for the taxon
}
}
\description{
Retrieve genes matching a gene identifier
}
\examples{
get_genes("DYRK1A")
get_genes(c("DYRK1A", "PTEN"))
}
\keyword{gene}
