% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{get_result_sets}
\alias{get_result_sets}
\title{Retrieve all result sets matching the provided criteria}
\usage{
get_result_sets(
  datasets,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)
}
\arguments{
\item{datasets}{A numerical dataset identifier or a dataset short name}

\item{raw}{\code{TRUE} to receive results as-is from Gemma, or \code{FALSE} to enable
parsing. Raw results usually contain additional fields and flags that are
omitted in the parsed results.}

\item{memoised}{Whether or not to save to cache for future calls with the
same inputs and use the result saved in cache if a result is already saved.
Doing \code{options(gemma.memoised = TRUE)} will ensure that the cache is always
used. Use \code{\link{forget_gemma_memoised}} to clear the cache.}

\item{file}{The name of a file to save the results to, or \code{NULL} to not write
results to a file. If \code{raw == TRUE}, the output will be the raw endpoint from the
API, likely a JSON or a gzip file. Otherwise, it will be a RDS file.}

\item{overwrite}{Whether or not to overwrite if a file exists at the specified
filename.}
}
\value{
A data table with the queried datasets' resultSet ID(s). A list if
\code{raw = TRUE}. Use
\code{\link{get_differential_expression_values}} to get differential expression
values (see examples). Use \code{\link{get_dataset_differential_expression_analyses}}
to get more detailed information about a result set.

The fields of the output data.table are:

\itemize{
\item \code{resultSet.id}: Internal ID given to the result set. Can be used to access the results using \code{\link{get_differential_expression_values}}
\item \code{factor.category}: What is the category splitting the experimental groups in the result set (e.g. disease )
\item \code{factor.levels}: What are the conditions that are compared in the result set (e.g control, bipolar disorder)

}
}
\description{
Retrieve all result sets matching the provided criteria
}
\examples{
resultSets <- get_result_sets("GSE2872")
get_differential_expression_values(resultSet = resultSets$resultSet.id[1])
}
\keyword{internal}
