% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{get_taxa}
\alias{get_taxa}
\title{Get taxa}
\usage{
get_taxa(memoised = getOption("gemma.memoised", FALSE))
}
\arguments{
\item{memoised}{Whether or not to save to cache for future calls with the
same inputs and use the result saved in cache if a result is already saved.
Doing \code{options(gemma.memoised = TRUE)} will ensure that the cache is always
used. Use \code{\link{forget_gemma_memoised}} to clear the cache.}
}
\value{
A data frame including the names, IDs and database information
about the taxons
}
\description{
Returns taxa and their versions used in Gemma
}
\examples{
get_taxa()
}
\keyword{misc}
