% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{getDataset}
\alias{getDataset}
\title{Access gene expression data and metadata}
\usage{
getDataset(
  dataset,
  filter = FALSE,
  type = "se",
  memoised = getOption("gemma.memoised", FALSE)
)
}
\arguments{
\item{dataset}{A dataset identifier.}

\item{filter}{The filtered version corresponds to what is used in most Gemma analyses, removing some probes/elements. Unfiltered includes all elements.}

\item{type}{"se"for a SummarizedExperiment or "eset" for Expression Set. We recommend using
SummarizedExperiments which are more recent. See the Summarized experiment
\href{https://bioconductor.org/packages/release/bioc/vignettes/SummarizedExperiment/inst/doc/SummarizedExperiment.html}{vignette}
or the ExpressionSet \href{https://bioconductor.org/packages/release/bioc/vignettes/Biobase/inst/doc/ExpressionSetIntroduction.pdf}{vignette}
for more details.}

\item{memoised}{Whether or not to save to cache for future calls with the same inputs
and use the result saved in cache if a result is already saved. Doing
\code{options(gemma.memoised = TRUE)} will ensure that the cache is always used.
Use \code{forgetGemmaMemoised} to clear the cache.
cache. Use \code{forgetGemmaMemoised} to clear the cache.}
}
\value{
A SummarizedExperiment or ExpressionSet of the queried dataset.
}
\description{
Return an annotated Bioconductor-compatible
data structure of the queried dataset, including expression data and
the experimental design.
}
\examples{
getDataset("GSE2018")
}
\keyword{dataset}
