% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{getDataset}
\alias{getDataset}
\title{Dataset expression and design}
\usage{
getDataset(
  dataset,
  filter = FALSE,
  type = "se",
  memoised = getOption("gemma.memoise", FALSE)
)
}
\arguments{
\item{dataset}{A dataset identifier.}

\item{filter}{The filtered version corresponds to what is used in most Gemma analyses, removing some probes/elements. Unfiltered includes all elements.}

\item{type}{"se"for a SummarizedExperiment or "eset" for Expression Set. We recommend using
SummarizedExperiments which are more recent. See the Summarized experiment}

\item{memoised}{Whether or not to cache results so future requests for the same data
will be faster. Use \code{forgetGemmaMemoised} to clear the cache.
\href{https://bioconductor.org/packages/release/bioc/vignettes/SummarizedExperiment/inst/doc/SummarizedExperiment.html}{vignette}
or the ExpressionSet \href{https://bioconductor.org/packages/release/bioc/vignettes/Biobase/inst/doc/ExpressionSetIntroduction.pdf}{vignette}
for more details.}
}
\value{
A SummarizedExperiment or ExpressionSet of the queried dataset.
}
\description{
Combines various endpoint calls to return an annotated Bioconductor-compatible
data structure of the queried dataset, including expression data and
the experimental design.
}
\examples{
\donttest{
getDataset("GSE2018")
}
}
\keyword{dataset}
