% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{getDatasetDE}
\alias{getDatasetDE}
\title{Get Differential Expression SummarizedExperiment}
\usage{
getDatasetDE(dataset)
}
\arguments{
\item{dataset}{A dataset identifier.}
}
\value{
A \code{\link[SummarizedExperiment]{SummarizedExperiment}} or \code{\link[Biobase]{ExpressionSet}} of the queried resultSet.
}
\description{
Combine the SummarizedExperiment
}
\examples{
getDatasetDE("GSE2018")
}
\keyword{dataset}
