% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{getDatasetDE}
\alias{getDatasetDE}
\title{Get Differential Expression}
\usage{
getDatasetDE(dataset)
}
\arguments{
\item{dataset}{A dataset identifier.}
}
\value{
A data table with differential expression values
}
\description{
Retrieves the differential expression resultSet(s) associated with the dataset.
If there is more than one resultSet, it will show the options and prompt the user
to choose one.
}
\examples{
getDatasetDE("GSE2018")
}
\keyword{dataset}
