% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{getDatasetDE}
\alias{getDatasetDE}
\title{Dataset differential expression}
\usage{
getDatasetDE(
  dataset = NA_character_,
  resultSet = NA_integer_,
  memoised = getOption("gemma.memoise", FALSE)
)
}
\arguments{
\item{dataset}{A dataset identifier.}

\item{resultSet}{A resultSet identifier.}

\item{memoised}{Whether or not to save to cache for future calls with the same inputs
and use the result saved in cache if a result is already saved. Doing
\code{options(gemma.memoised = TRUE)} will ensure that the catche is always used.
Use \code{forgetGemmaMemoised} to clear the cache.}
}
\value{
A list of data tables with differential expression
values per result set.
}
\description{
Retrieves the differential expression resultSet(s) associated with the dataset.
If there is more than one resultSet, use \code{\link[=getDatasetResultSets]{getDatasetResultSets()}} or \code{\link[=getDatasetDEA]{getDatasetDEA()}} to see
the options and get the ID you want. Alternatively, you can query the resultSet
directly if you know its ID beforehand.
}
\examples{
getDatasetDE("GSE2018")
}
\keyword{dataset}
