% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{getDatasetDE}
\alias{getDatasetDE}
\title{Dataset differential expression}
\usage{
getDatasetDE(
  dataset = NA_character_,
  resultSet = NA_integer_,
  all = FALSE,
  memoised = getOption("gemma.memoise", FALSE)
)
}
\arguments{
\item{dataset}{A dataset identifier.}

\item{resultSet}{A resultSet identifier.}

\item{all}{If TRUE, will download all differential expression resultSets for the dataset.}

\item{memoised}{Whether or not to cache results so future requests for the same data
will be faster. Use \code{forgetGemmaMemoised} to clear the cache.}
}
\value{
A data table with differential expression values. If there are multiple
resultSets and all = TRUE, a list of data tables with differential expression
values.
}
\description{
Retrieves the differential expression resultSet(s) associated with the dataset.
If there is more than one resultSet, use \code{\link[=getDatasetResultSets]{getDatasetResultSets()}} to see
the options and get the ID you want. Alternatively, you can query the resultSet
directly if you know its ID beforehand.
}
\examples{
getDatasetDE("GSE2018")
}
\keyword{dataset}
