% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{getDatasetDE}
\alias{getDatasetDE}
\title{Datasets differential expression levels
Retrieves differential expression levels for given datasets}
\usage{
getDatasetDE(
  datasets = NA_character_,
  keepNonSpecific = FALSE,
  diffExSet = NA_integer_,
  threshold = 100,
  limit = 100L,
  consolidate = NA_character_,
  raw = getOption("gemma.raw", F),
  async = getOption("gemma.async", F),
  memoised = getOption("gemma.memoise", F),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", F)
)
}
\arguments{
\item{datasets}{Optional, defaults to \code{empty}.
Limits the result to entities with given identifiers.
A list of identifiers, separated by commas (e.g:
\verb{GSE2871,GSE2869,GSE2868}). Identifiers can either be the Dataset ID or
its short name. Retrieval by ID is more efficient.
Only datasets that user has access to will be available.
Do not combine different identifiers in one query.}

\item{keepNonSpecific}{Optional, defaults to \code{false}.
If set to \code{false}, the response will only include elements that map
exclusively to each queried gene
If set to \code{true}, the response will include all elements that map to
each queried gene, even if they also map to other genes.}

\item{diffExSet}{Required, defaults to \code{empty}.
The ID of the differential expression set to retrieve the data from.
This value can be obtained through the \'Dataset differential analysis\'
endpoint in the \'Dataset endpoints\' category. See the \code{resultSetId} in
one of the response objects in said endpoint.}

\item{threshold}{Optional, defaults to \code{100.0}.
The threshold that the differential expression has to meet to be
included in the response.}

\item{limit}{Optional, defaults to \code{20}.
Limits the result to specified amount of objects. Use 0 for no limit.}

\item{consolidate}{Optional, defaults to \code{empty}.
What action to take when there is more than one element per gene in a
dataset.
The choices are:
\itemize{
\item \verb{[empty]} - will list all vectors separately
\item \code{pickmax} - only return the vector that has the highest expression
(mean over all its bioAssays)
\item \code{pickvar} - only return the vector with highest variance of
expression across its bioAssays
\item \code{average} - create a new vector that will average the bioAssay
values from all vectors
}}

\item{raw}{\code{FALSE} to receive results as-is from Gemma, or \code{TRUE} to enable
parsing.}

\item{async}{\code{TRUE} to run the API query on a separate worker, or \code{FALSE} to run
synchronously. See the \code{async} package for details.}

\item{memoised}{Whether or not to cache results so future requests for the same data
will be faster. Use \code{forgetGemmaMemoised} to clear the cache.}

\item{file}{The name of a file to save the results to, or \code{NULL} to not write
results to a file. If \code{raw == TRUE}, the output will be a JSON file.
Otherwise, it will be a RDS file.}

\item{overwrite}{Whether or not to overwrite if a file exists at the specified filename.}
}
\value{
The differential expression levels for each given experiment (experiment
expression levels value object) in the given differential expression
set.
If the experiment is not in the given diff. exp. set, an empty array is
returned.
A \verb{404 error} if the given identifier does not map to any object.
}
\description{
Datasets differential expression levels
Retrieves differential expression levels for given datasets
}
\keyword{dataset}
