% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{getDatasetPCA}
\alias{getDatasetPCA}
\title{Datasets pca component expression levels}
\usage{
getDatasetPCA(
  datasets = NA_character_,
  component = 1L,
  limit = 100L,
  keepNonSpecific = FALSE,
  consolidate = NA_character_,
  raw = getOption("gemma.raw", FALSE),
  async = getOption("gemma.async", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)
}
\arguments{
\item{datasets}{Optional, defaults to \code{empty}.
Limits the result to entities with given identifiers.
A list of identifiers, separated by commas (e.g:
\verb{GSE2871,GSE2869,GSE2868}). Identifiers can either be the Dataset ID or
its short name. Retrieval by ID is more efficient.
Only datasets that user has access to will be available.
Do not combine different identifiers in one query.}

\item{component}{Required, defaults to \code{empty}.
The pca component to limit the results to.}

\item{limit}{Optional, defaults to \code{20}.
Limits the result to specified amount of objects. Use 0 for no limit.}

\item{keepNonSpecific}{Optional, defaults to \code{false}.
If set to \code{false}, the response will only include elements that map
exclusively to each queried gene
If set to \code{true}, the response will include all elements that map to
each queried gene, even if they also map to other genes.}

\item{consolidate}{Optional, defaults to \code{empty}.
What action to take when there is more than one element per gene in a
dataset.
The choices are:
\itemize{
\item \verb{[empty]} - will list all vectors separately
\item \code{pickmax} - only return the vector that has the highest expression
(mean over all its bioAssays)
\item \code{pickvar} - only return the vector with highest variance of
expression across its bioAssays
\item \code{average} - create a new vector that will average the bioAssay
values from all vectors
}}

\item{raw}{\code{TRUE} to receive results as-is from Gemma, or \code{FALSE} to enable
parsing.}

\item{async}{\code{TRUE} to run the API query on a separate worker, or \code{FALSE} to run
synchronously. See the \code{async} package for details.}

\item{memoised}{Whether or not to cache results so future requests for the same data
will be faster. Use \code{forgetGemmaMemoised} to clear the cache.}

\item{file}{The name of a file to save the results to, or \code{NULL} to not write
results to a file. If \code{raw == TRUE}, the output will be a JSON file.
Otherwise, it will be a RDS file.}

\item{overwrite}{Whether or not to overwrite if a file exists at the specified filename.}
}
\value{
The expression levels for each given experiment (experiment expression
levels value object) of genes that are most correlated with the given
principal component.
A \verb{404 error} if the given identifier does not map to any object.
}
\description{
Retrieves the expression levels most correlated with the given principal
component
}
\examples{
dat <- getDatasetPCA("GSE2018")
str(dat$expr)
}
\keyword{dataset}
