% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{getDatasetTidy}
\alias{getDatasetTidy}
\title{Tidy expression and design}
\usage{
getDatasetTidy(
  dataset,
  filter = FALSE,
  memoised = getOption("gemma.memoise", FALSE)
)
}
\arguments{
\item{dataset}{A dataset identifier.}

\item{filter}{The filtered version corresponds to what is used in most Gemma analyses, removing some probes/elements. Unfiltered includes all elements.}

\item{memoised}{Whether or not to save to cache for future calls with the same inputs
and use the result saved in cache if a result is already saved. Doing
\code{options(gemma.memoised = TRUE)} will ensure that the cache is always used.
Use \code{forgetGemmaMemoised} to clear the cache.}
}
\value{
A tibble that combines the expression and design matrices.
}
\description{
Combines the expression and design matrices of the queried dataset into a
tibble for easy visualization and exploration with ggplot and the rest of the tidyverse.
}
\examples{

getDatasetTidy("GSE2018")
}
\keyword{dataset}
