% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{getExpressionSet}
\alias{getExpressionSet}
\title{Get ExpressionSet}
\usage{
getExpressionSet(dataset, filter)
}
\arguments{
\item{dataset}{A dataset identifier}

\item{filter}{If true, call returns filtered expression data.}
}
\value{
An annotated ExpressionSet for the queried dataset.
}
\description{
Combines various endpoint calls to return an annotated Bioconductor ExpressionSet
}
\examples{
getExpressionSet("GSE2018")
}
\keyword{dataset}
