% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{getGeneCoexpression}
\alias{getGeneCoexpression}
\title{Gene coexpression
Retrieves the coexpression of two given genes}
\usage{
getGeneCoexpression(
  gene = NA_character_,
  with = NA_character_,
  limit = 20L,
  stringency = 1L,
  raw = getOption("gemma.raw", F),
  async = getOption("gemma.async", F),
  memoised = getOption("gemma.memoise", F),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", F)
)
}
\arguments{
\item{gene}{Required, part of the URL path.
Can either be the NCBI ID (\code{1859}), Ensembl ID (\code{ENSG00000157540}) or
official symbol (\code{DYRK1A}) of the gene.
NCBI ID is the most efficient (and guaranteed to be unique) identifier.
[]{.glyphicon .glyphicon-th-large .glyphicon-exclamation-sign} Official
symbol represents a gene homologue for a random taxon, unless used in a
specific taxon (see Taxon Endpoints).}

\item{with}{Required, defaults to \code{empty}.
Can either be the NCBI ID (\code{1859}), Ensembl ID (\code{ENSG00000157540}) or
official symbol (\code{DYRK1A}) of the gene.
NCBI ID is the most efficient (and guaranteed to be unique) identifier.
[]{.glyphicon .glyphicon-th-large .glyphicon-exclamation-sign} Official
symbol represents a gene homologue for a random taxon, unless used in a
specific taxon (see Taxon Endpoints).}

\item{limit}{Optional, defaults to \code{20}.
Limits the result to specified amount of objects. Use 0 for no limit.}

\item{stringency}{Optional, defaults to \code{1}.
Sets the stringency of coexpression search.}

\item{raw}{\code{FALSE} to receive results as-is from Gemma, or \code{TRUE} to enable
parsing.}

\item{async}{\code{TRUE} to run the API query on a separate worker, or \code{FALSE} to run
synchronously. See the \code{async} package for details.}

\item{memoised}{Whether or not to cache results so future requests for the same data
will be faster. Use \code{forgetGemmaMemoised} to clear the cache.}

\item{file}{The name of a file to save the results to, or \code{NULL} to not write
results to a file. If \code{raw == TRUE}, the output will be a JSON file.
Otherwise, it will be a RDS file.}

\item{overwrite}{Whether or not to overwrite if a file exists at the specified filename.}
}
\value{
An array of gene coexpression data (coexpression meta value objects)
representing the coexpression of the two given genes.
A \verb{404 error} if the given identifier does not map to any object.
A \verb{400 error} if required parameters are missing.
}
\description{
Gene coexpression
Retrieves the coexpression of two given genes
}
\keyword{gene}
