% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{getGeneLocation}
\alias{getGeneLocation}
\title{getGeneLocation}
\usage{
getGeneLocation(
  gene = NA_character_,
  raw = FALSE,
  async = FALSE,
  memoised = FALSE,
  file = NA_character_,
  overwrite = FALSE
)
}
\description{
getGeneLocation
}
