% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{getGeneProbes}
\alias{getGeneProbes}
\title{Gene probes}
\usage{
getGeneProbes(
  gene,
  offset = 0L,
  limit = 20L,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)
}
\arguments{
\item{gene}{Required, part of the URL path.
Can either be the NCBI ID (\code{1859}), Ensembl ID (\code{ENSG00000157540}) or
official symbol (\code{DYRK1A}) of the gene.
NCBI ID is the most efficient (and guaranteed to be unique) identifier.
Official
symbol represents a gene homologue for a random taxon, unless used in a
specific taxon (see Taxon Endpoints).}

\item{offset}{Optional, defaults to \code{0}.
Skips the specified amount of objects when retrieving them from the
database.}

\item{limit}{Optional, defaults to 20. Limits the result to specified amount of
objects.}

\item{raw}{\code{TRUE} to receive results as-is from Gemma, or \code{FALSE} to enable
parsing. Raw results usually contain additional fields and flags that
are omitted in the parsed results.}

\item{memoised}{Whether or not to cache results so future requests for the same data
will be faster. Use \code{forgetGemmaMemoised} to clear the cache.}

\item{file}{The name of a file to save the results to, or \code{NULL} to not write
results to a file. If \code{raw == TRUE}, the output will be a JSON file.
Otherwise, it will be a RDS file.}

\item{overwrite}{Whether or not to overwrite if a file exists at the specified filename.}
}
\value{
An array of probes (composite sequence value objects) that are mapped to
this gene. Note, that it is possible for probes to map to multiple
genes.
A \verb{404 error} if the given identifier does not map to any object.
}
\description{
Retrieves the probes (composite sequences) with this gene
}
\examples{
dat <- getGeneProbes("DYRK1A")
str(dat, vec.len = 2)
}
\keyword{gene}
