% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{getGenes}
\alias{getGenes}
\title{Genes
Retrieves all genes matching the identifiers}
\usage{
getGenes(
  genes = NA_character_,
  raw = getOption("gemma.raw", F),
  async = getOption("gemma.async", F),
  memoised = getOption("gemma.memoise", F),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", F)
)
}
\arguments{
\item{genes}{Required, part of the URL path.
A list of identifiers, separated by commas (e.g: \verb{1859, 5728}).
Can either be the NCBI ID (\code{1859}), Ensembl ID (\code{ENSG00000157540}) or
official symbol (\code{DYRK1A}) of the gene.
NCBI ID is the most efficient (and guaranteed to be unique) identifier.
[]{.glyphicon .glyphicon-th-large .glyphicon-exclamation-sign} Official
symbol represents a gene homologue for a random taxon, unless used in a
specific taxon (see Taxon Endpoints).
Do not combine different identifiers in one query.}

\item{raw}{\code{FALSE} to receive results as-is from Gemma, or \code{TRUE} to enable
parsing.}

\item{async}{\code{TRUE} to run the API query on a separate worker, or \code{FALSE} to run
synchronously. See the \code{async} package for details.}

\item{memoised}{Whether or not to cache results so future requests for the same data
will be faster. Use \code{forgetGemmaMemoised} to clear the cache.}

\item{file}{The name of a file to save the results to, or \code{NULL} to not write
results to a file. If \code{raw == TRUE}, the output will be a JSON file.
Otherwise, it will be a RDS file.}

\item{overwrite}{Whether or not to overwrite if a file exists at the specified filename.}
}
\value{
An array of value objects representing the objects that matched the
query.
Empty array if no objects matched.
A \verb{400 error} if required parameters are missing.
}
\description{
Genes
Retrieves all genes matching the identifiers
}
