% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{getGenesAtLocation}
\alias{getGenesAtLocation}
\title{Genes at location
Finds genes overlapping a given region.}
\usage{
getGenesAtLocation(
  taxon = NA_character_,
  chromosome = NA_character_,
  strand = "+",
  start = NA_integer_,
  size = NA_integer_,
  raw = getOption("gemma.raw", F),
  async = getOption("gemma.async", F),
  memoised = getOption("gemma.memoise", F),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", F)
)
}
\arguments{
\item{taxon}{Not required, part of the URL path.
can either be Taxon ID, Taxon NCBI ID, or one of its string identifiers:
scientific name, common name
It is recommended to use Taxon ID for efficiency.
Please note, that not all taxa have all the possible identifiers
available.
Use the \'All Taxa\' endpoint to retrieve the necessary information. For
convenience, below is a list of officially supported taxa:
ID   Comm.name   Scient.name                NcbiID

1    human       Homo sapiens               9606
2    mouse       Mus musculus               10090
3    rat         Rattus norvegicus          10116
11   yeast       Saccharomyces cerevisiae   4932
12   zebrafish   Danio rerio                7955
13   fly         Drosophila melanogaster    7227
14   worm        Caenorhabditis elegans     6239}

\item{chromosome}{Required, defaults to \code{empty}.
The chromosome of the query location. Eg: \code{3}, \code{21}, \code{X}}

\item{strand}{Optional, defaults to \code{+}.
Can either be \code{+} or \code{-}.
[]{.glyphicon .glyphicon-th-large .glyphicon-exclamation-sign} This is a
WIP parameter and does currently not do anything
[]{.glyphicon .glyphicon-th-large .glyphicon-exclamation-sign} When
using in scripts, remember to URL-encode the \'+\' plus character (see
the compiled URL below).}

\item{start}{Required, defaults to \code{empty}.
Number of the start nucleotide of the desired region.}

\item{size}{Required, defaults to \code{empty}.
Amount of nucleotides in the desired region (i.e. the length of the
region).}

\item{raw}{\code{FALSE} to receive results as-is from Gemma, or \code{TRUE} to enable
parsing.}

\item{async}{\code{TRUE} to run the API query on a separate worker, or \code{FALSE} to run
synchronously. See the \code{async} package for details.}

\item{memoised}{Whether or not to cache results so future requests for the same data
will be faster. Use \code{forgetGemmaMemoised} to clear the cache.}

\item{file}{The name of a file to save the results to, or \code{NULL} to not write
results to a file. If \code{raw == TRUE}, the output will be a JSON file.
Otherwise, it will be a RDS file.}

\item{overwrite}{Whether or not to overwrite if a file exists at the specified filename.}
}
\value{
An array of genes (gene value objects) that overlap the given region.
A \verb{404 error} if the given identifier does not map to any object.
A \verb{400 error} if required parameters are missing.
}
\description{
Genes at location
Finds genes overlapping a given region.
}
