% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{getGenesInfo}
\alias{getGenesInfo}
\title{Genes}
\usage{
getGenesInfo(
  genes,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE),
  attributes = getOption("gemma.attributes", TRUE)
)
}
\arguments{
\item{genes}{Required, part of the URL path.
A list of identifiers, separated by commas (e.g: \verb{1859, 5728}).
Can either be the NCBI ID (\code{1859}), Ensembl ID (\code{ENSG00000157540}) or
official symbol (\code{DYRK1A}) of the gene.
NCBI ID is the most efficient (and guaranteed to be unique) identifier.
Official
symbol represents a gene homologue for a random taxon, unless used in a
specific taxon (see Taxon Endpoints).
Do not combine different identifiers in one query.}

\item{raw}{\code{TRUE} to receive results as-is from Gemma, or \code{FALSE} to enable
parsing. Raw results usually contain additional fields and flags that are
omitted in the parsed results.}

\item{memoised}{Whether or not to save to cache for future calls with the
same inputs and use the result saved in cache if a result is already saved.
Doing \code{options(gemma.memoised = TRUE)} will ensure that the cache is always
used. Use \\code{\\link{forgetGemmaMemoised}} to clear the cache.}

\item{file}{The name of a file to save the results to, or \code{NULL} to not write
results to a file. If \code{raw == TRUE}, the output will be a JSON file. Otherwise,
it will be a RDS file.}

\item{overwrite}{Whether or not to overwrite if a file exists at the specified
filename.}

\item{attributes}{If \code{TRUE} additional information from the call will be added
into the output object's attributes such as offset and available elements.}
}
\value{
A data table with information about the queried gene(s).
}
\description{
Retrieves all genes matching the identifiers
}
\examples{
getGenesInfo("DYRK1A")
getGenesInfo(c("DYRK1A", "PTEN"))
}
\keyword{gene}
