% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{getPlatformDatasets}
\alias{getPlatformDatasets}
\title{getPlatformDatasets}
\usage{
getPlatformDatasets(
  platform = NA_character_,
  offset = 0L,
  limit = 20L,
  raw = FALSE,
  async = FALSE,
  memoised = FALSE
)
}
\description{
getPlatformDatasets
}
