% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{getPlatformElementGenes}
\alias{getPlatformElementGenes}
\title{getPlatformElementGenes}
\usage{
getPlatformElementGenes(
  platform = NA_character_,
  element = NA_character_,
  offset = 0L,
  limit = 20L,
  raw = FALSE,
  async = FALSE,
  memoised = FALSE,
  file = NA_character_,
  overwrite = FALSE
)
}
\description{
getPlatformElementGenes
}
