% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{getPlatformElements}
\alias{getPlatformElements}
\title{Platform elements}
\usage{
getPlatformElements(
  platform = NA_character_,
  element = NA_character_,
  offset = 0L,
  limit = 20L,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)
}
\arguments{
\item{platform}{Required, part of the URL path.
Can either be the platform ID or its short name (e.g: \code{GPL1355})
Retrieval by ID is more efficient.
Only platforms that user has access to will be available.}

\item{element}{Optional, defaults to \code{empty}.
Limits the result to entities with given identifiers.
A list of identifiers, separated by commas (e.g:
\verb{AFFX_Rat_beta-actin_M_at, AFFX_Rat_Hexokinase_M_at}).
Can either be probes name or IDs.
Do not combine different identifiers in one query.
When
using in scripts, remember to URL-encode any forward slashes in the
probe name (see the compiled URL below).}

\item{offset}{Optional, defaults to \code{0}.
Skips the specified amount of objects when retrieving them from the
database.}

\item{limit}{Optional, defaults to \code{20}.
Limits the result to specified amount of objects. Use 0 for no limit.}

\item{raw}{\code{TRUE} to receive results as-is from Gemma, or \code{FALSE} to enable
parsing.}

\item{memoised}{Whether or not to cache results so future requests for the same data
will be faster. Use \code{forgetGemmaMemoised} to clear the cache.}

\item{file}{The name of a file to save the results to, or \code{NULL} to not write
results to a file. If \code{raw == TRUE}, the output will be a JSON file.
Otherwise, it will be a RDS file.}

\item{overwrite}{Whether or not to overwrite if a file exists at the specified filename.}
}
\value{
An array of elements (composite sequence value objects) of the given
platform.
Empty collection, if no elements matched the \code{elements} parameter.
A \verb{404 error} if the given identifier does not map to any object.
}
\description{
Retrieves the composite sequences (elements) for the given platform
}
\examples{
dat <- getPlatformElements("GPL1355")
str(dat, vec.len = 1, max.level = 1)
}
\keyword{platform}
