% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{getTaxa}
\alias{getTaxa}
\title{Taxa}
\usage{
getTaxa(
  taxa = NA_character_,
  raw = getOption("gemma.raw", F),
  async = getOption("gemma.async", F),
  memoised = getOption("gemma.memoise", F),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", F)
)
}
\arguments{
\item{taxa}{Optional, defaults to \code{empty}.
Limits the result to entities with given identifiers.
A list of identifiers, separated by commas (e.g: \verb{human, mouse, fly}).
Identifiers can be the any of the following:
\itemize{
\item taxon ID
\item scientific name
\item common name
Retrieval by ID is more efficient.
Do not combine different identifiers in one query.
For convenience, below is a list of officially supported taxa
ID   Comm.name   Scient.name                NcbiID
}

1    human       Homo sapiens               9606
2    mouse       Mus musculus               10090
3    rat         Rattus norvegicus          10116
11   yeast       Saccharomyces cerevisiae   4932
12   zebrafish   Danio rerio                7955
13   fly         Drosophila melanogaster    7227
14   worm        Caenorhabditis elegans     6239}

\item{raw}{\code{TRUE} to receive results as-is from Gemma, or \code{FALSE} to enable
parsing.}

\item{async}{\code{TRUE} to run the API query on a separate worker, or \code{FALSE} to run
synchronously. See the \code{async} package for details.}

\item{memoised}{Whether or not to cache results so future requests for the same data
will be faster. Use \code{forgetGemmaMemoised} to clear the cache.}

\item{file}{The name of a file to save the results to, or \code{NULL} to not write
results to a file. If \code{raw == TRUE}, the output will be a JSON file.
Otherwise, it will be a RDS file.}

\item{overwrite}{Whether or not to overwrite if a file exists at the specified filename.}
}
\value{
An array of taxa (taxon value objects) matching the given identifiers.
A \verb{400 error} if all identifiers are invalid.
An array of all available taxa, if no identifiers were provided.
}
\description{
List taxa filtered by given parameters
}
\keyword{taxon}
