% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{getTaxa}
\alias{getTaxa}
\title{getTaxa
List taxa filtered by given parameters}
\usage{
getTaxa(
  taxon = NA_character_,
  raw = getOption("gemma.raw", F),
  async = getOption("gemma.async", F),
  memoised = getOption("gemma.memoise", F),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", F)
)
}
\arguments{
\item{taxon}{Not required, part of the URL path.
can either be Taxon ID, Taxon NCBI ID, or one of its string identifiers:
scientific name, common name
It is recommended to use Taxon ID for efficiency.
Please note, that not all taxa have all the possible identifiers
available.
Use the \'All Taxa\' endpoint to retrieve the necessary information. For
convenience, below is a list of officially supported taxa:
ID   Comm.name   Scient.name                NcbiID

1    human       Homo sapiens               9606
2    mouse       Mus musculus               10090
3    rat         Rattus norvegicus          10116
11   yeast       Saccharomyces cerevisiae   4932
12   zebrafish   Danio rerio                7955
13   fly         Drosophila melanogaster    7227
14   worm        Caenorhabditis elegans     6239}

\item{raw}{\code{FALSE} to receive results as-is from Gemma, or \code{TRUE} to enable
parsing.}

\item{async}{\code{TRUE} to run the API query on a separate worker, or \code{FALSE} to run
synchronously. See the \code{async} package for details.}

\item{memoised}{Whether or not to cache results so future requests for the same data
will be faster. Use \code{forgetGemmaMemoised} to clear the cache.}

\item{file}{The name of a file to save the results to, or \code{NULL} to not write
results to a file. If \code{raw == TRUE}, the output will be a JSON file.
Otherwise, it will be a RDS file.}

\item{overwrite}{Whether or not to overwrite if a file exists at the specified filename.}
}
\value{
An array of taxa (taxon value objects) matching the given identifiers.
A \verb{400 error} if all identifiers are invalid.
An array of all available taxa, if no identifiers were provided.
}
\description{
getTaxa
List taxa filtered by given parameters
}
