% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{getTaxonPhenotypes}
\alias{getTaxonPhenotypes}
\title{getTaxonPhenotypes
Loads all phenotypes for the given taxon.}
\usage{
getTaxonPhenotypes(
  taxon = NA_character_,
  editableOnly = FALSE,
  tree = FALSE,
  raw = FALSE,
  async = FALSE,
  memoised = FALSE,
  file = NA_character_,
  overwrite = FALSE
)
}
\arguments{
\item{taxon}{<p class='description-frow'>Not required, part of the URL path.</p><p>   can either be Taxon ID, Taxon NCBI ID, or one of its string identifiers:   scientific name, common name</p><p>It is recommended to use Taxon ID for efficiency.</p><p>Please note, that not all taxa have all the possible identifiers available.</p><p>Use the 'All Taxa' endpoint to retrieve the necessary information. For convenience, below is a list of officially supported taxa: </p><table><tr><th> ID </th><th>Comm.name</th><th>Scient.name</th><th>NcbiID</th></tr><tr><td>  1 </td><td> human           </td><td> Homo sapiens              </td><td>    9606 </td></tr><tr><td>  2 </td><td> mouse           </td><td> Mus musculus              </td><td>   10090 </td></tr><tr><td>  3 </td><td> rat             </td><td> Rattus norvegicus         </td><td>   10116 </td></tr><tr><td> 11 </td><td> yeast           </td><td> Saccharomyces cerevisiae  </td><td>    4932 </td></tr><tr><td> 12 </td><td> zebrafish       </td><td> Danio rerio               </td><td>    7955 </td></tr><tr><td> 13 </td><td> fly             </td><td> Drosophila melanogaster   </td><td>    7227 </td></tr><tr><td> 14 </td><td> worm            </td><td> Caenorhabditis elegans    </td><td>    6239 </td></tr></table>}

\item{editableOnly}{<p class='description-frow'>Optional, defaults to <code>false</code>.</p><p>Whether to only list editable objects.</p>}

\item{tree}{<p class='description-frow'>Optional, defaults to <code>false</code>.</p><p>Whether the returned structure should be an actual tree (nested JSON objects).</p> <p>   Default is false - the tree is flattened and the tree structure information is stored as   the values of the returned object.</p>}

\item{raw}{<p><code>FALSE</code> to receive results as-is from Gemma, or <code>TRUE</code> to enable parsing.</p>}

\item{async}{<p><code>TRUE</code> to run the API query on a separate worker, or <code>FALSE</code> to run synchronously. See the <code>async</code> package for details.</p>}

\item{memoised}{<p>Whether or not to cache results so future requests for the same data will be faster. Use <code>forgetGemmaMemoised</code> to clear the cache.</p>}

\item{file}{<p>The name of a file to save the results to, or <code>NULL</code> to not write results to a file. If <code>raw == TRUE</code>, the output will be a JSON file. Otherwise, it will be a RDS file.</p>}

\item{overwrite}{<p>Whether or not to overwrite if a file exists at the specified filename.</p>}
}
\value{
<p>If <code>tree = false</code>, an array of simple tree value objects.</p><p>   If <code>tree = true</code>, an array of tree characteristic value objects, that will represent root nodes of   a phenotype tree. Each node has its child nodes in the <code>children</code> array property. There will be    exactly 3 root nodes - for a disease ontology tree, human phenotype ontology tree and a mammalian phenotype ontology tree.   If there are no terms for the given taxon in any of the ontologies, the relevant root node will be null.</p><p>A <code>404 error</code> if the given identifier does not map to any object.</p>
}
\description{
getTaxonPhenotypes
Loads all phenotypes for the given taxon.
}
