% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{getTaxonPhenotypes}
\alias{getTaxonPhenotypes}
\title{getTaxonPhenotypes}
\usage{
getTaxonPhenotypes(
  taxon = NA_character_,
  editableOnly = FALSE,
  tree = FALSE,
  raw = FALSE,
  async = FALSE,
  memoised = FALSE
)
}
\description{
getTaxonPhenotypes
}
