% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{getTaxonPhenotypes}
\alias{getTaxonPhenotypes}
\title{Taxon phenotypes
Loads all phenotypes for the given taxon.}
\usage{
getTaxonPhenotypes(
  taxon = NA_character_,
  editableOnly = FALSE,
  tree = FALSE,
  raw = getOption("gemma.raw", F),
  async = getOption("gemma.async", F),
  memoised = getOption("gemma.memoise", F),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", F)
)
}
\arguments{
\item{taxon}{Not required, part of the URL path.
can either be Taxon ID, Taxon NCBI ID, or one of its string identifiers:
scientific name, common name
It is recommended to use Taxon ID for efficiency.
Please note, that not all taxa have all the possible identifiers
available.
Use the \'All Taxa\' endpoint to retrieve the necessary information. For
convenience, below is a list of officially supported taxa:
ID   Comm.name   Scient.name                NcbiID

1    human       Homo sapiens               9606
2    mouse       Mus musculus               10090
3    rat         Rattus norvegicus          10116
11   yeast       Saccharomyces cerevisiae   4932
12   zebrafish   Danio rerio                7955
13   fly         Drosophila melanogaster    7227
14   worm        Caenorhabditis elegans     6239}

\item{editableOnly}{Optional, defaults to \code{false}.
Whether to only list editable objects.}

\item{tree}{Optional, defaults to \code{false}.
Whether the returned structure should be an actual tree (nested JSON
objects).
Default is false - the tree is flattened and the tree structure
information is stored as the values of the returned object.}

\item{raw}{\code{FALSE} to receive results as-is from Gemma, or \code{TRUE} to enable
parsing.}

\item{async}{\code{TRUE} to run the API query on a separate worker, or \code{FALSE} to run
synchronously. See the \code{async} package for details.}

\item{memoised}{Whether or not to cache results so future requests for the same data
will be faster. Use \code{forgetGemmaMemoised} to clear the cache.}

\item{file}{The name of a file to save the results to, or \code{NULL} to not write
results to a file. If \code{raw == TRUE}, the output will be a JSON file.
Otherwise, it will be a RDS file.}

\item{overwrite}{Whether or not to overwrite if a file exists at the specified filename.}
}
\value{
If \code{tree = false}, an array of simple tree value objects.
If \code{tree = true}, an array of tree characteristic value objects, that
will represent root nodes of a phenotype tree. Each node has its child
nodes in the \code{children} array property. There will be exactly 3 root
nodes - for a disease ontology tree, human phenotype ontology tree and a
mammalian phenotype ontology tree. If there are no terms for the given
taxon in any of the ontologies, the relevant root node will be null.
A \verb{404 error} if the given identifier does not map to any object.
}
\description{
Taxon phenotypes
Loads all phenotypes for the given taxon.
}
