% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{getTidyDataset}
\alias{getTidyDataset}
\title{Get Tidy Dataset and Design}
\usage{
getTidyDataset(dataset, filter = TRUE)
}
\arguments{
\item{dataset}{A dataset identifier.}

\item{filter}{The filtered version corresponds to what is used in most Gemma analyses, removing some probes/elements. Unfiltered includes all elements.}
}
\value{
A \code{\link[tibble]{tibble}}.
}
\description{
Combines the expression and design matrix of the queried dataset into a
tibble for easy visualization and exploration with \code{\link[ggplot2]{ggplot}} and the rest of the tidyverse.
}
\examples{
getTidyDataset("GSE2018")
}
\keyword{dataset}
