% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{memgetDatasetDE}
\alias{memgetDatasetDE}
\title{Memoise getDatasetDE}
\usage{
memgetDatasetDE(
  datasets = NA_character_,
  keepNonSpecific = FALSE,
  diffExSet = NA_integer_,
  threshold = 100,
  limit = 100L,
  consolidate = NA_character_,
  raw = TRUE,
  async = FALSE,
  memoised = FALSE
)
}
\description{
Memoise getDatasetDE
}
