% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{memgetDatasetDesign}
\alias{memgetDatasetDesign}
\title{Memoise getDatasetDesign}
\usage{
memgetDatasetDesign(
  dataset = NA_character_,
  raw = getOption("gemma.raw", FALSE),
  async = getOption("gemma.async", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)
}
\description{
Memoise getDatasetDesign
}
\keyword{internal}
