% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{memgetDatasetPlatforms}
\alias{memgetDatasetPlatforms}
\title{Memoise getDatasetPlatforms}
\usage{
memgetDatasetPlatforms(
  dataset = NA_character_,
  raw = getOption("gemma.raw", F),
  async = getOption("gemma.async", F),
  memoised = getOption("gemma.memoise", F),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", F)
)
}
\description{
Memoise getDatasetPlatforms
}
\keyword{internal}
