% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{memgetEvidenceOnTaxon}
\alias{memgetEvidenceOnTaxon}
\title{Memoise getEvidenceOnTaxon}
\usage{
memgetEvidenceOnTaxon(
  taxon = NA_character_,
  gene = NA_character_,
  raw = getOption("gemma.raw", F),
  async = getOption("gemma.async", F),
  memoised = getOption("gemma.memoise", F),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", F)
)
}
\description{
Memoise getEvidenceOnTaxon
}
