% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{memgetGeneGO}
\alias{memgetGeneGO}
\title{Memoise getGeneGO}
\usage{
memgetGeneGO(
  gene = NA_character_,
  raw = FALSE,
  async = FALSE,
  memoised = FALSE,
  file = NA_character_,
  overwrite = FALSE
)
}
\description{
Memoise getGeneGO
}
