% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{memgetGeneLocationOnTaxon}
\alias{memgetGeneLocationOnTaxon}
\title{Memoise getGeneLocationOnTaxon}
\usage{
memgetGeneLocationOnTaxon(
  taxon = NA_character_,
  gene = NA_character_,
  raw = getOption("gemma.raw", FALSE),
  async = getOption("gemma.async", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)
}
\description{
Memoise getGeneLocationOnTaxon
}
\keyword{internal}
