% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{memgetPlatformDatasets}
\alias{memgetPlatformDatasets}
\title{Memoise getPlatformDatasets}
\usage{
memgetPlatformDatasets(
  platform = NA_character_,
  offset = 0L,
  limit = 20L,
  raw = getOption("gemma.raw", FALSE),
  async = getOption("gemma.async", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)
}
\description{
Memoise getPlatformDatasets
}
\keyword{internal}
