% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{memgetPlatformElementGenes}
\alias{memgetPlatformElementGenes}
\title{Memoise getPlatformElementGenes}
\usage{
memgetPlatformElementGenes(
  platform = NA_character_,
  element = NA_character_,
  offset = 0L,
  limit = 20L,
  raw = getOption("gemma.raw", F),
  async = getOption("gemma.async", F),
  memoised = getOption("gemma.memoise", F),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", F)
)
}
\description{
Memoise getPlatformElementGenes
}
