% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{memgetPlatformElementGenes}
\alias{memgetPlatformElementGenes}
\title{Memoise getPlatformElementGenes}
\usage{
memgetPlatformElementGenes(
  platform = NA_character_,
  element = NA_character_,
  offset = 0L,
  limit = 20L,
  raw = FALSE,
  async = FALSE,
  memoised = FALSE
)
}
\description{
Memoise getPlatformElementGenes
}
