% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{memgetPlatformElements}
\alias{memgetPlatformElements}
\title{Memoise getPlatformElements}
\usage{
memgetPlatformElements(
  platform = NA_character_,
  offset = 0L,
  limit = 20L,
  raw = TRUE,
  async = FALSE,
  memoised = FALSE
)
}
\description{
Memoise getPlatformElements
}
