% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{memgetTaxonPhenotypes}
\alias{memgetTaxonPhenotypes}
\title{Memoise getTaxonPhenotypes}
\usage{
memgetTaxonPhenotypes(
  taxon = NA_character_,
  editableOnly = FALSE,
  tree = FALSE,
  raw = getOption("gemma.raw", F),
  async = getOption("gemma.async", F),
  memoised = getOption("gemma.memoise", F),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", F)
)
}
\description{
Memoise getTaxonPhenotypes
}
