% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{memsearchDatasets}
\alias{memsearchDatasets}
\title{Memoise searchDatasets}
\usage{
memsearchDatasets(
  taxon = NA_character_,
  query = NA_character_,
  filter = NA_character_,
  offset = 0L,
  limit = 0L,
  sort = "+id",
  raw = getOption("gemma.raw", F),
  async = getOption("gemma.async", F),
  memoised = getOption("gemma.memoise", F),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", F)
)
}
\description{
Memoise searchDatasets
}
\keyword{internal}
