% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{platformInfo}
\alias{platformInfo}
\title{platformInfo}
\usage{
platformInfo(
  platform = NA_character_,
  request = NA_character_,
  ...,
  raw = FALSE,
  async = FALSE,
  memoised = FALSE,
  file = NA_character_,
  overwrite = FALSE
)
}
\description{
platformInfo
}
