% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processAnnotations}
\alias{processAnnotations}
\title{Processes JSON as annotations}
\usage{
processAnnotations(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A processed data.table
}
\description{
Processes JSON as annotations
}
