% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processAnnotations}
\alias{processAnnotations}
\title{Processes JSON as annotations}
\usage{
processAnnotations(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A data table with information about the annotations of the queried
dataset. A list if \code{raw = TRUE}.A \code{404 error} if the given
identifier does not map to any object.

The fields of the output data.table are:

\itemize{
\item \code{class.name}: Name of the annotation class (e.g. organism part)
\item \code{class.URI}: URI for the annotation class
\item \code{term.name}: Name of the annotation term (e.g. lung)
\item \code{term.URI}: URI for the annotation term
}
}
\description{
Processes JSON as annotations
}
\keyword{internal}
