% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processBibliographicReferenceValueObject}
\alias{processBibliographicReferenceValueObject}
\title{processBibliographicReferenceValueObject}
\usage{
processBibliographicReferenceValueObject(d)
}
\value{
A data.table containing references.
The fields of the output data.table are:

\itemize{
\item \code{title}: Title of the paper
\item \code{authors}: List of authors
\item \code{citation}: Citation string for the paper
\item \code{pmid}: PMID of the paper
\item \code{url}: Pubmed URL of the paper
}
}
\description{
processBibliographicReferenceValueObject
}
