% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processDatasets}
\alias{processDatasets}
\title{Processes JSON as a vector of datasets}
\usage{
processDatasets(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A data table with information about the queried dataset(s). A list if
\code{raw = TRUE}. Returns an empty list if no datasets matched. A successful
response may contain 'Geeq' information, which aims to provide a unified
metric to measure experiments by the quality of their data, and their
suitability for use in Gemma. You can
read more about the geeq properties \href{https://pavlidislab.github.io/Gemma/geeq.html}{here}.

The fields of the output data.table are:

\itemize{
\item \code{ee.ShortName}: Shortname given to the dataset within Gemma. Often corresponds to accession ID
\item \code{ee.Name}: Full title of the dataset
\item \code{ee.ID}: Internal ID of the dataset.
\item \code{ee.Description}: Description of the dataset
\item \code{ee.Public}: Is the dataset publicly available.
\item \code{ee.Troubled}: Did an automatic process within gemma or a curator mark the dataset as "troubled"
\item \code{ee.Accession}: Accession ID of the dataset in the external database it was taken from
\item \code{ee.Database}: The name of the database where the dataset was taken from
\item \code{ee.URL}: URL of the original database
\item \code{ee.SampleCount}: Number of samples in the dataset
\item \code{ee.batchEffect}: A text field describing whether the dataset has batch effects
\item \code{ee.batchCorrected}: Whether batch correction has been performed on the dataset
\item \code{geeq.batchConfound}: 0 if batch info isn't available, -1 if batch counfoud is detected, 1 if batch information is available and no batch confound found
\item \code{geeq.batchEffect}: -1 if batch p value < 0.0001, 1 if batch p value > 0.1, 0 if otherwise and when there is no batch information is available or when the data is confounded with batches.
\item \code{geeq.qScore}: Data quality score given to the dataset by Gemma.
\item \code{geeq.sScore}: Suitability score given to the dataset by Gemma. Refers to factors like batches, platforms and other aspects of experimental design
\item \code{taxon.Name}: The taxa of the study. In Gemma each study will include a single species. If the original source has samples from multiple species, they will be split into different studies within Gemma
\item \code{taxon.ID}: Internal ID given to the taxon by Gemma
}
}
\description{
Processes JSON as a vector of datasets
}
\keyword{internal}
