% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processDEA}
\alias{processDEA}
\title{Processes JSON as a differential expression analysis}
\usage{
processDEA(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A data table with information about the differential expression
analysis of the queried dataset. Note that this funciton does not return
differential expression values themselves. Use \code{\link{get_differential_expression_values}}
to get differential expression values (see examples).

The fields of the output data.table are:

\itemize{
\item \code{result.ID}: Result set ID of the differential expression analysis.
May represent multiple factors in a single model.
\item \code{contrast.ID}: Id of the specific contrast factor. Together with the result.ID
they uniquely represent a given contrast.
\item \code{experiment.ID}: Id of the source experiment
\item \code{baseline.category}: Category for the contrast
\item \code{baseline.categoryURI}: URI for the baseline category
\item \code{baseline.factors}: Characteristics of the baseline. This field is a data.table
\item \code{experimental.factors}: Characteristics of the experimental group. This field is a data.table
\item \code{isSubset}: TRUE if the result set belong to a subset, FALSE if not. Subsets are created when performing differential expression to avoid unhelpful comparisons.
\item \code{subsetFactor}: Characteristics of the subset. This field is a data.table
\item \code{probes.analyzed}: Number of probesets represented in the contrast
\item \code{genes.analyzed}: Number of genes represented in the contrast
}
}
\description{
Processes JSON as a differential expression analysis
}
\keyword{internal}
