% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processDifferentialExpressionAnalysisResultSetValueObject}
\alias{processDifferentialExpressionAnalysisResultSetValueObject}
\title{Process JSON of a result set}
\usage{
processDifferentialExpressionAnalysisResultSetValueObject(d)
}
\value{
A data table with information about the queried result sets. Note that this function does not return
differential expression values themselves. Use \code{\link{get_differential_expression_values}}
to get differential expression values

\itemize{
\item \code{result.ID}: Result set ID of the differential expression analysis.
May represent multiple factors in a single model.
\item \code{contrast.ID}: Id of the specific contrast factor. Together with the result.ID
they uniquely represent a given contrast.
\item \code{experiment.ID}: Id of the source experiment
\item \code{factor.category}: Category for the contrast
\item \code{factor.category.URI}: URI for the contrast category
\item \code{factor.ID}: ID of the factor
\item \code{baseline.factors}: Characteristics of the baseline. This field is a data.table
\item \code{experimental.factors}: Characteristics of the experimental group. This field is a data.table
\item \code{isSubset}: TRUE if the result set belong to a subset, FALSE if not. Subsets are created when performing differential expression to avoid unhelpful comparisons.
\item \code{subsetFactor}: Characteristics of the subset. This field is a data.table
}
}
\description{
Process JSON of a result set
}
\keyword{internal}
