% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processElements}
\alias{processElements}
\title{Processes JSON as a vector of elements}
\usage{
processElements(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A data table with information about the probes representing a gene across
all platrofms. A list if \code{raw = TRUE}.
A \code{404 error} if the given identifier does not map to any genes.

The fields of the output data.table are:

\itemize{
\item \code{mapping.name}: Name of the mapping. Typically the probeset name
\item \code{mappping.description}: A free text field providing optional information about the mapping
\item \code{platform.ShortName}: Shortname of the platform given by Gemma. Typically the GPL identifier.
\item \code{platform.Name}: Full name of the platform
\item \code{platform.ID}: Id number of the platform given by Gemma
\item \code{platform.taxon}: Species the platform was designed for
\item \code{platform.taxonID}: Id number of the species given by Gemma
\item \code{platform.type}: Type of the platform.
\item \code{platform.description}: Free text field describing the platform.
\item \code{platform.troubled}: Whether the platform is marked as troubled by a Gemma curator.
}
}
\description{
Processes JSON as a vector of elements
}
\keyword{internal}
