% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processFile}
\alias{processFile}
\title{Processes a response as a gzip file}
\usage{
processFile(response)
}
\arguments{
\item{response}{The response from an \code{http_get} request}
}
\value{
A processed data.table
}
\description{
Processes a response as a gzip file
}
\keyword{internal}
