% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processGemmaArray}
\alias{processGemmaArray}
\title{Processes JSON as an array}
\usage{
processGemmaArray(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A data table with information about the probes representing the gene
across different platforms.
}
\description{
Processes JSON as an array
}
\keyword{internal}
