% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processGenes}
\alias{processGenes}
\title{Processes JSON as a vector of genes}
\usage{
processGenes(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A data table with information about the querried gene(s)
A list if \code{raw = TRUE}.

The fields of the output data.table are:

\itemize{
\item \code{gene.Symbol}: Symbol for the gene
\item \code{gene.Ensembl}: Ensembl ID for the gene
\item \code{gene.NCBI}: NCBI id for the gene
\item \code{gene.Name}: Name of the gene
\item \code{gene.MFX.Rank}: Multifunctionality rank for the gene
\item \code{taxon.Name}: Name of the species
\item \code{taxon.Scientific}: Scientific name for the taxon
\item \code{taxon.ID}: Internal identifier given to the species by Gemma
\item \code{taxon.NCBI}: NCBI ID of the taxon
\item \code{taxon.Database.Name}: Underlying database used in Gemma for the taxon
\item \code{taxon.Database.ID}: ID of the underyling database used in Gemma for the taxon
}
}
\description{
Processes JSON as a vector of genes
}
\keyword{internal}
