% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_processors.R
\name{processGenes}
\alias{processGenes}
\title{Processes JSON as a vector of genes}
\usage{
processGenes(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A processed data.table
}
\description{
Processes JSON as a vector of genes
}
\keyword{internal}
