% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processGenes}
\alias{processGenes}
\title{Processes JSON as a vector of genes}
\usage{
processGenes(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A data table with information about the querried gene(s)
A list if \code{raw = TRUE}.

The fields of the output data.table are:

\itemize{
\item \code{gene.symbol}: Symbol for the gene
\item \code{gene.ensembl}: Ensembl ID for the gene
\item \code{gene.NCBI}: NCBI id for the gene
\item \code{gene.name}: Name of the gene
\item \code{gene.aliases}: Gene aliases. Each row includes a vector
\item \code{gene.MFX.rank}: Multifunctionality rank for the gene
\item \code{taxon.name}: Name of the species
\item \code{taxon.scientific}: Scientific name for the taxon
\item \code{taxon.ID}: Internal identifier given to the species by Gemma
\item \code{taxon.NCBI}: NCBI ID of the taxon
\item \code{taxon.database.name}: Underlying database used in Gemma for the taxon
\item \code{taxon.database.ID}: ID of the underlying database used in Gemma for the taxon
}
}
\description{
Processes JSON as a vector of genes
}
\keyword{internal}
