% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processGO}
\alias{processGO}
\title{Processes JSON as GO terms}
\usage{
processGO(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A data table with information about the GO terms assigned to the
queried gene. A list if \code{raw = TRUE}. A \code{404 error} if the given identifier does not map to any
object.

The fields of the output data.table are:

\itemize{
\item \code{term.Name}: Name of the term
\item \code{term.ID}: ID of the term
\item \code{term.URI}: URI of the term
}
}
\description{
Processes JSON as GO terms
}
\keyword{internal}
