% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processSamples}
\alias{processSamples}
\title{Processes JSON as a vector of samples}
\usage{
processSamples(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A processed data.table
}
\description{
Processes JSON as a vector of samples
}
\keyword{internal}
