% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_processors.R
\name{processSearchAnnotations}
\alias{processSearchAnnotations}
\title{Processes JSON as an annotation}
\usage{
processSearchAnnotations(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A processed data.table
}
\description{
Processes JSON as an annotation
}
\keyword{internal}
