% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processSearchAnnotations}
\alias{processSearchAnnotations}
\title{Processes JSON as an annotation}
\usage{
processSearchAnnotations(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A data table with annotations (annotation search result value objects)
matching the given identifiers. A list if \code{raw = TRUE}. A \code{400 error} if required parameters are missing.

The fields of the output data.table are:

\itemize{
\item \code{category.Name}: Category that the annotation belongs to
\item \code{category.URI}: URI for the category.Name
\item \code{value.Name}: Annotation term
\item \code{value.URI}: URI for the value.Name
}
}
\description{
Processes JSON as an annotation
}
\keyword{internal}
