% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{processTaxon}
\alias{processTaxon}
\title{Processes JSON as a vector of taxa}
\usage{
processTaxon(d)
}
\arguments{
\item{d}{The JSON to process}
}
\value{
A processed data.table

\itemize{
\item \code{taxon.name}: Name of the species
\item \code{taxon.scientific}: Scientific name for the taxon
\item \code{taxon.ID}: Internal identifier given to the species by Gemma
\item \code{taxon.NCBI}: NCBI ID of the taxon
\item \code{taxon.database.name}: Underlying database used in Gemma for the taxon
\item \code{taxon.database.ID}: ID of the underyling database used in Gemma for the taxon
}
}
\description{
Processes JSON as a vector of taxa
}
\keyword{internal}
