% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_allEndpoints.R
\name{searchAnnotations}
\alias{searchAnnotations}
\title{Annotation search}
\usage{
searchAnnotations(
  query,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE),
  attributes = getOption("gemma.attributes", TRUE)
)
}
\arguments{
\item{query}{Required, defaults to \code{empty}.
The search query. Either plain text ('traumatic'), or an ontology term
URI ('http://purl.obolibrary.org/obo/UBERON_0002048'). Datasets that
contain the given string in their short of full name will also be
matched ('GSE201', 'Bronchoalveolar lavage samples'.
Can be multiple identifiers separated by commas.
When
using in scripts, remember to URL-encode any forward slashes in the
phenotype value URIs (see the compiled URL below).}

\item{raw}{\code{TRUE} to receive results as-is from Gemma, or \code{FALSE} to enable
parsing. Raw results usually contain additional fields and flags that are
omitted in the parsed results.}

\item{memoised}{Whether or not to save to cache for future calls with the
same inputs and use the result saved in cache if a result is already saved.
Doing \code{options(gemma.memoised = TRUE)} will ensure that the cache is always
used. Use \\code{\\link{forgetGemmaMemoised}} to clear the cache.}

\item{file}{The name of a file to save the results to, or \code{NULL} to not write
results to a file. If \code{raw == TRUE}, the output will be a JSON file. Otherwise,
it will be a RDS file.}

\item{overwrite}{Whether or not to overwrite if a file exists at the specified
filename.}

\item{attributes}{If \code{TRUE} additional information from the call will be added
into the output object's attributes such as offset and available elements.}
}
\value{
A data table with annotations (annotation search result value objects) matching the given identifiers. A \verb{400 error} if required parameters are missing.
}
\description{
Does a search for annotations based on the given string
}
\examples{
annots <- searchAnnotations("traumatic")
}
\keyword{misc}
