% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{searchAnnotations}
\alias{searchAnnotations}
\title{Annotation search}
\usage{
searchAnnotations(
  query = NA_character_,
  raw = getOption("gemma.raw", FALSE),
  async = getOption("gemma.async", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)
}
\arguments{
\item{query}{Required, defaults to \code{empty}.
The search query. Either plain text ('traumatic'), or an ontology term
URI ('http://purl.obolibrary.org/obo/UBERON_0002048'). Datasets that
contain the given string in their short of full name will also be
matched ('GSE201', 'Bronchoalveolar lavage samples'.
Can be multiple identifiers separated by commas.
When
using in scripts, remember to URL-encode any forward slashes in the
phenotype value URIs (see the compiled URL below).}

\item{raw}{\code{TRUE} to receive results as-is from Gemma, or \code{FALSE} to enable
parsing.}

\item{async}{\code{TRUE} to run the API query on a separate worker, or \code{FALSE} to run
synchronously. See the \code{async} package for details.}

\item{memoised}{Whether or not to cache results so future requests for the same data
will be faster. Use \code{forgetGemmaMemoised} to clear the cache.}

\item{file}{The name of a file to save the results to, or \code{NULL} to not write
results to a file. If \code{raw == TRUE}, the output will be a JSON file.
Otherwise, it will be a RDS file.}

\item{overwrite}{Whether or not to overwrite if a file exists at the specified filename.}
}
\value{
An array of annotations (annotation search result value objects)
matching the given identifiers.
A \verb{400 error} if required parameters are missing.
}
\description{
Does a search for annotations based on the given string
}
\examples{
searchAnnotations("traumatic")
}
\keyword{annotation}
