% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allEndpoints.R
\name{searchDatasets}
\alias{searchDatasets}
\title{Dataset search}
\usage{
searchDatasets(
  query,
  taxon = NA_character_,
  offset = 0L,
  limit = 20L,
  sort = "+id",
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)
}
\arguments{
\item{query}{Required, defaults to \code{empty}.
The search query. Either plain text ('traumatic'), or an ontology term
URI ('http://purl.obolibrary.org/obo/UBERON_0002048'). Datasets that
contain the given string in their short of full name will also be
matched ('GSE201', 'Bronchoalveolar lavage samples'.
Can be multiple identifiers separated by commas.
When
using in scripts, remember to URL-encode any forward slashes in the
phenotype value URIs (see the compiled URL below).}

\item{taxon}{Not required, part of the URL path. can either be Taxon ID, Taxon NCBI
ID, or one of its string identifiers: scientific name, common name.}

\item{offset}{Optional, defaults to \code{0}.
Skips the specified amount of objects when retrieving them from the
database.}

\item{limit}{Optional, defaults to 20. Limits the result to specified amount of
objects.}

\item{sort}{Optional, defaults to \code{+id}.
Sets the ordering property and direction.
Format is \verb{[+,-][property name]}. E.g. \code{-accession} will translate to
descending ordering by the 'Accession' property.
Note that this does [not guarantee the order of the returned
entities!]{.description-imp} This is merely a signal to how the data
should be pre-sorted before the limit and offset are applied.
Nested properties are also supported (recursively).
E.g: \code{+curationDetails.lastTroubledEvent.date}
When
using in scripts, remember to URL-encode the '+' plus character (see
the compiled URL below).}

\item{raw}{\code{TRUE} to receive results as-is from Gemma, or \code{FALSE} to enable
parsing. Raw results usually contain additional fields and flags that
are omitted in the parsed results.}

\item{memoised}{Whether or not to cache results so future requests for the same data
will be faster. Use \code{forgetGemmaMemoised} to clear the cache.}

\item{file}{The name of a file to save the results to, or \code{NULL} to not write
results to a file. If \code{raw == TRUE}, the output will be a JSON file.
Otherwise, it will be a RDS file.}

\item{overwrite}{Whether or not to overwrite if a file exists at the specified filename.}
}
\value{
An array of datasets (expression experiment value objects) that are
annotated with the given ontology terms, or, in case of plaintext query,
experiments that contain the given words (name, short name, accession,
tags)
If an ontology term URI is given, the results will also include datasets
that are associated with the descendants of the term.
The search only only checks the annotations value, not the category
(which is also an ontology term).
}
\description{
Does a search for datasets containing annotations, short name or full
name matching the given string
}
\examples{
dat <- searchDatasets("bipolar")
str(dat)
}
\keyword{dataset}
