% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{setGemmaUser}
\alias{setGemmaUser}
\title{Set Gemma User}
\usage{
setGemmaUser(username = NULL, password = NULL)
}
\description{
Allows the user to access information that requires logging in to Gemma. To log out, run \code{setGemmaUser} without specifying the username or password.
}
\examples{
setGemmaUser('username','password') # login
setGemmaUser() # logout

}
\keyword{misc}
