% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{visualize_dataset}
\alias{visualize_dataset}
\title{Quick differential expression visualization}
\usage{
visualize_dataset(
  dataset,
  genes = NULL,
  adj_p_filter = 1,
  rank_filter = Inf,
  ...
)
}
\arguments{
\item{dataset}{Name or id of a dataset or a SummarizedExperiment as returned by \code{\link{get_dataset_object}}}

\item{genes}{An optional list of gene symbols or identifiers}

\item{adj_p_filter}{P value threshold to filter genes}

\item{rank_filter}{Positive integer or a ratio.}

\item{...}{passed to \code{\link{get_dataset_object}} if dataset is a name}
}
\value{
A list of pheatmap plots for every differential expression contrast of an experiment.
Each plot will be subsetted to exclusively include the relevant subset for the contrast.
}
\description{
Quick differential expression visualization
}
\keyword{dataset}
